sf POLYGON

constructs an sf of POLYGON objects

sf_polygon(
  obj = NULL,
  x = NULL,
  y = NULL,
  z = NULL,
  m = NULL,
  polygon_id = NULL,
  linestring_id = NULL,
  close = TRUE,
  keep = FALSE,
  list_columns = NULL
)

Arguments

obj

sorted matrix or data.frame

x

x geometry column

y

y geometry column

z

z geometry column

m

m geometry column

polygon_id

column of ids for polygons

linestring_id

column of ids for lines (within polygons)

close

logical indicating whether polygons should be closed. If TRUE, all polygons will be checked and force closed if possible

keep

logical indicating if the non-geometry and non-id columns should be kept. if TRUE you must supply the geometry and id columns, and only the first row of each geometry is kept. See Keeping Properties.

list_columns

vector of column names to turn into a list.

Value

sf object of POLYGON geometries

notes

sfheaders functions do not perform any validity checks on the geometries. Nor do they set Coordinate Reference Systems, EPSG, PROJ4 or precision attributes.

The data.frame and matrices you send into the sfheader functions must be ordered.

Keeping Properties

Setting keep = TRUE will retain any columns not specified as a coordinate (x, y, z, m) or an id (e.g., linestring_id, polygon_id) of the input obj.

You can use list_columns to specify which of the properties will be turned into a list, thus keeping all the values in the column. For columns not specified in list_columns, only the first row of the column is kept

The sf_* functions assume the input obj is a long data.frame / matrix, where any properties are repeated down the table for the same geometry.

Examples

m <- matrix(c(0,0,0,0,1,1), ncol = 2 )
sf_polygon( m )
#>   id               geometry
#> 1  1 0, 0, 0, 0, 0, 1, 1, 0

m <- matrix(c(0,0,0,0,0,1,0,1,1,1,2,2,1,2,3,1,3,4), ncol = 3, byrow = TRUE)
sf_polygon( obj = m )
#>   id                                                      geometry
#> 1  1 0, 0, 0, 1, 1, 1, 0, 0, 0, 1, 2, 2, 3, 0, 0, 1, 1, 2, 3, 4, 0
sf_polygon( obj = m, polygon_id = 1 )
#>   id               geometry
#> 1  0 0, 0, 1, 0, 0, 1, 1, 0
#> 2  1 2, 2, 3, 2, 2, 3, 4, 2
sf_polygon( obj = m, linestring_id = 1 )
#>   id                                       geometry
#> 1  1 0, 0, 1, 0, 0, 1, 1, 0, 2, 2, 3, 2, 2, 3, 4, 2

sf_polygon( obj = m, linestring_id = 1, polygon_id = 1 )
#>   id               geometry
#> 1  0 0, 0, 1, 0, 0, 1, 1, 0
#> 2  1 2, 2, 3, 2, 2, 3, 4, 2

sf_polygon( obj = m, x = 2, y = 3 )
#>   id                                 geometry
#> 1  1 0, 0, 1, 2, 2, 3, 0, 0, 1, 1, 2, 3, 4, 0
sf_polygon( obj = m, x = 1, y = 2, z = 3 )
#>   id                                                      geometry
#> 1  1 0, 0, 0, 1, 1, 1, 0, 0, 0, 1, 2, 2, 3, 0, 0, 1, 1, 2, 3, 4, 0
sf_polygon( obj = m, x = 2, y = 3, linestring_id = 1, polygon_id = 1 )
#>   id               geometry
#> 1  0 0, 0, 1, 0, 0, 1, 1, 0
#> 2  1 2, 2, 3, 2, 2, 3, 4, 2

df <- data.frame(
  ml_id = c(1,1,1,1,1,1,1,1,1,2,2,2,2,2,2)
  , l_id = c(1,1,1,2,2,2,3,3,3,1,1,1,2,2,2)
  , x = rnorm(15)
  , y = rnorm(15)
  , z = rnorm(15)
  , m = rnorm(15)
)

sf_polygon( obj = df, x = "x", y = "y")
#>   id
#> 1  1
#>                                                                                                                                                                                                                                                                                                                                                                                                 geometry
#> 1 1.17471632, -0.03359987, 0.93299715, 0.33646381, 0.97054568, 0.44654142, -0.27889505, 1.06176213, -0.70416464, 0.76131838, 0.38890002, -0.37295066, -0.45998313, 1.45669338, 0.37128253, 1.17471632, 1.31659976, 0.35459108, 0.08399519, 0.37145891, 1.36361860, -1.05763890, 1.14926344, 1.01312534, 1.49017902, 0.17754417, 0.37816506, -0.40763147, 0.49977437, -0.67397030, 2.13768870, 1.31659976
sf_polygon( obj = df, x = "x", y = "y", z = "z")
#>   id
#> 1  1
#>                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                           geometry
#> 1 1.17471632, -0.03359987, 0.93299715, 0.33646381, 0.97054568, 0.44654142, -0.27889505, 1.06176213, -0.70416464, 0.76131838, 0.38890002, -0.37295066, -0.45998313, 1.45669338, 0.37128253, 1.17471632, 1.31659976, 0.35459108, 0.08399519, 0.37145891, 1.36361860, -1.05763890, 1.14926344, 1.01312534, 1.49017902, 0.17754417, 0.37816506, -0.40763147, 0.49977437, -0.67397030, 2.13768870, 1.31659976, -0.65666914, 0.19151246, -1.04489896, 0.61148223, 0.09801032, -1.60669668, -0.89432051, -1.40476524, 0.01253939, -0.68425097, -1.11669247, -0.82806644, -0.62734241, 0.22329230, 0.24257430, -0.65666914
sf_polygon( obj = df, x = "x", y = "y", z = "z", m = "m")
#>   id
#> 1  1
#>                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                   geometry
#> 1 1.17471632, -0.03359987, 0.93299715, 0.33646381, 0.97054568, 0.44654142, -0.27889505, 1.06176213, -0.70416464, 0.76131838, 0.38890002, -0.37295066, -0.45998313, 1.45669338, 0.37128253, 1.17471632, 1.31659976, 0.35459108, 0.08399519, 0.37145891, 1.36361860, -1.05763890, 1.14926344, 1.01312534, 1.49017902, 0.17754417, 0.37816506, -0.40763147, 0.49977437, -0.67397030, 2.13768870, 1.31659976, -0.65666914, 0.19151246, -1.04489896, 0.61148223, 0.09801032, -1.60669668, -0.89432051, -1.40476524, 0.01253939, -0.68425097, -1.11669247, -0.82806644, -0.62734241, 0.22329230, 0.24257430, -0.65666914, 1.69614748, 0.75331532, -1.28748814, 1.21324312, -1.49664272, -0.63341116, -0.23354666, -2.60381802, 0.41794893, 0.29087332, -1.36372031, -0.11426559, 1.39767897, 1.21168999, -0.40464779, 1.69614748

sf_polygon( obj = df, x = 2, y = 3)
#>   id
#> 1  1
#>                                                                                                                                                                                                                                                                                                                                                                                              geometry
#> 1 1.00000000, 1.00000000, 1.00000000, 2.00000000, 2.00000000, 2.00000000, 3.00000000, 3.00000000, 3.00000000, 1.00000000, 1.00000000, 1.00000000, 2.00000000, 2.00000000, 2.00000000, 1.00000000, 1.17471632, -0.03359987, 0.93299715, 0.33646381, 0.97054568, 0.44654142, -0.27889505, 1.06176213, -0.70416464, 0.76131838, 0.38890002, -0.37295066, -0.45998313, 1.45669338, 0.37128253, 1.17471632
sf_polygon( obj = df, x = 2, y = 3, z = 4)
#>   id
#> 1  1
#>                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                 geometry
#> 1 1.00000000, 1.00000000, 1.00000000, 2.00000000, 2.00000000, 2.00000000, 3.00000000, 3.00000000, 3.00000000, 1.00000000, 1.00000000, 1.00000000, 2.00000000, 2.00000000, 2.00000000, 1.00000000, 1.17471632, -0.03359987, 0.93299715, 0.33646381, 0.97054568, 0.44654142, -0.27889505, 1.06176213, -0.70416464, 0.76131838, 0.38890002, -0.37295066, -0.45998313, 1.45669338, 0.37128253, 1.17471632, 1.31659976, 0.35459108, 0.08399519, 0.37145891, 1.36361860, -1.05763890, 1.14926344, 1.01312534, 1.49017902, 0.17754417, 0.37816506, -0.40763147, 0.49977437, -0.67397030, 2.13768870, 1.31659976
sf_polygon( obj = df, x = 2, y = 3, z = 4, m = 5)
#>   id
#> 1  1
#>                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                           geometry
#> 1 1.00000000, 1.00000000, 1.00000000, 2.00000000, 2.00000000, 2.00000000, 3.00000000, 3.00000000, 3.00000000, 1.00000000, 1.00000000, 1.00000000, 2.00000000, 2.00000000, 2.00000000, 1.00000000, 1.17471632, -0.03359987, 0.93299715, 0.33646381, 0.97054568, 0.44654142, -0.27889505, 1.06176213, -0.70416464, 0.76131838, 0.38890002, -0.37295066, -0.45998313, 1.45669338, 0.37128253, 1.17471632, 1.31659976, 0.35459108, 0.08399519, 0.37145891, 1.36361860, -1.05763890, 1.14926344, 1.01312534, 1.49017902, 0.17754417, 0.37816506, -0.40763147, 0.49977437, -0.67397030, 2.13768870, 1.31659976, -0.65666914, 0.19151246, -1.04489896, 0.61148223, 0.09801032, -1.60669668, -0.89432051, -1.40476524, 0.01253939, -0.68425097, -1.11669247, -0.82806644, -0.62734241, 0.22329230, 0.24257430, -0.65666914

sf_polygon( obj = df, polygon_id = "ml_id", linestring_id = "l_id" )
#>   ml_id
#> 1     1
#> 2     2
#>                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                           geometry
#> 1 1.17471632, -0.03359987, 0.93299715, 1.17471632, 1.31659976, 0.35459108, 0.08399519, 1.31659976, -0.65666914, 0.19151246, -1.04489896, -0.65666914, 1.69614748, 0.75331532, -1.28748814, 1.69614748, 0.33646381, 0.97054568, 0.44654142, 0.33646381, 0.37145891, 1.36361860, -1.05763890, 0.37145891, 0.61148223, 0.09801032, -1.60669668, 0.61148223, 1.21324312, -1.49664272, -0.63341116, 1.21324312, -0.27889505, 1.06176213, -0.70416464, -0.27889505, 1.14926344, 1.01312534, 1.49017902, 1.14926344, -0.89432051, -1.40476524, 0.01253939, -0.89432051, -0.23354666, -2.60381802, 0.41794893, -0.23354666
#> 2                                                                                                                                                                                                                                     0.7613184, 0.3889000, -0.3729507, 0.7613184, 0.1775442, 0.3781651, -0.4076315, 0.1775442, -0.6842510, -1.1166925, -0.8280664, -0.6842510, 0.2908733, -1.3637203, -0.1142656, 0.2908733, -0.4599831, 1.4566934, 0.3712825, -0.4599831, 0.4997744, -0.6739703, 2.1376887, 0.4997744, -0.6273424, 0.2232923, 0.2425743, -0.6273424, 1.3976790, 1.2116900, -0.4046478, 1.3976790
sf_polygon( obj = df, polygon_id = 1, linestring_id = 2 )
#>   id
#> 1  1
#> 2  2
#>                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                           geometry
#> 1 1.17471632, -0.03359987, 0.93299715, 1.17471632, 1.31659976, 0.35459108, 0.08399519, 1.31659976, -0.65666914, 0.19151246, -1.04489896, -0.65666914, 1.69614748, 0.75331532, -1.28748814, 1.69614748, 0.33646381, 0.97054568, 0.44654142, 0.33646381, 0.37145891, 1.36361860, -1.05763890, 0.37145891, 0.61148223, 0.09801032, -1.60669668, 0.61148223, 1.21324312, -1.49664272, -0.63341116, 1.21324312, -0.27889505, 1.06176213, -0.70416464, -0.27889505, 1.14926344, 1.01312534, 1.49017902, 1.14926344, -0.89432051, -1.40476524, 0.01253939, -0.89432051, -0.23354666, -2.60381802, 0.41794893, -0.23354666
#> 2                                                                                                                                                                                                                                     0.7613184, 0.3889000, -0.3729507, 0.7613184, 0.1775442, 0.3781651, -0.4076315, 0.1775442, -0.6842510, -1.1166925, -0.8280664, -0.6842510, 0.2908733, -1.3637203, -0.1142656, 0.2908733, -0.4599831, 1.4566934, 0.3712825, -0.4599831, 0.4997744, -0.6739703, 2.1376887, 0.4997744, -0.6273424, 0.2232923, 0.2425743, -0.6273424, 1.3976790, 1.2116900, -0.4046478, 1.3976790

## keeping properties
df <- data.frame(
  ml_id = c(1,1,1,1,1,1,1,1,1,2,2,2,2,2,2)
  , l_id = c(1,1,1,2,2,2,3,3,3,1,1,1,2,2,2)
  , x = rnorm(15)
  , y = rnorm(15)
  , z = rnorm(15)
  , m = rnorm(15)
  , val = letters[1:15]
  , stringsAsFactors = FALSE
)

## using keep = TRUE means the first row of all non-geometries are kept
sf_polygon(
  obj = df
  , polygon_id = "ml_id"
  , linestring_id = "l_id"
  , x = "x"
  , y = "y"
  , keep = TRUE
)
#>   ml_id           z            m val
#> 1     1 -0.01593882 -0.679341375   a
#> 2     2  1.07465069 -0.004812903   j
#>                                                                                                                                                                                                                                                                                geometry
#> 1 -0.9981487, -0.2263870, -0.2114960, -0.9981487, 0.4154389, -0.9731385, -1.7888938, 0.4154389, 0.8451759, -0.2968676, -0.9762262, 0.8451759, 0.2360719, 0.2776883, -0.6408947, 0.2360719, -1.4545424, 0.9748924, 1.5947642, -1.4545424, -0.1483551, -0.6906724, -1.6499874, -0.1483551
#> 2                                                                                                 -1.6782345, 0.9302221, 0.1325310, -1.6782345, -1.6894634, 0.4395878, -0.8791112, -1.6894634, 0.1431183, 0.4582015, -0.1050574, 0.1431183, 0.3283822, 1.0775859, -0.2243974, 0.3283822

## use 'list_column' to specify columns where you want to keep all the values
sf_polygon(
  obj = df
  , polygon_id = "ml_id"
  , linestring_id = "l_id"
  , x = "x"
  , y = "y"
  , keep = TRUE
  , list_columns = "val"
)
#>   ml_id           z            m                                val
#> 1     1 -0.01593882 -0.679341375 a, b, c, a, d, e, f, d, g, h, i, g
#> 2     2  1.07465069 -0.004812903             j, k, l, j, m, n, o, m
#>                                                                                                                                                                                                                                                                                geometry
#> 1 -0.9981487, -0.2263870, -0.2114960, -0.9981487, 0.4154389, -0.9731385, -1.7888938, 0.4154389, 0.8451759, -0.2968676, -0.9762262, 0.8451759, 0.2360719, 0.2776883, -0.6408947, 0.2360719, -1.4545424, 0.9748924, 1.5947642, -1.4545424, -0.1483551, -0.6906724, -1.6499874, -0.1483551
#> 2                                                                                                 -1.6782345, 0.9302221, 0.1325310, -1.6782345, -1.6894634, 0.4395878, -0.8791112, -1.6894634, 0.1431183, 0.4582015, -0.1050574, 0.1431183, 0.3283822, 1.0775859, -0.2243974, 0.3283822