sfc MULTILINESTRING

constructs an sfc of MULTILINESTRING objects

sfc_multilinestring(
  obj = NULL,
  x = NULL,
  y = NULL,
  z = NULL,
  m = NULL,
  multilinestring_id = NULL,
  linestring_id = NULL
)

Arguments

obj

sorted matrix or data.frame

x

x geometry column

y

y geometry column

z

z geometry column

m

m geometry column

multilinestring_id

column of ids for multilinestrings

linestring_id

column of ids for linestrings (within multilinestrings)

Value

sfc object of MULTILINESTRING geometries

notes

sfheaders functions do not perform any validity checks on the geometries. Nor do they set Coordinate Reference Systems, EPSG, PROJ4 or precision attributes.

The data.frame and matrices you send into the sfheader functions must be ordered.

Examples

m <- matrix(c(0,0,0,0,1,1), ncol = 3 )
sfc_multilinestring( m )
#> [[1]]
#> [[1]]
#>      [,1] [,2] [,3]
#> [1,]    0    0    1
#> [2,]    0    0    1
#> 
#> attr(,"class")
#> [1] "XYZ"             "MULTILINESTRING" "sfg"            
#> 
#> attr(,"n_empty")
#> [1] 0
#> attr(,"crs")
#> $input
#> [1] NA
#> 
#> $wkt
#> [1] NA
#> 
#> attr(,"class")
#> [1] "crs"
#> attr(,"class")
#> [1] "sfc_MULTILINESTRING" "sfc"                
#> attr(,"precision")
#> [1] 0
#> attr(,"bbox")
#> xmin ymin xmax ymax 
#>    0    0    0    0 
#> attr(,"class")
#> [1] "bbox"
#> attr(,"z_range")
#> zmin zmax 
#>    1    1 
#> attr(,"class")
#> [1] "z_range"

m <- matrix(c(0,0,0,0,0,1,0,1,1,1,2,2,1,2,3), ncol = 3, byrow = TRUE)
sfc_multilinestring( obj = m )
#> [[1]]
#> [[1]]
#>      [,1] [,2] [,3]
#> [1,]    0    0    0
#> [2,]    0    0    1
#> [3,]    0    1    1
#> [4,]    1    2    2
#> [5,]    1    2    3
#> 
#> attr(,"class")
#> [1] "XYZ"             "MULTILINESTRING" "sfg"            
#> 
#> attr(,"n_empty")
#> [1] 0
#> attr(,"crs")
#> $input
#> [1] NA
#> 
#> $wkt
#> [1] NA
#> 
#> attr(,"class")
#> [1] "crs"
#> attr(,"class")
#> [1] "sfc_MULTILINESTRING" "sfc"                
#> attr(,"precision")
#> [1] 0
#> attr(,"bbox")
#> xmin ymin xmax ymax 
#>    0    0    1    2 
#> attr(,"class")
#> [1] "bbox"
#> attr(,"z_range")
#> zmin zmax 
#>    0    3 
#> attr(,"class")
#> [1] "z_range"
sfc_multilinestring( obj = m, multilinestring_id = 1 )
#> [[1]]
#> [[1]]
#>      [,1] [,2]
#> [1,]    0    0
#> [2,]    0    1
#> [3,]    1    1
#> 
#> attr(,"class")
#> [1] "XY"              "MULTILINESTRING" "sfg"            
#> 
#> [[2]]
#> [[1]]
#>      [,1] [,2]
#> [1,]    2    2
#> [2,]    2    3
#> 
#> attr(,"class")
#> [1] "XY"              "MULTILINESTRING" "sfg"            
#> 
#> attr(,"n_empty")
#> [1] 0
#> attr(,"crs")
#> $input
#> [1] NA
#> 
#> $wkt
#> [1] NA
#> 
#> attr(,"class")
#> [1] "crs"
#> attr(,"class")
#> [1] "sfc_MULTILINESTRING" "sfc"                
#> attr(,"precision")
#> [1] 0
#> attr(,"bbox")
#> xmin ymin xmax ymax 
#>    0    0    2    3 
#> attr(,"class")
#> [1] "bbox"
sfc_multilinestring( obj = m, linestring_id = 1 )
#> [[1]]
#> [[1]]
#>      [,1] [,2]
#> [1,]    0    0
#> [2,]    0    1
#> [3,]    1    1
#> 
#> [[2]]
#>      [,1] [,2]
#> [1,]    2    2
#> [2,]    2    3
#> 
#> attr(,"class")
#> [1] "XY"              "MULTILINESTRING" "sfg"            
#> 
#> attr(,"n_empty")
#> [1] 0
#> attr(,"crs")
#> $input
#> [1] NA
#> 
#> $wkt
#> [1] NA
#> 
#> attr(,"class")
#> [1] "crs"
#> attr(,"class")
#> [1] "sfc_MULTILINESTRING" "sfc"                
#> attr(,"precision")
#> [1] 0
#> attr(,"bbox")
#> xmin ymin xmax ymax 
#>    0    0    2    3 
#> attr(,"class")
#> [1] "bbox"

sfc_multilinestring( obj = m, linestring_id = 1, multilinestring_id = 1 )
#> [[1]]
#> [[1]]
#>      [,1] [,2]
#> [1,]    0    0
#> [2,]    0    1
#> [3,]    1    1
#> 
#> attr(,"class")
#> [1] "XY"              "MULTILINESTRING" "sfg"            
#> 
#> [[2]]
#> [[1]]
#>      [,1] [,2]
#> [1,]    2    2
#> [2,]    2    3
#> 
#> attr(,"class")
#> [1] "XY"              "MULTILINESTRING" "sfg"            
#> 
#> attr(,"n_empty")
#> [1] 0
#> attr(,"crs")
#> $input
#> [1] NA
#> 
#> $wkt
#> [1] NA
#> 
#> attr(,"class")
#> [1] "crs"
#> attr(,"class")
#> [1] "sfc_MULTILINESTRING" "sfc"                
#> attr(,"precision")
#> [1] 0
#> attr(,"bbox")
#> xmin ymin xmax ymax 
#>    0    0    2    3 
#> attr(,"class")
#> [1] "bbox"

sfc_multilinestring( obj = m, x = 2, y = 3 )
#> [[1]]
#> [[1]]
#>      [,1] [,2]
#> [1,]    0    0
#> [2,]    0    1
#> [3,]    1    1
#> [4,]    2    2
#> [5,]    2    3
#> 
#> attr(,"class")
#> [1] "XY"              "MULTILINESTRING" "sfg"            
#> 
#> attr(,"n_empty")
#> [1] 0
#> attr(,"crs")
#> $input
#> [1] NA
#> 
#> $wkt
#> [1] NA
#> 
#> attr(,"class")
#> [1] "crs"
#> attr(,"class")
#> [1] "sfc_MULTILINESTRING" "sfc"                
#> attr(,"precision")
#> [1] 0
#> attr(,"bbox")
#> xmin ymin xmax ymax 
#>    0    0    2    3 
#> attr(,"class")
#> [1] "bbox"
sfc_multilinestring( obj = m, x = 1, y = 2, z = 3 )
#> [[1]]
#> [[1]]
#>      [,1] [,2] [,3]
#> [1,]    0    0    0
#> [2,]    0    0    1
#> [3,]    0    1    1
#> [4,]    1    2    2
#> [5,]    1    2    3
#> 
#> attr(,"class")
#> [1] "XYZ"             "MULTILINESTRING" "sfg"            
#> 
#> attr(,"n_empty")
#> [1] 0
#> attr(,"crs")
#> $input
#> [1] NA
#> 
#> $wkt
#> [1] NA
#> 
#> attr(,"class")
#> [1] "crs"
#> attr(,"class")
#> [1] "sfc_MULTILINESTRING" "sfc"                
#> attr(,"precision")
#> [1] 0
#> attr(,"bbox")
#> xmin ymin xmax ymax 
#>    0    0    1    2 
#> attr(,"class")
#> [1] "bbox"
#> attr(,"z_range")
#> zmin zmax 
#>    0    3 
#> attr(,"class")
#> [1] "z_range"
sfc_multilinestring( obj = m, x = 2, y = 3, linestring_id = 1, multilinestring_id = 1 )
#> [[1]]
#> [[1]]
#>      [,1] [,2]
#> [1,]    0    0
#> [2,]    0    1
#> [3,]    1    1
#> 
#> attr(,"class")
#> [1] "XY"              "MULTILINESTRING" "sfg"            
#> 
#> [[2]]
#> [[1]]
#>      [,1] [,2]
#> [1,]    2    2
#> [2,]    2    3
#> 
#> attr(,"class")
#> [1] "XY"              "MULTILINESTRING" "sfg"            
#> 
#> attr(,"n_empty")
#> [1] 0
#> attr(,"crs")
#> $input
#> [1] NA
#> 
#> $wkt
#> [1] NA
#> 
#> attr(,"class")
#> [1] "crs"
#> attr(,"class")
#> [1] "sfc_MULTILINESTRING" "sfc"                
#> attr(,"precision")
#> [1] 0
#> attr(,"bbox")
#> xmin ymin xmax ymax 
#>    0    0    2    3 
#> attr(,"class")
#> [1] "bbox"

df <- data.frame(
  ml_id = c(1,1,1,1,1,1,1,1,2,2,2,2,2)
  , l_id = c(1,1,1,2,2,3,3,3,1,1,1,2,2)
  , x = rnorm(13)
  , y = rnorm(13)
  , z = rnorm(13)
  , m = rnorm(13)
)

sfc_multilinestring( obj = df, x = "x", y = "y")
#> [[1]]
#> [[1]]
#>             [,1]        [,2]
#>  [1,] -0.3703176 -1.11399124
#>  [2,] -0.2087384  0.53917943
#>  [3,] -0.3350529 -0.68633544
#>  [4,]  0.6801246 -1.10115239
#>  [5,]  0.3544453 -0.24730541
#>  [6,] -0.6983728 -0.73679035
#>  [7,] -0.9662909  1.01284679
#>  [8,] -0.2564673  0.22157536
#>  [9,]  0.6276523 -1.41090402
#> [10,]  0.5690997 -1.25799324
#> [11,]  0.3286720 -0.42695265
#> [12,] -0.4589994 -0.51574209
#> [13,] -0.6176589  0.09812316
#> 
#> attr(,"class")
#> [1] "XY"              "MULTILINESTRING" "sfg"            
#> 
#> attr(,"n_empty")
#> [1] 0
#> attr(,"crs")
#> $input
#> [1] NA
#> 
#> $wkt
#> [1] NA
#> 
#> attr(,"class")
#> [1] "crs"
#> attr(,"class")
#> [1] "sfc_MULTILINESTRING" "sfc"                
#> attr(,"precision")
#> [1] 0
#> attr(,"bbox")
#>       xmin       ymin       xmax       ymax 
#> -0.9662909 -1.4109040  0.6801246  1.0128468 
#> attr(,"class")
#> [1] "bbox"
sfc_multilinestring( obj = df, x = "x", y = "y", z = "z")
#> [[1]]
#> [[1]]
#>             [,1]        [,2]       [,3]
#>  [1,] -0.3703176 -1.11399124  2.0407415
#>  [2,] -0.2087384  0.53917943  0.8655794
#>  [3,] -0.3350529 -0.68633544  1.8283222
#>  [4,]  0.6801246 -1.10115239 -1.3416264
#>  [5,]  0.3544453 -0.24730541  2.0328684
#>  [6,] -0.6983728 -0.73679035  1.0299999
#>  [7,] -0.9662909  1.01284679 -1.5709782
#>  [8,] -0.2564673  0.22157536  0.2307289
#>  [9,]  0.6276523 -1.41090402  2.0294839
#> [10,]  0.5690997 -1.25799324  1.0252736
#> [11,]  0.3286720 -0.42695265  0.5063174
#> [12,] -0.4589994 -0.51574209 -1.4966278
#> [13,] -0.6176589  0.09812316  1.0722054
#> 
#> attr(,"class")
#> [1] "XYZ"             "MULTILINESTRING" "sfg"            
#> 
#> attr(,"n_empty")
#> [1] 0
#> attr(,"crs")
#> $input
#> [1] NA
#> 
#> $wkt
#> [1] NA
#> 
#> attr(,"class")
#> [1] "crs"
#> attr(,"class")
#> [1] "sfc_MULTILINESTRING" "sfc"                
#> attr(,"precision")
#> [1] 0
#> attr(,"bbox")
#>       xmin       ymin       xmax       ymax 
#> -0.9662909 -1.4109040  0.6801246  1.0128468 
#> attr(,"class")
#> [1] "bbox"
#> attr(,"z_range")
#>      zmin      zmax 
#> -1.570978  2.040742 
#> attr(,"class")
#> [1] "z_range"
sfc_multilinestring( obj = df, x = "x", y = "y", z = "z", m = "m")
#> [[1]]
#> [[1]]
#>             [,1]        [,2]       [,3]       [,4]
#>  [1,] -0.3703176 -1.11399124  2.0407415 -0.3827659
#>  [2,] -0.2087384  0.53917943  0.8655794 -1.0595030
#>  [3,] -0.3350529 -0.68633544  1.8283222  1.2831800
#>  [4,]  0.6801246 -1.10115239 -1.3416264  0.6857082
#>  [5,]  0.3544453 -0.24730541  2.0328684  1.2886374
#>  [6,] -0.6983728 -0.73679035  1.0299999 -0.2279991
#>  [7,] -0.9662909  1.01284679 -1.5709782  1.0411002
#>  [8,] -0.2564673  0.22157536  0.2307289  0.4514257
#>  [9,]  0.6276523 -1.41090402  2.0294839 -0.3534748
#> [10,]  0.5690997 -1.25799324  1.0252736  0.6959350
#> [11,]  0.3286720 -0.42695265  0.5063174  0.3204103
#> [12,] -0.4589994 -0.51574209 -1.4966278 -0.5574678
#> [13,] -0.6176589  0.09812316  1.0722054 -1.1369414
#> 
#> attr(,"class")
#> [1] "XYZM"            "MULTILINESTRING" "sfg"            
#> 
#> attr(,"n_empty")
#> [1] 0
#> attr(,"crs")
#> $input
#> [1] NA
#> 
#> $wkt
#> [1] NA
#> 
#> attr(,"class")
#> [1] "crs"
#> attr(,"class")
#> [1] "sfc_MULTILINESTRING" "sfc"                
#> attr(,"precision")
#> [1] 0
#> attr(,"bbox")
#>       xmin       ymin       xmax       ymax 
#> -0.9662909 -1.4109040  0.6801246  1.0128468 
#> attr(,"class")
#> [1] "bbox"
#> attr(,"z_range")
#>      zmin      zmax 
#> -1.570978  2.040742 
#> attr(,"class")
#> [1] "z_range"
#> attr(,"m_range")
#>      mmin      mmax 
#> -1.136941  1.288637 
#> attr(,"class")
#> [1] "m_range"

sfc_multilinestring( obj = df, x = 2, y = 3)
#> [[1]]
#> [[1]]
#>       [,1]       [,2]
#>  [1,]    1 -0.3703176
#>  [2,]    1 -0.2087384
#>  [3,]    1 -0.3350529
#>  [4,]    2  0.6801246
#>  [5,]    2  0.3544453
#>  [6,]    3 -0.6983728
#>  [7,]    3 -0.9662909
#>  [8,]    3 -0.2564673
#>  [9,]    1  0.6276523
#> [10,]    1  0.5690997
#> [11,]    1  0.3286720
#> [12,]    2 -0.4589994
#> [13,]    2 -0.6176589
#> 
#> attr(,"class")
#> [1] "XY"              "MULTILINESTRING" "sfg"            
#> 
#> attr(,"n_empty")
#> [1] 0
#> attr(,"crs")
#> $input
#> [1] NA
#> 
#> $wkt
#> [1] NA
#> 
#> attr(,"class")
#> [1] "crs"
#> attr(,"class")
#> [1] "sfc_MULTILINESTRING" "sfc"                
#> attr(,"precision")
#> [1] 0
#> attr(,"bbox")
#>       xmin       ymin       xmax       ymax 
#>  1.0000000 -0.9662909  3.0000000  0.6801246 
#> attr(,"class")
#> [1] "bbox"
sfc_multilinestring( obj = df, x = 2, y = 3, z = 4)
#> [[1]]
#> [[1]]
#>       [,1]       [,2]        [,3]
#>  [1,]    1 -0.3703176 -1.11399124
#>  [2,]    1 -0.2087384  0.53917943
#>  [3,]    1 -0.3350529 -0.68633544
#>  [4,]    2  0.6801246 -1.10115239
#>  [5,]    2  0.3544453 -0.24730541
#>  [6,]    3 -0.6983728 -0.73679035
#>  [7,]    3 -0.9662909  1.01284679
#>  [8,]    3 -0.2564673  0.22157536
#>  [9,]    1  0.6276523 -1.41090402
#> [10,]    1  0.5690997 -1.25799324
#> [11,]    1  0.3286720 -0.42695265
#> [12,]    2 -0.4589994 -0.51574209
#> [13,]    2 -0.6176589  0.09812316
#> 
#> attr(,"class")
#> [1] "XYZ"             "MULTILINESTRING" "sfg"            
#> 
#> attr(,"n_empty")
#> [1] 0
#> attr(,"crs")
#> $input
#> [1] NA
#> 
#> $wkt
#> [1] NA
#> 
#> attr(,"class")
#> [1] "crs"
#> attr(,"class")
#> [1] "sfc_MULTILINESTRING" "sfc"                
#> attr(,"precision")
#> [1] 0
#> attr(,"bbox")
#>       xmin       ymin       xmax       ymax 
#>  1.0000000 -0.9662909  3.0000000  0.6801246 
#> attr(,"class")
#> [1] "bbox"
#> attr(,"z_range")
#>      zmin      zmax 
#> -1.410904  1.012847 
#> attr(,"class")
#> [1] "z_range"
sfc_multilinestring( obj = df, x = 2, y = 3, z = 4, m = 5)
#> [[1]]
#> [[1]]
#>       [,1]       [,2]        [,3]       [,4]
#>  [1,]    1 -0.3703176 -1.11399124  2.0407415
#>  [2,]    1 -0.2087384  0.53917943  0.8655794
#>  [3,]    1 -0.3350529 -0.68633544  1.8283222
#>  [4,]    2  0.6801246 -1.10115239 -1.3416264
#>  [5,]    2  0.3544453 -0.24730541  2.0328684
#>  [6,]    3 -0.6983728 -0.73679035  1.0299999
#>  [7,]    3 -0.9662909  1.01284679 -1.5709782
#>  [8,]    3 -0.2564673  0.22157536  0.2307289
#>  [9,]    1  0.6276523 -1.41090402  2.0294839
#> [10,]    1  0.5690997 -1.25799324  1.0252736
#> [11,]    1  0.3286720 -0.42695265  0.5063174
#> [12,]    2 -0.4589994 -0.51574209 -1.4966278
#> [13,]    2 -0.6176589  0.09812316  1.0722054
#> 
#> attr(,"class")
#> [1] "XYZM"            "MULTILINESTRING" "sfg"            
#> 
#> attr(,"n_empty")
#> [1] 0
#> attr(,"crs")
#> $input
#> [1] NA
#> 
#> $wkt
#> [1] NA
#> 
#> attr(,"class")
#> [1] "crs"
#> attr(,"class")
#> [1] "sfc_MULTILINESTRING" "sfc"                
#> attr(,"precision")
#> [1] 0
#> attr(,"bbox")
#>       xmin       ymin       xmax       ymax 
#>  1.0000000 -0.9662909  3.0000000  0.6801246 
#> attr(,"class")
#> [1] "bbox"
#> attr(,"z_range")
#>      zmin      zmax 
#> -1.410904  1.012847 
#> attr(,"class")
#> [1] "z_range"
#> attr(,"m_range")
#>      mmin      mmax 
#> -1.570978  2.040742 
#> attr(,"class")
#> [1] "m_range"

sfc_multilinestring( obj = df, multilinestring_id = "ml_id", linestring_id = "l_id" )
#> [[1]]
#> [[1]]
#>            [,1]       [,2]      [,3]       [,4]
#> [1,] -0.3703176 -1.1139912 2.0407415 -0.3827659
#> [2,] -0.2087384  0.5391794 0.8655794 -1.0595030
#> [3,] -0.3350529 -0.6863354 1.8283222  1.2831800
#> 
#> [[2]]
#>           [,1]       [,2]      [,3]      [,4]
#> [1,] 0.6801246 -1.1011524 -1.341626 0.6857082
#> [2,] 0.3544453 -0.2473054  2.032868 1.2886374
#> 
#> [[3]]
#>            [,1]       [,2]       [,3]       [,4]
#> [1,] -0.6983728 -0.7367903  1.0299999 -0.2279991
#> [2,] -0.9662909  1.0128468 -1.5709782  1.0411002
#> [3,] -0.2564673  0.2215754  0.2307289  0.4514257
#> 
#> attr(,"class")
#> [1] "XYZM"            "MULTILINESTRING" "sfg"            
#> 
#> [[2]]
#> [[1]]
#>           [,1]       [,2]      [,3]       [,4]
#> [1,] 0.6276523 -1.4109040 2.0294839 -0.3534748
#> [2,] 0.5690997 -1.2579932 1.0252736  0.6959350
#> [3,] 0.3286720 -0.4269526 0.5063174  0.3204103
#> 
#> [[2]]
#>            [,1]        [,2]      [,3]       [,4]
#> [1,] -0.4589994 -0.51574209 -1.496628 -0.5574678
#> [2,] -0.6176589  0.09812316  1.072205 -1.1369414
#> 
#> attr(,"class")
#> [1] "XYZM"            "MULTILINESTRING" "sfg"            
#> 
#> attr(,"n_empty")
#> [1] 0
#> attr(,"crs")
#> $input
#> [1] NA
#> 
#> $wkt
#> [1] NA
#> 
#> attr(,"class")
#> [1] "crs"
#> attr(,"class")
#> [1] "sfc_MULTILINESTRING" "sfc"                
#> attr(,"precision")
#> [1] 0
#> attr(,"bbox")
#>       xmin       ymin       xmax       ymax 
#> -0.9662909 -1.4109040  0.6801246  1.0128468 
#> attr(,"class")
#> [1] "bbox"
#> attr(,"z_range")
#>      zmin      zmax 
#> -1.570978  2.040742 
#> attr(,"class")
#> [1] "z_range"
#> attr(,"m_range")
#>      mmin      mmax 
#> -1.136941  1.288637 
#> attr(,"class")
#> [1] "m_range"
sfc_multilinestring( obj = df, multilinestring_id = 1, linestring_id = 2 )
#> [[1]]
#> [[1]]
#>            [,1]       [,2]      [,3]       [,4]
#> [1,] -0.3703176 -1.1139912 2.0407415 -0.3827659
#> [2,] -0.2087384  0.5391794 0.8655794 -1.0595030
#> [3,] -0.3350529 -0.6863354 1.8283222  1.2831800
#> 
#> [[2]]
#>           [,1]       [,2]      [,3]      [,4]
#> [1,] 0.6801246 -1.1011524 -1.341626 0.6857082
#> [2,] 0.3544453 -0.2473054  2.032868 1.2886374
#> 
#> [[3]]
#>            [,1]       [,2]       [,3]       [,4]
#> [1,] -0.6983728 -0.7367903  1.0299999 -0.2279991
#> [2,] -0.9662909  1.0128468 -1.5709782  1.0411002
#> [3,] -0.2564673  0.2215754  0.2307289  0.4514257
#> 
#> attr(,"class")
#> [1] "XYZM"            "MULTILINESTRING" "sfg"            
#> 
#> [[2]]
#> [[1]]
#>           [,1]       [,2]      [,3]       [,4]
#> [1,] 0.6276523 -1.4109040 2.0294839 -0.3534748
#> [2,] 0.5690997 -1.2579932 1.0252736  0.6959350
#> [3,] 0.3286720 -0.4269526 0.5063174  0.3204103
#> 
#> [[2]]
#>            [,1]        [,2]      [,3]       [,4]
#> [1,] -0.4589994 -0.51574209 -1.496628 -0.5574678
#> [2,] -0.6176589  0.09812316  1.072205 -1.1369414
#> 
#> attr(,"class")
#> [1] "XYZM"            "MULTILINESTRING" "sfg"            
#> 
#> attr(,"n_empty")
#> [1] 0
#> attr(,"crs")
#> $input
#> [1] NA
#> 
#> $wkt
#> [1] NA
#> 
#> attr(,"class")
#> [1] "crs"
#> attr(,"class")
#> [1] "sfc_MULTILINESTRING" "sfc"                
#> attr(,"precision")
#> [1] 0
#> attr(,"bbox")
#>       xmin       ymin       xmax       ymax 
#> -0.9662909 -1.4109040  0.6801246  1.0128468 
#> attr(,"class")
#> [1] "bbox"
#> attr(,"z_range")
#>      zmin      zmax 
#> -1.570978  2.040742 
#> attr(,"class")
#> [1] "z_range"
#> attr(,"m_range")
#>      mmin      mmax 
#> -1.136941  1.288637 
#> attr(,"class")
#> [1] "m_range"