sf POLYGON

constructs an sf of POLYGON objects

sf_polygon(
  obj = NULL,
  x = NULL,
  y = NULL,
  z = NULL,
  m = NULL,
  polygon_id = NULL,
  linestring_id = NULL,
  close = TRUE,
  keep = FALSE,
  list_columns = NULL
)

Arguments

obj

sorted matrix or data.frame

x

x geometry column

y

y geometry column

z

z geometry column

m

m geometry column

polygon_id

column of ids for polygons

linestring_id

column of ids for lines (within polygons)

close

logical indicating whether polygons should be closed. If TRUE, all polygons will be checked and force closed if possible

keep

logical indicating if the non-geometry and non-id columns should be kept. if TRUE you must supply the geometry and id columns, and only the first row of each geometry is kept. See Keeping Properties.

list_columns

vector of column names to turn into a list.

Value

sf object of POLYGON geometries

notes

sfheaders functions do not perform any validity checks on the geometries. Nor do they set Coordinate Reference Systems, EPSG, PROJ4 or precision attributes.

The data.frame and matrices you send into the sfheader functions must be ordered.

Keeping Properties

Setting keep = TRUE will retain any columns not specified as a coordinate (x, y, z, m) or an id (e.g., linestring_id, polygon_id) of the input obj.

You can use list_columns to specify which of the properties will be turned into a list, thus keeping all the values in the column. For columns not specified in list_columns, only the first row of the column is kept

The sf_* functions assume the input obj is a long data.frame / matrix, where any properties are repeated down the table for the same geometry.

Examples

m <- matrix(c(0,0,0,0,1,1), ncol = 2 )
sf_polygon( m )
#>   id               geometry
#> 1  1 0, 0, 0, 0, 0, 1, 1, 0

m <- matrix(c(0,0,0,0,0,1,0,1,1,1,2,2,1,2,3,1,3,4), ncol = 3, byrow = TRUE)
sf_polygon( obj = m )
#>   id                                                      geometry
#> 1  1 0, 0, 0, 1, 1, 1, 0, 0, 0, 1, 2, 2, 3, 0, 0, 1, 1, 2, 3, 4, 0
sf_polygon( obj = m, polygon_id = 1 )
#>   id               geometry
#> 1  0 0, 0, 1, 0, 0, 1, 1, 0
#> 2  1 2, 2, 3, 2, 2, 3, 4, 2
sf_polygon( obj = m, linestring_id = 1 )
#>   id                                       geometry
#> 1  1 0, 0, 1, 0, 0, 1, 1, 0, 2, 2, 3, 2, 2, 3, 4, 2

sf_polygon( obj = m, linestring_id = 1, polygon_id = 1 )
#>   id               geometry
#> 1  0 0, 0, 1, 0, 0, 1, 1, 0
#> 2  1 2, 2, 3, 2, 2, 3, 4, 2

sf_polygon( obj = m, x = 2, y = 3 )
#>   id                                 geometry
#> 1  1 0, 0, 1, 2, 2, 3, 0, 0, 1, 1, 2, 3, 4, 0
sf_polygon( obj = m, x = 1, y = 2, z = 3 )
#>   id                                                      geometry
#> 1  1 0, 0, 0, 1, 1, 1, 0, 0, 0, 1, 2, 2, 3, 0, 0, 1, 1, 2, 3, 4, 0
sf_polygon( obj = m, x = 2, y = 3, linestring_id = 1, polygon_id = 1 )
#>   id               geometry
#> 1  0 0, 0, 1, 0, 0, 1, 1, 0
#> 2  1 2, 2, 3, 2, 2, 3, 4, 2

df <- data.frame(
  ml_id = c(1,1,1,1,1,1,1,1,1,2,2,2,2,2,2)
  , l_id = c(1,1,1,2,2,2,3,3,3,1,1,1,2,2,2)
  , x = rnorm(15)
  , y = rnorm(15)
  , z = rnorm(15)
  , m = rnorm(15)
)

sf_polygon( obj = df, x = "x", y = "y")
#>   id
#> 1  1
#>                                                                                                                                                                                                                                                                                                                                                                                                           geometry
#> 1 -1.38920112, 1.21060858, 0.50855363, -0.30353507, -0.02424957, 0.74218576, -0.43510645, 1.09316323, -1.35569833, -0.28705576, 0.47956232, 0.13415949, 0.56276493, 0.25351181, 1.38470512, -1.38920112, -0.55381864, 0.24841391, 0.75742694, -0.18190104, 1.49513834, -0.62304023, -0.23497217, -1.19594019, -0.83227576, -0.79387711, -0.81678475, -0.36835961, 0.28523840, 1.87091345, -1.45671432, -0.55381864
sf_polygon( obj = df, x = "x", y = "y", z = "z")
#>   id
#> 1  1
#>                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                  geometry
#> 1 -1.38920112, 1.21060858, 0.50855363, -0.30353507, -0.02424957, 0.74218576, -0.43510645, 1.09316323, -1.35569833, -0.28705576, 0.47956232, 0.13415949, 0.56276493, 0.25351181, 1.38470512, -1.38920112, -0.55381864, 0.24841391, 0.75742694, -0.18190104, 1.49513834, -0.62304023, -0.23497217, -1.19594019, -0.83227576, -0.79387711, -0.81678475, -0.36835961, 0.28523840, 1.87091345, -1.45671432, -0.55381864, 0.63996999, -0.32752354, 0.68055015, 0.13940139, -0.72212165, 0.13073376, 0.77703763, -0.57328698, -1.28111810, -1.32525322, 0.32570487, -0.42373633, 0.27151726, 2.43702521, -1.78919934, 0.63996999
sf_polygon( obj = df, x = "x", y = "y", z = "z", m = "m")
#>   id
#> 1  1
#>                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                         geometry
#> 1 -1.38920112, 1.21060858, 0.50855363, -0.30353507, -0.02424957, 0.74218576, -0.43510645, 1.09316323, -1.35569833, -0.28705576, 0.47956232, 0.13415949, 0.56276493, 0.25351181, 1.38470512, -1.38920112, -0.55381864, 0.24841391, 0.75742694, -0.18190104, 1.49513834, -0.62304023, -0.23497217, -1.19594019, -0.83227576, -0.79387711, -0.81678475, -0.36835961, 0.28523840, 1.87091345, -1.45671432, -0.55381864, 0.63996999, -0.32752354, 0.68055015, 0.13940139, -0.72212165, 0.13073376, 0.77703763, -0.57328698, -1.28111810, -1.32525322, 0.32570487, -0.42373633, 0.27151726, 2.43702521, -1.78919934, 0.63996999, 0.21704851, 0.05074869, 2.87420308, 2.47810674, 1.37098905, 0.13216352, -1.28580779, -0.02919050, 0.40364567, -1.13282009, 0.04404341, -0.47216248, -0.68302064, -0.31160084, -0.15175599, 0.21704851

sf_polygon( obj = df, x = 2, y = 3)
#>   id
#> 1  1
#>                                                                                                                                                                                                                                                                                                                                                                                                geometry
#> 1 1.00000000, 1.00000000, 1.00000000, 2.00000000, 2.00000000, 2.00000000, 3.00000000, 3.00000000, 3.00000000, 1.00000000, 1.00000000, 1.00000000, 2.00000000, 2.00000000, 2.00000000, 1.00000000, -1.38920112, 1.21060858, 0.50855363, -0.30353507, -0.02424957, 0.74218576, -0.43510645, 1.09316323, -1.35569833, -0.28705576, 0.47956232, 0.13415949, 0.56276493, 0.25351181, 1.38470512, -1.38920112
sf_polygon( obj = df, x = 2, y = 3, z = 4)
#>   id
#> 1  1
#>                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                           geometry
#> 1 1.00000000, 1.00000000, 1.00000000, 2.00000000, 2.00000000, 2.00000000, 3.00000000, 3.00000000, 3.00000000, 1.00000000, 1.00000000, 1.00000000, 2.00000000, 2.00000000, 2.00000000, 1.00000000, -1.38920112, 1.21060858, 0.50855363, -0.30353507, -0.02424957, 0.74218576, -0.43510645, 1.09316323, -1.35569833, -0.28705576, 0.47956232, 0.13415949, 0.56276493, 0.25351181, 1.38470512, -1.38920112, -0.55381864, 0.24841391, 0.75742694, -0.18190104, 1.49513834, -0.62304023, -0.23497217, -1.19594019, -0.83227576, -0.79387711, -0.81678475, -0.36835961, 0.28523840, 1.87091345, -1.45671432, -0.55381864
sf_polygon( obj = df, x = 2, y = 3, z = 4, m = 5)
#>   id
#> 1  1
#>                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                  geometry
#> 1 1.00000000, 1.00000000, 1.00000000, 2.00000000, 2.00000000, 2.00000000, 3.00000000, 3.00000000, 3.00000000, 1.00000000, 1.00000000, 1.00000000, 2.00000000, 2.00000000, 2.00000000, 1.00000000, -1.38920112, 1.21060858, 0.50855363, -0.30353507, -0.02424957, 0.74218576, -0.43510645, 1.09316323, -1.35569833, -0.28705576, 0.47956232, 0.13415949, 0.56276493, 0.25351181, 1.38470512, -1.38920112, -0.55381864, 0.24841391, 0.75742694, -0.18190104, 1.49513834, -0.62304023, -0.23497217, -1.19594019, -0.83227576, -0.79387711, -0.81678475, -0.36835961, 0.28523840, 1.87091345, -1.45671432, -0.55381864, 0.63996999, -0.32752354, 0.68055015, 0.13940139, -0.72212165, 0.13073376, 0.77703763, -0.57328698, -1.28111810, -1.32525322, 0.32570487, -0.42373633, 0.27151726, 2.43702521, -1.78919934, 0.63996999

sf_polygon( obj = df, polygon_id = "ml_id", linestring_id = "l_id" )
#>   ml_id
#> 1     1
#> 2     2
#>                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                 geometry
#> 1 -1.38920112, 1.21060858, 0.50855363, -1.38920112, -0.55381864, 0.24841391, 0.75742694, -0.55381864, 0.63996999, -0.32752354, 0.68055015, 0.63996999, 0.21704851, 0.05074869, 2.87420308, 0.21704851, -0.30353507, -0.02424957, 0.74218576, -0.30353507, -0.18190104, 1.49513834, -0.62304023, -0.18190104, 0.13940139, -0.72212165, 0.13073376, 0.13940139, 2.47810674, 1.37098905, 0.13216352, 2.47810674, -0.43510645, 1.09316323, -1.35569833, -0.43510645, -0.23497217, -1.19594019, -0.83227576, -0.23497217, 0.77703763, -0.57328698, -1.28111810, 0.77703763, -1.28580779, -0.02919050, 0.40364567, -1.28580779
#> 2                                                                                                                                                                                                       -0.28705576, 0.47956232, 0.13415949, -0.28705576, -0.79387711, -0.81678475, -0.36835961, -0.79387711, -1.32525322, 0.32570487, -0.42373633, -1.32525322, -1.13282009, 0.04404341, -0.47216248, -1.13282009, 0.56276493, 0.25351181, 1.38470512, 0.56276493, 0.28523840, 1.87091345, -1.45671432, 0.28523840, 0.27151726, 2.43702521, -1.78919934, 0.27151726, -0.68302064, -0.31160084, -0.15175599, -0.68302064
sf_polygon( obj = df, polygon_id = 1, linestring_id = 2 )
#>   id
#> 1  1
#> 2  2
#>                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                 geometry
#> 1 -1.38920112, 1.21060858, 0.50855363, -1.38920112, -0.55381864, 0.24841391, 0.75742694, -0.55381864, 0.63996999, -0.32752354, 0.68055015, 0.63996999, 0.21704851, 0.05074869, 2.87420308, 0.21704851, -0.30353507, -0.02424957, 0.74218576, -0.30353507, -0.18190104, 1.49513834, -0.62304023, -0.18190104, 0.13940139, -0.72212165, 0.13073376, 0.13940139, 2.47810674, 1.37098905, 0.13216352, 2.47810674, -0.43510645, 1.09316323, -1.35569833, -0.43510645, -0.23497217, -1.19594019, -0.83227576, -0.23497217, 0.77703763, -0.57328698, -1.28111810, 0.77703763, -1.28580779, -0.02919050, 0.40364567, -1.28580779
#> 2                                                                                                                                                                                                       -0.28705576, 0.47956232, 0.13415949, -0.28705576, -0.79387711, -0.81678475, -0.36835961, -0.79387711, -1.32525322, 0.32570487, -0.42373633, -1.32525322, -1.13282009, 0.04404341, -0.47216248, -1.13282009, 0.56276493, 0.25351181, 1.38470512, 0.56276493, 0.28523840, 1.87091345, -1.45671432, 0.28523840, 0.27151726, 2.43702521, -1.78919934, 0.27151726, -0.68302064, -0.31160084, -0.15175599, -0.68302064

## keeping properties
df <- data.frame(
  ml_id = c(1,1,1,1,1,1,1,1,1,2,2,2,2,2,2)
  , l_id = c(1,1,1,2,2,2,3,3,3,1,1,1,2,2,2)
  , x = rnorm(15)
  , y = rnorm(15)
  , z = rnorm(15)
  , m = rnorm(15)
  , val = letters[1:15]
  , stringsAsFactors = FALSE
)

## using keep = TRUE means the first row of all non-geometries are kept
sf_polygon(
  obj = df
  , polygon_id = "ml_id"
  , linestring_id = "l_id"
  , x = "x"
  , y = "y"
  , keep = TRUE
)
#>   ml_id          z         m val
#> 1     1 -1.9969618 0.1149350   a
#> 2     2 -0.3519723 0.3648937   j
#>                                                                                                                                                                                                                                                                                                  geometry
#> 1 0.97567801, 0.28317341, -0.13234516, 0.97567801, 0.34604333, -0.74048976, 0.72323628, 0.34604333, 1.39524598, 1.44327743, 0.40862458, 1.39524598, 0.56076103, 0.04131937, 3.25100676, 0.56076103, -0.08745764, -0.38371878, 1.17138831, -0.08745764, -0.10792283, -1.58394860, -1.59117428, -0.10792283
#> 2                                                                                                   1.13939892, -1.64946283, 1.30956415, 1.13939892, 1.03963157, 0.60569028, -0.75035974, 1.03963157, 2.17960575, 0.66938374, -0.01856669, 2.17960575, -1.36956062, -0.80844892, -2.02330037, -1.36956062

## use 'list_column' to specify columns where you want to keep all the values
sf_polygon(
  obj = df
  , polygon_id = "ml_id"
  , linestring_id = "l_id"
  , x = "x"
  , y = "y"
  , keep = TRUE
  , list_columns = "val"
)
#>   ml_id          z         m                                val
#> 1     1 -1.9969618 0.1149350 a, b, c, a, d, e, f, d, g, h, i, g
#> 2     2 -0.3519723 0.3648937             j, k, l, j, m, n, o, m
#>                                                                                                                                                                                                                                                                                                  geometry
#> 1 0.97567801, 0.28317341, -0.13234516, 0.97567801, 0.34604333, -0.74048976, 0.72323628, 0.34604333, 1.39524598, 1.44327743, 0.40862458, 1.39524598, 0.56076103, 0.04131937, 3.25100676, 0.56076103, -0.08745764, -0.38371878, 1.17138831, -0.08745764, -0.10792283, -1.58394860, -1.59117428, -0.10792283
#> 2                                                                                                   1.13939892, -1.64946283, 1.30956415, 1.13939892, 1.03963157, 0.60569028, -0.75035974, 1.03963157, 2.17960575, 0.66938374, -0.01856669, 2.17960575, -1.36956062, -0.80844892, -2.02330037, -1.36956062