sf POLYGON

constructs an sf of POLYGON objects

sf_polygon(
  obj = NULL,
  x = NULL,
  y = NULL,
  z = NULL,
  m = NULL,
  polygon_id = NULL,
  linestring_id = NULL,
  close = TRUE,
  keep = FALSE,
  list_columns = NULL
)

Arguments

obj

sorted matrix or data.frame

x

x geometry column

y

y geometry column

z

z geometry column

m

m geometry column

polygon_id

column of ids for polygons

linestring_id

column of ids for lines (within polygons)

close

logical indicating whether polygons should be closed. If TRUE, all polygons will be checked and force closed if possible

keep

logical indicating if the non-geometry and non-id columns should be kept. if TRUE you must supply the geometry and id columns, and only the first row of each geometry is kept. See Keeping Properties.

list_columns

vector of column names to turn into a list.

Value

sf object of POLYGON geometries

notes

sfheaders functions do not perform any validity checks on the geometries. Nor do they set Coordinate Reference Systems, EPSG, PROJ4 or precision attributes.

The data.frame and matrices you send into the sfheader functions must be ordered.

Keeping Properties

Setting keep = TRUE will retain any columns not specified as a coordinate (x, y, z, m) or an id (e.g., linestring_id, polygon_id) of the input obj.

You can use list_columns to specify which of the properties will be turned into a list, thus keeping all the values in the column. For columns not specified in list_columns, only the first row of the column is kept

The sf_* functions assume the input obj is a long data.frame / matrix, where any properties are repeated down the table for the same geometry.

Examples

m <- matrix(c(0,0,0,0,1,1), ncol = 2 )
sf_polygon( m )
#>   id               geometry
#> 1  1 0, 0, 0, 0, 0, 1, 1, 0

m <- matrix(c(0,0,0,0,0,1,0,1,1,1,2,2,1,2,3,1,3,4), ncol = 3, byrow = TRUE)
sf_polygon( obj = m )
#>   id                                                      geometry
#> 1  1 0, 0, 0, 1, 1, 1, 0, 0, 0, 1, 2, 2, 3, 0, 0, 1, 1, 2, 3, 4, 0
sf_polygon( obj = m, polygon_id = 1 )
#>   id               geometry
#> 1  0 0, 0, 1, 0, 0, 1, 1, 0
#> 2  1 2, 2, 3, 2, 2, 3, 4, 2
sf_polygon( obj = m, linestring_id = 1 )
#>   id                                       geometry
#> 1  1 0, 0, 1, 0, 0, 1, 1, 0, 2, 2, 3, 2, 2, 3, 4, 2

sf_polygon( obj = m, linestring_id = 1, polygon_id = 1 )
#>   id               geometry
#> 1  0 0, 0, 1, 0, 0, 1, 1, 0
#> 2  1 2, 2, 3, 2, 2, 3, 4, 2

sf_polygon( obj = m, x = 2, y = 3 )
#>   id                                 geometry
#> 1  1 0, 0, 1, 2, 2, 3, 0, 0, 1, 1, 2, 3, 4, 0
sf_polygon( obj = m, x = 1, y = 2, z = 3 )
#>   id                                                      geometry
#> 1  1 0, 0, 0, 1, 1, 1, 0, 0, 0, 1, 2, 2, 3, 0, 0, 1, 1, 2, 3, 4, 0
sf_polygon( obj = m, x = 2, y = 3, linestring_id = 1, polygon_id = 1 )
#>   id               geometry
#> 1  0 0, 0, 1, 0, 0, 1, 1, 0
#> 2  1 2, 2, 3, 2, 2, 3, 4, 2

df <- data.frame(
  ml_id = c(1,1,1,1,1,1,1,1,1,2,2,2,2,2,2)
  , l_id = c(1,1,1,2,2,2,3,3,3,1,1,1,2,2,2)
  , x = rnorm(15)
  , y = rnorm(15)
  , z = rnorm(15)
  , m = rnorm(15)
)

sf_polygon( obj = df, x = "x", y = "y")
#>   id
#> 1  1
#>                                                                                                                                                                                                                                                                                                                                                                                                     geometry
#> 1 1.21247490, -1.41826660, 1.74270602, 0.73443997, -1.35680744, 0.65004819, -0.64991480, -1.07936556, -0.94936085, -0.24992486, -0.03058543, 0.11489022, 0.47103515, 0.17483045, 0.97655857, 1.21247490, 0.19047465, -0.78868167, 0.12990042, -0.58685313, -0.33296462, 0.23908091, 0.85384008, -0.30024815, 1.83882062, 0.65534114, 0.44107155, 1.58001310, 0.86657305, -0.17137848, 0.94374944, 0.19047465
sf_polygon( obj = df, x = "x", y = "y", z = "z")
#>   id
#> 1  1
#>                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                          geometry
#> 1 1.21247490, -1.41826660, 1.74270602, 0.73443997, -1.35680744, 0.65004819, -0.64991480, -1.07936556, -0.94936085, -0.24992486, -0.03058543, 0.11489022, 0.47103515, 0.17483045, 0.97655857, 1.21247490, 0.19047465, -0.78868167, 0.12990042, -0.58685313, -0.33296462, 0.23908091, 0.85384008, -0.30024815, 1.83882062, 0.65534114, 0.44107155, 1.58001310, 0.86657305, -0.17137848, 0.94374944, 0.19047465, 0.83024651, 1.83890251, -0.06170319, -2.34197885, 0.80880488, -1.58844179, 1.41451648, 0.58841211, 1.59154816, 0.40137976, -0.28771273, -0.22694538, 0.67911834, 0.26293040, 0.52953513, 0.83024651
sf_polygon( obj = df, x = "x", y = "y", z = "z", m = "m")
#>   id
#> 1  1
#>                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                    geometry
#> 1 1.21247490, -1.41826660, 1.74270602, 0.73443997, -1.35680744, 0.65004819, -0.64991480, -1.07936556, -0.94936085, -0.24992486, -0.03058543, 0.11489022, 0.47103515, 0.17483045, 0.97655857, 1.21247490, 0.19047465, -0.78868167, 0.12990042, -0.58685313, -0.33296462, 0.23908091, 0.85384008, -0.30024815, 1.83882062, 0.65534114, 0.44107155, 1.58001310, 0.86657305, -0.17137848, 0.94374944, 0.19047465, 0.83024651, 1.83890251, -0.06170319, -2.34197885, 0.80880488, -1.58844179, 1.41451648, 0.58841211, 1.59154816, 0.40137976, -0.28771273, -0.22694538, 0.67911834, 0.26293040, 0.52953513, 0.83024651, -1.51996012, 0.33358634, -0.43046014, -1.37665389, -1.33228743, 0.84861914, -0.33118354, 0.90832228, -1.34235619, 0.40687913, -0.81297877, -2.15152702, 0.18522604, 0.23035971, -0.67776483, -1.51996012

sf_polygon( obj = df, x = 2, y = 3)
#>   id
#> 1  1
#>                                                                                                                                                                                                                                                                                                                                                                                                geometry
#> 1 1.00000000, 1.00000000, 1.00000000, 2.00000000, 2.00000000, 2.00000000, 3.00000000, 3.00000000, 3.00000000, 1.00000000, 1.00000000, 1.00000000, 2.00000000, 2.00000000, 2.00000000, 1.00000000, 1.21247490, -1.41826660, 1.74270602, 0.73443997, -1.35680744, 0.65004819, -0.64991480, -1.07936556, -0.94936085, -0.24992486, -0.03058543, 0.11489022, 0.47103515, 0.17483045, 0.97655857, 1.21247490
sf_polygon( obj = df, x = 2, y = 3, z = 4)
#>   id
#> 1  1
#>                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                     geometry
#> 1 1.00000000, 1.00000000, 1.00000000, 2.00000000, 2.00000000, 2.00000000, 3.00000000, 3.00000000, 3.00000000, 1.00000000, 1.00000000, 1.00000000, 2.00000000, 2.00000000, 2.00000000, 1.00000000, 1.21247490, -1.41826660, 1.74270602, 0.73443997, -1.35680744, 0.65004819, -0.64991480, -1.07936556, -0.94936085, -0.24992486, -0.03058543, 0.11489022, 0.47103515, 0.17483045, 0.97655857, 1.21247490, 0.19047465, -0.78868167, 0.12990042, -0.58685313, -0.33296462, 0.23908091, 0.85384008, -0.30024815, 1.83882062, 0.65534114, 0.44107155, 1.58001310, 0.86657305, -0.17137848, 0.94374944, 0.19047465
sf_polygon( obj = df, x = 2, y = 3, z = 4, m = 5)
#>   id
#> 1  1
#>                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                          geometry
#> 1 1.00000000, 1.00000000, 1.00000000, 2.00000000, 2.00000000, 2.00000000, 3.00000000, 3.00000000, 3.00000000, 1.00000000, 1.00000000, 1.00000000, 2.00000000, 2.00000000, 2.00000000, 1.00000000, 1.21247490, -1.41826660, 1.74270602, 0.73443997, -1.35680744, 0.65004819, -0.64991480, -1.07936556, -0.94936085, -0.24992486, -0.03058543, 0.11489022, 0.47103515, 0.17483045, 0.97655857, 1.21247490, 0.19047465, -0.78868167, 0.12990042, -0.58685313, -0.33296462, 0.23908091, 0.85384008, -0.30024815, 1.83882062, 0.65534114, 0.44107155, 1.58001310, 0.86657305, -0.17137848, 0.94374944, 0.19047465, 0.83024651, 1.83890251, -0.06170319, -2.34197885, 0.80880488, -1.58844179, 1.41451648, 0.58841211, 1.59154816, 0.40137976, -0.28771273, -0.22694538, 0.67911834, 0.26293040, 0.52953513, 0.83024651

sf_polygon( obj = df, polygon_id = "ml_id", linestring_id = "l_id" )
#>   ml_id
#> 1     1
#> 2     2
#>                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                 geometry
#> 1 1.21247490, -1.41826660, 1.74270602, 1.21247490, 0.19047465, -0.78868167, 0.12990042, 0.19047465, 0.83024651, 1.83890251, -0.06170319, 0.83024651, -1.51996012, 0.33358634, -0.43046014, -1.51996012, 0.73443997, -1.35680744, 0.65004819, 0.73443997, -0.58685313, -0.33296462, 0.23908091, -0.58685313, -2.34197885, 0.80880488, -1.58844179, -2.34197885, -1.37665389, -1.33228743, 0.84861914, -1.37665389, -0.64991480, -1.07936556, -0.94936085, -0.64991480, 0.85384008, -0.30024815, 1.83882062, 0.85384008, 1.41451648, 0.58841211, 1.59154816, 1.41451648, -0.33118354, 0.90832228, -1.34235619, -0.33118354
#> 2                                                                                                                                                                                                                -0.24992486, -0.03058543, 0.11489022, -0.24992486, 0.65534114, 0.44107155, 1.58001310, 0.65534114, 0.40137976, -0.28771273, -0.22694538, 0.40137976, 0.40687913, -0.81297877, -2.15152702, 0.40687913, 0.47103515, 0.17483045, 0.97655857, 0.47103515, 0.86657305, -0.17137848, 0.94374944, 0.86657305, 0.67911834, 0.26293040, 0.52953513, 0.67911834, 0.18522604, 0.23035971, -0.67776483, 0.18522604
sf_polygon( obj = df, polygon_id = 1, linestring_id = 2 )
#>   id
#> 1  1
#> 2  2
#>                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                 geometry
#> 1 1.21247490, -1.41826660, 1.74270602, 1.21247490, 0.19047465, -0.78868167, 0.12990042, 0.19047465, 0.83024651, 1.83890251, -0.06170319, 0.83024651, -1.51996012, 0.33358634, -0.43046014, -1.51996012, 0.73443997, -1.35680744, 0.65004819, 0.73443997, -0.58685313, -0.33296462, 0.23908091, -0.58685313, -2.34197885, 0.80880488, -1.58844179, -2.34197885, -1.37665389, -1.33228743, 0.84861914, -1.37665389, -0.64991480, -1.07936556, -0.94936085, -0.64991480, 0.85384008, -0.30024815, 1.83882062, 0.85384008, 1.41451648, 0.58841211, 1.59154816, 1.41451648, -0.33118354, 0.90832228, -1.34235619, -0.33118354
#> 2                                                                                                                                                                                                                -0.24992486, -0.03058543, 0.11489022, -0.24992486, 0.65534114, 0.44107155, 1.58001310, 0.65534114, 0.40137976, -0.28771273, -0.22694538, 0.40137976, 0.40687913, -0.81297877, -2.15152702, 0.40687913, 0.47103515, 0.17483045, 0.97655857, 0.47103515, 0.86657305, -0.17137848, 0.94374944, 0.86657305, 0.67911834, 0.26293040, 0.52953513, 0.67911834, 0.18522604, 0.23035971, -0.67776483, 0.18522604

## keeping properties
df <- data.frame(
  ml_id = c(1,1,1,1,1,1,1,1,1,2,2,2,2,2,2)
  , l_id = c(1,1,1,2,2,2,3,3,3,1,1,1,2,2,2)
  , x = rnorm(15)
  , y = rnorm(15)
  , z = rnorm(15)
  , m = rnorm(15)
  , val = letters[1:15]
  , stringsAsFactors = FALSE
)

## using keep = TRUE means the first row of all non-geometries are kept
sf_polygon(
  obj = df
  , polygon_id = "ml_id"
  , linestring_id = "l_id"
  , x = "x"
  , y = "y"
  , keep = TRUE
)
#>   ml_id         z          m val
#> 1     1 0.9808348 -0.0154752   a
#> 2     2 0.4938913 -1.5757360   j
#>                                                                                                                                                                                                                                                                                                        geometry
#> 1 -0.56470626, 0.03094559, -0.68515953, -0.56470626, -0.05549116, -0.39502507, 0.57938255, -0.05549116, -0.14914027, -1.51952020, -0.13399760, -0.14914027, 0.16303184, -1.21185963, 1.79608356, 0.16303184, 1.41429228, 0.85674742, -1.42585084, 1.41429228, -0.15506839, -0.05385049, 1.22162006, -0.15506839
#> 2                                                                                                      -0.44549755, -0.02580986, 0.90414229, -0.44549755, -0.94158065, 0.67343772, 0.29399370, -0.94158065, -0.80682210, -1.40701029, -0.08162283, -0.80682210, 0.46894593, 1.35695955, -0.31795078, 0.46894593

## use 'list_column' to specify columns where you want to keep all the values
sf_polygon(
  obj = df
  , polygon_id = "ml_id"
  , linestring_id = "l_id"
  , x = "x"
  , y = "y"
  , keep = TRUE
  , list_columns = "val"
)
#>   ml_id         z          m                                val
#> 1     1 0.9808348 -0.0154752 a, b, c, a, d, e, f, d, g, h, i, g
#> 2     2 0.4938913 -1.5757360             j, k, l, j, m, n, o, m
#>                                                                                                                                                                                                                                                                                                        geometry
#> 1 -0.56470626, 0.03094559, -0.68515953, -0.56470626, -0.05549116, -0.39502507, 0.57938255, -0.05549116, -0.14914027, -1.51952020, -0.13399760, -0.14914027, 0.16303184, -1.21185963, 1.79608356, 0.16303184, 1.41429228, 0.85674742, -1.42585084, 1.41429228, -0.15506839, -0.05385049, 1.22162006, -0.15506839
#> 2                                                                                                      -0.44549755, -0.02580986, 0.90414229, -0.44549755, -0.94158065, 0.67343772, 0.29399370, -0.94158065, -0.80682210, -1.40701029, -0.08162283, -0.80682210, 0.46894593, 1.35695955, -0.31795078, 0.46894593