sf POLYGON

constructs an sf of POLYGON objects

sf_polygon(
  obj = NULL,
  x = NULL,
  y = NULL,
  z = NULL,
  m = NULL,
  polygon_id = NULL,
  linestring_id = NULL,
  close = TRUE,
  keep = FALSE,
  list_columns = NULL
)

Arguments

obj

sorted matrix or data.frame

x

x geometry column

y

y geometry column

z

z geometry column

m

m geometry column

polygon_id

column of ids for polygons

linestring_id

column of ids for lines (within polygons)

close

logical indicating whether polygons should be closed. If TRUE, all polygons will be checked and force closed if possible

keep

logical indicating if the non-geometry and non-id columns should be kept. if TRUE you must supply the geometry and id columns, and only the first row of each geometry is kept. See Keeping Properties.

list_columns

vector of column names to turn into a list.

Value

sf object of POLYGON geometries

notes

sfheaders functions do not perform any validity checks on the geometries. Nor do they set Coordinate Reference Systems, EPSG, PROJ4 or precision attributes.

The data.frame and matrices you send into the sfheader functions must be ordered.

Keeping Properties

Setting keep = TRUE will retain any columns not specified as a coordinate (x, y, z, m) or an id (e.g., linestring_id, polygon_id) of the input obj.

You can use list_columns to specify which of the properties will be turned into a list, thus keeping all the values in the column. For columns not specified in list_columns, only the first row of the column is kept

The sf_* functions assume the input obj is a long data.frame / matrix, where any properties are repeated down the table for the same geometry.

Examples

m <- matrix(c(0,0,0,0,1,1), ncol = 2 )
sf_polygon( m )
#>   id               geometry
#> 1  1 0, 0, 0, 0, 0, 1, 1, 0

m <- matrix(c(0,0,0,0,0,1,0,1,1,1,2,2,1,2,3,1,3,4), ncol = 3, byrow = TRUE)
sf_polygon( obj = m )
#>   id                                                      geometry
#> 1  1 0, 0, 0, 1, 1, 1, 0, 0, 0, 1, 2, 2, 3, 0, 0, 1, 1, 2, 3, 4, 0
sf_polygon( obj = m, polygon_id = 1 )
#>   id               geometry
#> 1  0 0, 0, 1, 0, 0, 1, 1, 0
#> 2  1 2, 2, 3, 2, 2, 3, 4, 2
sf_polygon( obj = m, linestring_id = 1 )
#>   id                                       geometry
#> 1  1 0, 0, 1, 0, 0, 1, 1, 0, 2, 2, 3, 2, 2, 3, 4, 2

sf_polygon( obj = m, linestring_id = 1, polygon_id = 1 )
#>   id               geometry
#> 1  0 0, 0, 1, 0, 0, 1, 1, 0
#> 2  1 2, 2, 3, 2, 2, 3, 4, 2

sf_polygon( obj = m, x = 2, y = 3 )
#>   id                                 geometry
#> 1  1 0, 0, 1, 2, 2, 3, 0, 0, 1, 1, 2, 3, 4, 0
sf_polygon( obj = m, x = 1, y = 2, z = 3 )
#>   id                                                      geometry
#> 1  1 0, 0, 0, 1, 1, 1, 0, 0, 0, 1, 2, 2, 3, 0, 0, 1, 1, 2, 3, 4, 0
sf_polygon( obj = m, x = 2, y = 3, linestring_id = 1, polygon_id = 1 )
#>   id               geometry
#> 1  0 0, 0, 1, 0, 0, 1, 1, 0
#> 2  1 2, 2, 3, 2, 2, 3, 4, 2

df <- data.frame(
  ml_id = c(1,1,1,1,1,1,1,1,1,2,2,2,2,2,2)
  , l_id = c(1,1,1,2,2,2,3,3,3,1,1,1,2,2,2)
  , x = rnorm(15)
  , y = rnorm(15)
  , z = rnorm(15)
  , m = rnorm(15)
)

sf_polygon( obj = df, x = "x", y = "y")
#>   id
#> 1  1
#>                                                                                                                                                                                                                                                                                                                                                                                                        geometry
#> 1 -0.15536034, 0.53727788, -0.67173324, -0.98300682, 0.18652556, -1.26965240, 0.32054519, 0.22250500, -0.27421351, -0.12969106, 0.38133631, -1.13303327, 0.85822972, -0.61313170, 0.03401142, -0.15536034, -0.32235681, -0.22117590, 0.41289479, 0.13943855, -0.71901411, 0.80917276, -1.20710216, 0.93721277, 0.35784235, 1.91266220, -0.09474154, 0.86517265, 1.37992257, 1.83290264, 0.64831966, -0.32235681
sf_polygon( obj = df, x = "x", y = "y", z = "z")
#>   id
#> 1  1
#>                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                geometry
#> 1 -0.15536034, 0.53727788, -0.67173324, -0.98300682, 0.18652556, -1.26965240, 0.32054519, 0.22250500, -0.27421351, -0.12969106, 0.38133631, -1.13303327, 0.85822972, -0.61313170, 0.03401142, -0.15536034, -0.32235681, -0.22117590, 0.41289479, 0.13943855, -0.71901411, 0.80917276, -1.20710216, 0.93721277, 0.35784235, 1.91266220, -0.09474154, 0.86517265, 1.37992257, 1.83290264, 0.64831966, -0.32235681, -1.01462665, -1.73160670, 1.51397315, 0.06310170, 1.30454400, 0.64039856, 1.48877742, -0.53110755, -0.90279818, -0.15714875, -0.18551235, -1.51965987, 1.31035525, 1.32500020, 0.76743724, -1.01462665
sf_polygon( obj = df, x = "x", y = "y", z = "z", m = "m")
#>   id
#> 1  1
#>                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                       geometry
#> 1 -0.155360343, 0.537277880, -0.671733245, -0.983006815, 0.186525564, -1.269652403, 0.320545190, 0.222504998, -0.274213514, -0.129691058, 0.381336308, -1.133033274, 0.858229723, -0.613131698, 0.034011418, -0.155360343, -0.322356812, -0.221175899, 0.412894788, 0.139438546, -0.719014106, 0.809172759, -1.207102164, 0.937212770, 0.357842355, 1.912662198, -0.094741541, 0.865172646, 1.379922574, 1.832902643, 0.648319661, -0.322356812, -1.014626648, -1.731606696, 1.513973146, 0.063101701, 1.304543995, 0.640398562, 1.488777415, -0.531107546, -0.902798177, -0.157148754, -0.185512350, -1.519659867, 1.310355251, 1.325000197, 0.767437243, -1.014626648, 0.005731243, 1.118113938, 0.820615518, 1.470039551, -0.076570492, -0.761763534, -0.572477122, -0.499814086, 0.333603337, 0.918790175, -0.249856782, -0.888750244, 0.550877390, -0.003685816, 0.007598280, 0.005731243

sf_polygon( obj = df, x = 2, y = 3)
#>   id
#> 1  1
#>                                                                                                                                                                                                                                                                                                                                                                                                  geometry
#> 1 1.00000000, 1.00000000, 1.00000000, 2.00000000, 2.00000000, 2.00000000, 3.00000000, 3.00000000, 3.00000000, 1.00000000, 1.00000000, 1.00000000, 2.00000000, 2.00000000, 2.00000000, 1.00000000, -0.15536034, 0.53727788, -0.67173324, -0.98300682, 0.18652556, -1.26965240, 0.32054519, 0.22250500, -0.27421351, -0.12969106, 0.38133631, -1.13303327, 0.85822972, -0.61313170, 0.03401142, -0.15536034
sf_polygon( obj = df, x = 2, y = 3, z = 4)
#>   id
#> 1  1
#>                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                        geometry
#> 1 1.00000000, 1.00000000, 1.00000000, 2.00000000, 2.00000000, 2.00000000, 3.00000000, 3.00000000, 3.00000000, 1.00000000, 1.00000000, 1.00000000, 2.00000000, 2.00000000, 2.00000000, 1.00000000, -0.15536034, 0.53727788, -0.67173324, -0.98300682, 0.18652556, -1.26965240, 0.32054519, 0.22250500, -0.27421351, -0.12969106, 0.38133631, -1.13303327, 0.85822972, -0.61313170, 0.03401142, -0.15536034, -0.32235681, -0.22117590, 0.41289479, 0.13943855, -0.71901411, 0.80917276, -1.20710216, 0.93721277, 0.35784235, 1.91266220, -0.09474154, 0.86517265, 1.37992257, 1.83290264, 0.64831966, -0.32235681
sf_polygon( obj = df, x = 2, y = 3, z = 4, m = 5)
#>   id
#> 1  1
#>                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                geometry
#> 1 1.00000000, 1.00000000, 1.00000000, 2.00000000, 2.00000000, 2.00000000, 3.00000000, 3.00000000, 3.00000000, 1.00000000, 1.00000000, 1.00000000, 2.00000000, 2.00000000, 2.00000000, 1.00000000, -0.15536034, 0.53727788, -0.67173324, -0.98300682, 0.18652556, -1.26965240, 0.32054519, 0.22250500, -0.27421351, -0.12969106, 0.38133631, -1.13303327, 0.85822972, -0.61313170, 0.03401142, -0.15536034, -0.32235681, -0.22117590, 0.41289479, 0.13943855, -0.71901411, 0.80917276, -1.20710216, 0.93721277, 0.35784235, 1.91266220, -0.09474154, 0.86517265, 1.37992257, 1.83290264, 0.64831966, -0.32235681, -1.01462665, -1.73160670, 1.51397315, 0.06310170, 1.30454400, 0.64039856, 1.48877742, -0.53110755, -0.90279818, -0.15714875, -0.18551235, -1.51965987, 1.31035525, 1.32500020, 0.76743724, -1.01462665

sf_polygon( obj = df, polygon_id = "ml_id", linestring_id = "l_id" )
#>   ml_id
#> 1     1
#> 2     2
#>                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                geometry
#> 1 -0.155360343, 0.537277880, -0.671733245, -0.155360343, -0.322356812, -0.221175899, 0.412894788, -0.322356812, -1.014626648, -1.731606696, 1.513973146, -1.014626648, 0.005731243, 1.118113938, 0.820615518, 0.005731243, -0.983006815, 0.186525564, -1.269652403, -0.983006815, 0.139438546, -0.719014106, 0.809172759, 0.139438546, 0.063101701, 1.304543995, 0.640398562, 0.063101701, 1.470039551, -0.076570492, -0.761763534, 1.470039551, 0.320545190, 0.222504998, -0.274213514, 0.320545190, -1.207102164, 0.937212770, 0.357842355, -1.207102164, 1.488777415, -0.531107546, -0.902798177, 1.488777415, -0.572477122, -0.499814086, 0.333603337, -0.572477122
#> 2                                                                                                                                                                                                                            -0.129691058, 0.381336308, -1.133033274, -0.129691058, 1.912662198, -0.094741541, 0.865172646, 1.912662198, -0.157148754, -0.185512350, -1.519659867, -0.157148754, 0.918790175, -0.249856782, -0.888750244, 0.918790175, 0.858229723, -0.613131698, 0.034011418, 0.858229723, 1.379922574, 1.832902643, 0.648319661, 1.379922574, 1.310355251, 1.325000197, 0.767437243, 1.310355251, 0.550877390, -0.003685816, 0.007598280, 0.550877390
sf_polygon( obj = df, polygon_id = 1, linestring_id = 2 )
#>   id
#> 1  1
#> 2  2
#>                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                geometry
#> 1 -0.155360343, 0.537277880, -0.671733245, -0.155360343, -0.322356812, -0.221175899, 0.412894788, -0.322356812, -1.014626648, -1.731606696, 1.513973146, -1.014626648, 0.005731243, 1.118113938, 0.820615518, 0.005731243, -0.983006815, 0.186525564, -1.269652403, -0.983006815, 0.139438546, -0.719014106, 0.809172759, 0.139438546, 0.063101701, 1.304543995, 0.640398562, 0.063101701, 1.470039551, -0.076570492, -0.761763534, 1.470039551, 0.320545190, 0.222504998, -0.274213514, 0.320545190, -1.207102164, 0.937212770, 0.357842355, -1.207102164, 1.488777415, -0.531107546, -0.902798177, 1.488777415, -0.572477122, -0.499814086, 0.333603337, -0.572477122
#> 2                                                                                                                                                                                                                            -0.129691058, 0.381336308, -1.133033274, -0.129691058, 1.912662198, -0.094741541, 0.865172646, 1.912662198, -0.157148754, -0.185512350, -1.519659867, -0.157148754, 0.918790175, -0.249856782, -0.888750244, 0.918790175, 0.858229723, -0.613131698, 0.034011418, 0.858229723, 1.379922574, 1.832902643, 0.648319661, 1.379922574, 1.310355251, 1.325000197, 0.767437243, 1.310355251, 0.550877390, -0.003685816, 0.007598280, 0.550877390

## keeping properties
df <- data.frame(
  ml_id = c(1,1,1,1,1,1,1,1,1,2,2,2,2,2,2)
  , l_id = c(1,1,1,2,2,2,3,3,3,1,1,1,2,2,2)
  , x = rnorm(15)
  , y = rnorm(15)
  , z = rnorm(15)
  , m = rnorm(15)
  , val = letters[1:15]
  , stringsAsFactors = FALSE
)

## using keep = TRUE means the first row of all non-geometries are kept
sf_polygon(
  obj = df
  , polygon_id = "ml_id"
  , linestring_id = "l_id"
  , x = "x"
  , y = "y"
  , keep = TRUE
)
#>   ml_id         z          m val
#> 1     1 3.0454769 -0.4839348   a
#> 2     2 0.4580781  0.1134499   j
#>                                                                                                                                                                                                                                                                                                       geometry
#> 1 -0.63305754, 1.67408860, 1.15885389, -0.63305754, -0.19912919, 0.26741684, 0.47112234, -0.19912919, -1.03313933, 0.27950199, -1.56678680, -1.03313933, 0.06819896, 0.60414077, -0.05196202, 0.06819896, -0.93960589, -0.58635515, 0.16858590, -0.93960589, -0.46115719, 0.46036090, -0.45411053, -0.46115719
#> 2                                                                                                                       -0.2922863, -0.3018597, -1.1991880, -0.2922863, 1.1575050, -1.0577738, 1.5382489, 1.1575050, -0.4160917, -0.7301155, 1.1047246, -0.4160917, 0.1776384, 0.1697684, 1.0378085, 0.1776384

## use 'list_column' to specify columns where you want to keep all the values
sf_polygon(
  obj = df
  , polygon_id = "ml_id"
  , linestring_id = "l_id"
  , x = "x"
  , y = "y"
  , keep = TRUE
  , list_columns = "val"
)
#>   ml_id         z          m                                val
#> 1     1 3.0454769 -0.4839348 a, b, c, a, d, e, f, d, g, h, i, g
#> 2     2 0.4580781  0.1134499             j, k, l, j, m, n, o, m
#>                                                                                                                                                                                                                                                                                                       geometry
#> 1 -0.63305754, 1.67408860, 1.15885389, -0.63305754, -0.19912919, 0.26741684, 0.47112234, -0.19912919, -1.03313933, 0.27950199, -1.56678680, -1.03313933, 0.06819896, 0.60414077, -0.05196202, 0.06819896, -0.93960589, -0.58635515, 0.16858590, -0.93960589, -0.46115719, 0.46036090, -0.45411053, -0.46115719
#> 2                                                                                                                       -0.2922863, -0.3018597, -1.1991880, -0.2922863, 1.1575050, -1.0577738, 1.5382489, 1.1575050, -0.4160917, -0.7301155, 1.1047246, -0.4160917, 0.1776384, 0.1697684, 1.0378085, 0.1776384