sf POLYGON

constructs an sf of POLYGON objects

sf_polygon(
  obj = NULL,
  x = NULL,
  y = NULL,
  z = NULL,
  m = NULL,
  polygon_id = NULL,
  linestring_id = NULL,
  close = TRUE,
  keep = FALSE,
  list_columns = NULL
)

Arguments

obj

sorted matrix or data.frame

x

x geometry column

y

y geometry column

z

z geometry column

m

m geometry column

polygon_id

column of ids for polygons

linestring_id

column of ids for lines (within polygons)

close

logical indicating whether polygons should be closed. If TRUE, all polygons will be checked and force closed if possible

keep

logical indicating if the non-geometry and non-id columns should be kept. if TRUE you must supply the geometry and id columns, and only the first row of each geometry is kept. See Keeping Properties.

list_columns

vector of column names to turn into a list.

Value

sf object of POLYGON geometries

notes

sfheaders functions do not perform any validity checks on the geometries. Nor do they set Coordinate Reference Systems, EPSG, PROJ4 or precision attributes.

The data.frame and matrices you send into the sfheader functions must be ordered.

Keeping Properties

Setting keep = TRUE will retain any columns not specified as a coordinate (x, y, z, m) or an id (e.g., linestring_id, polygon_id) of the input obj.

You can use list_columns to specify which of the properties will be turned into a list, thus keeping all the values in the column. For columns not specified in list_columns, only the first row of the column is kept

The sf_* functions assume the input obj is a long data.frame / matrix, where any properties are repeated down the table for the same geometry.

Examples

m <- matrix(c(0,0,0,0,1,1), ncol = 2 )
sf_polygon( m )
#>   id               geometry
#> 1  1 0, 0, 0, 0, 0, 1, 1, 0

m <- matrix(c(0,0,0,0,0,1,0,1,1,1,2,2,1,2,3,1,3,4), ncol = 3, byrow = TRUE)
sf_polygon( obj = m )
#>   id                                                      geometry
#> 1  1 0, 0, 0, 1, 1, 1, 0, 0, 0, 1, 2, 2, 3, 0, 0, 1, 1, 2, 3, 4, 0
sf_polygon( obj = m, polygon_id = 1 )
#>   id               geometry
#> 1  0 0, 0, 1, 0, 0, 1, 1, 0
#> 2  1 2, 2, 3, 2, 2, 3, 4, 2
sf_polygon( obj = m, linestring_id = 1 )
#>   id                                       geometry
#> 1  1 0, 0, 1, 0, 0, 1, 1, 0, 2, 2, 3, 2, 2, 3, 4, 2

sf_polygon( obj = m, linestring_id = 1, polygon_id = 1 )
#>   id               geometry
#> 1  0 0, 0, 1, 0, 0, 1, 1, 0
#> 2  1 2, 2, 3, 2, 2, 3, 4, 2

sf_polygon( obj = m, x = 2, y = 3 )
#>   id                                 geometry
#> 1  1 0, 0, 1, 2, 2, 3, 0, 0, 1, 1, 2, 3, 4, 0
sf_polygon( obj = m, x = 1, y = 2, z = 3 )
#>   id                                                      geometry
#> 1  1 0, 0, 0, 1, 1, 1, 0, 0, 0, 1, 2, 2, 3, 0, 0, 1, 1, 2, 3, 4, 0
sf_polygon( obj = m, x = 2, y = 3, linestring_id = 1, polygon_id = 1 )
#>   id               geometry
#> 1  0 0, 0, 1, 0, 0, 1, 1, 0
#> 2  1 2, 2, 3, 2, 2, 3, 4, 2

df <- data.frame(
  ml_id = c(1,1,1,1,1,1,1,1,1,2,2,2,2,2,2)
  , l_id = c(1,1,1,2,2,2,3,3,3,1,1,1,2,2,2)
  , x = rnorm(15)
  , y = rnorm(15)
  , z = rnorm(15)
  , m = rnorm(15)
)

sf_polygon( obj = df, x = "x", y = "y")
#>   id
#> 1  1
#>                                                                                                                                                                                                                                                                                                                                                                                                                                     geometry
#> 1 1.512738117, 0.500661016, 0.578869043, 1.288935661, -3.189859822, 0.060469912, 1.293751674, -1.174243647, -0.835355867, 0.610871626, 0.331662500, 1.143064057, -0.632014447, -0.498994682, -1.050680851, 1.512738117, -0.239127227, -1.358785338, 1.614207113, 2.377089701, 1.793844169, 0.990258217, 0.948142447, -0.338381392, -0.412977457, 2.374744848, 0.295458792, 0.506110127, 0.517428935, 0.009555378, -0.272603463, -0.239127227
sf_polygon( obj = df, x = "x", y = "y", z = "z")
#>   id
#> 1  1
#>                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                          geometry
#> 1 1.512738117, 0.500661016, 0.578869043, 1.288935661, -3.189859822, 0.060469912, 1.293751674, -1.174243647, -0.835355867, 0.610871626, 0.331662500, 1.143064057, -0.632014447, -0.498994682, -1.050680851, 1.512738117, -0.239127227, -1.358785338, 1.614207113, 2.377089701, 1.793844169, 0.990258217, 0.948142447, -0.338381392, -0.412977457, 2.374744848, 0.295458792, 0.506110127, 0.517428935, 0.009555378, -0.272603463, -0.239127227, 0.183501119, -0.319188796, 0.882807721, 0.327286373, 1.676438369, 0.125332866, -0.620141083, 0.016487582, -0.309906639, 0.715295238, -0.270926677, 0.584101678, -1.103228775, 0.490744178, 0.270167344, 0.183501119
sf_polygon( obj = df, x = "x", y = "y", z = "z", m = "m")
#>   id
#> 1  1
#>                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                  geometry
#> 1 1.512738117, 0.500661016, 0.578869043, 1.288935661, -3.189859822, 0.060469912, 1.293751674, -1.174243647, -0.835355867, 0.610871626, 0.331662500, 1.143064057, -0.632014447, -0.498994682, -1.050680851, 1.512738117, -0.239127227, -1.358785338, 1.614207113, 2.377089701, 1.793844169, 0.990258217, 0.948142447, -0.338381392, -0.412977457, 2.374744848, 0.295458792, 0.506110127, 0.517428935, 0.009555378, -0.272603463, -0.239127227, 0.183501119, -0.319188796, 0.882807721, 0.327286373, 1.676438369, 0.125332866, -0.620141083, 0.016487582, -0.309906639, 0.715295238, -0.270926677, 0.584101678, -1.103228775, 0.490744178, 0.270167344, 0.183501119, -1.373456356, -1.024999010, -1.045272216, -0.431857112, -1.841159765, 2.169452827, 0.511941887, 0.841224507, -1.155744923, 0.170374774, 0.069919138, 0.518921118, 0.727841394, -1.739704292, 1.016873329, -1.373456356

sf_polygon( obj = df, x = 2, y = 3)
#>   id
#> 1  1
#>                                                                                                                                                                                                                                                                                                                                                                                               geometry
#> 1 1.00000000, 1.00000000, 1.00000000, 2.00000000, 2.00000000, 2.00000000, 3.00000000, 3.00000000, 3.00000000, 1.00000000, 1.00000000, 1.00000000, 2.00000000, 2.00000000, 2.00000000, 1.00000000, 1.51273812, 0.50066102, 0.57886904, 1.28893566, -3.18985982, 0.06046991, 1.29375167, -1.17424365, -0.83535587, 0.61087163, 0.33166250, 1.14306406, -0.63201445, -0.49899468, -1.05068085, 1.51273812
sf_polygon( obj = df, x = 2, y = 3, z = 4)
#>   id
#> 1  1
#>                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                     geometry
#> 1 1.000000000, 1.000000000, 1.000000000, 2.000000000, 2.000000000, 2.000000000, 3.000000000, 3.000000000, 3.000000000, 1.000000000, 1.000000000, 1.000000000, 2.000000000, 2.000000000, 2.000000000, 1.000000000, 1.512738117, 0.500661016, 0.578869043, 1.288935661, -3.189859822, 0.060469912, 1.293751674, -1.174243647, -0.835355867, 0.610871626, 0.331662500, 1.143064057, -0.632014447, -0.498994682, -1.050680851, 1.512738117, -0.239127227, -1.358785338, 1.614207113, 2.377089701, 1.793844169, 0.990258217, 0.948142447, -0.338381392, -0.412977457, 2.374744848, 0.295458792, 0.506110127, 0.517428935, 0.009555378, -0.272603463, -0.239127227
sf_polygon( obj = df, x = 2, y = 3, z = 4, m = 5)
#>   id
#> 1  1
#>                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                          geometry
#> 1 1.000000000, 1.000000000, 1.000000000, 2.000000000, 2.000000000, 2.000000000, 3.000000000, 3.000000000, 3.000000000, 1.000000000, 1.000000000, 1.000000000, 2.000000000, 2.000000000, 2.000000000, 1.000000000, 1.512738117, 0.500661016, 0.578869043, 1.288935661, -3.189859822, 0.060469912, 1.293751674, -1.174243647, -0.835355867, 0.610871626, 0.331662500, 1.143064057, -0.632014447, -0.498994682, -1.050680851, 1.512738117, -0.239127227, -1.358785338, 1.614207113, 2.377089701, 1.793844169, 0.990258217, 0.948142447, -0.338381392, -0.412977457, 2.374744848, 0.295458792, 0.506110127, 0.517428935, 0.009555378, -0.272603463, -0.239127227, 0.183501119, -0.319188796, 0.882807721, 0.327286373, 1.676438369, 0.125332866, -0.620141083, 0.016487582, -0.309906639, 0.715295238, -0.270926677, 0.584101678, -1.103228775, 0.490744178, 0.270167344, 0.183501119

sf_polygon( obj = df, polygon_id = "ml_id", linestring_id = "l_id" )
#>   ml_id
#> 1     1
#> 2     2
#>                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                             geometry
#> 1 1.51273812, 0.50066102, 0.57886904, 1.51273812, -0.23912723, -1.35878534, 1.61420711, -0.23912723, 0.18350112, -0.31918880, 0.88280772, 0.18350112, -1.37345636, -1.02499901, -1.04527222, -1.37345636, 1.28893566, -3.18985982, 0.06046991, 1.28893566, 2.37708970, 1.79384417, 0.99025822, 2.37708970, 0.32728637, 1.67643837, 0.12533287, 0.32728637, -0.43185711, -1.84115976, 2.16945283, -0.43185711, 1.29375167, -1.17424365, -0.83535587, 1.29375167, 0.94814245, -0.33838139, -0.41297746, 0.94814245, -0.62014108, 0.01648758, -0.30990664, -0.62014108, 0.51194189, 0.84122451, -1.15574492, 0.51194189
#> 2                                                                                                                                                                            0.610871626, 0.331662500, 1.143064057, 0.610871626, 2.374744848, 0.295458792, 0.506110127, 2.374744848, 0.715295238, -0.270926677, 0.584101678, 0.715295238, 0.170374774, 0.069919138, 0.518921118, 0.170374774, -0.632014447, -0.498994682, -1.050680851, -0.632014447, 0.517428935, 0.009555378, -0.272603463, 0.517428935, -1.103228775, 0.490744178, 0.270167344, -1.103228775, 0.727841394, -1.739704292, 1.016873329, 0.727841394
sf_polygon( obj = df, polygon_id = 1, linestring_id = 2 )
#>   id
#> 1  1
#> 2  2
#>                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                             geometry
#> 1 1.51273812, 0.50066102, 0.57886904, 1.51273812, -0.23912723, -1.35878534, 1.61420711, -0.23912723, 0.18350112, -0.31918880, 0.88280772, 0.18350112, -1.37345636, -1.02499901, -1.04527222, -1.37345636, 1.28893566, -3.18985982, 0.06046991, 1.28893566, 2.37708970, 1.79384417, 0.99025822, 2.37708970, 0.32728637, 1.67643837, 0.12533287, 0.32728637, -0.43185711, -1.84115976, 2.16945283, -0.43185711, 1.29375167, -1.17424365, -0.83535587, 1.29375167, 0.94814245, -0.33838139, -0.41297746, 0.94814245, -0.62014108, 0.01648758, -0.30990664, -0.62014108, 0.51194189, 0.84122451, -1.15574492, 0.51194189
#> 2                                                                                                                                                                            0.610871626, 0.331662500, 1.143064057, 0.610871626, 2.374744848, 0.295458792, 0.506110127, 2.374744848, 0.715295238, -0.270926677, 0.584101678, 0.715295238, 0.170374774, 0.069919138, 0.518921118, 0.170374774, -0.632014447, -0.498994682, -1.050680851, -0.632014447, 0.517428935, 0.009555378, -0.272603463, 0.517428935, -1.103228775, 0.490744178, 0.270167344, -1.103228775, 0.727841394, -1.739704292, 1.016873329, 0.727841394

## keeping properties
df <- data.frame(
  ml_id = c(1,1,1,1,1,1,1,1,1,2,2,2,2,2,2)
  , l_id = c(1,1,1,2,2,2,3,3,3,1,1,1,2,2,2)
  , x = rnorm(15)
  , y = rnorm(15)
  , z = rnorm(15)
  , m = rnorm(15)
  , val = letters[1:15]
  , stringsAsFactors = FALSE
)

## using keep = TRUE means the first row of all non-geometries are kept
sf_polygon(
  obj = df
  , polygon_id = "ml_id"
  , linestring_id = "l_id"
  , x = "x"
  , y = "y"
  , keep = TRUE
)
#>   ml_id          z          m val
#> 1     1 -0.3573661 -0.1276583   a
#> 2     2 -0.8725965  0.2017193   j
#>                                                                                                                                                                                                                                                                               geometry
#> 1 -0.2938867, -0.4894186, 1.4720363, -0.2938867, -1.1728285, -0.7180245, -0.5639766, -1.1728285, 0.1733687, 0.7718210, -1.9586201, 0.1733687, 0.7380382, -1.2036981, 1.5686419, 0.7380382, -0.4793672, 0.6260620, 0.1790756, -0.4793672, -0.4649657, -0.9184742, 0.8585787, -0.4649657
#> 2                                                                              -0.57485386, 0.40660767, 0.41193014, -0.57485386, -0.41724868, 0.01145704, 0.89514049, -0.41724868, -1.38159219, -0.93981993, -0.19988495, -1.38159219, 0.46752144, -0.66604128, 0.02753222, 0.46752144

## use 'list_column' to specify columns where you want to keep all the values
sf_polygon(
  obj = df
  , polygon_id = "ml_id"
  , linestring_id = "l_id"
  , x = "x"
  , y = "y"
  , keep = TRUE
  , list_columns = "val"
)
#>   ml_id          z          m                                val
#> 1     1 -0.3573661 -0.1276583 a, b, c, a, d, e, f, d, g, h, i, g
#> 2     2 -0.8725965  0.2017193             j, k, l, j, m, n, o, m
#>                                                                                                                                                                                                                                                                               geometry
#> 1 -0.2938867, -0.4894186, 1.4720363, -0.2938867, -1.1728285, -0.7180245, -0.5639766, -1.1728285, 0.1733687, 0.7718210, -1.9586201, 0.1733687, 0.7380382, -1.2036981, 1.5686419, 0.7380382, -0.4793672, 0.6260620, 0.1790756, -0.4793672, -0.4649657, -0.9184742, 0.8585787, -0.4649657
#> 2                                                                              -0.57485386, 0.40660767, 0.41193014, -0.57485386, -0.41724868, 0.01145704, 0.89514049, -0.41724868, -1.38159219, -0.93981993, -0.19988495, -1.38159219, 0.46752144, -0.66604128, 0.02753222, 0.46752144