sf POLYGON

constructs an sf of POLYGON objects

sf_polygon(obj = NULL, x = NULL, y = NULL, z = NULL, m = NULL,
  polygon_id = NULL, linestring_id = NULL)

Arguments

obj
sorted matrix or data.frame

x
x geometry column

y
y geometry column

z
z geometry column

m
m geometry column

polygon_id
column of ids for polygons

linestring_id
column of ids for lines (within polygons)

notes

sfheaders functions do not perform any validity checks on the geometries. Nor do they set Coordinate Reference Systems, EPSG, PROJ4 or precision attributes.

The data.frame and matrices you send into the sfheader functions must be ordered.

Examples

m <- matrix(c(0,0,0,0,1,1), ncol = 3 )
sf_polygon( m )

#>   id         geometry
#> 1  1 0, 0, 0, 0, 1, 1

m <- matrix(c(0,0,0,0,0,1,0,1,1,1,2,2,1,2,3), ncol = 3, byrow = TRUE)
sf_polygon( obj = m )

#>   id                                    geometry
#> 1  1 0, 0, 0, 1, 1, 0, 0, 1, 2, 2, 0, 1, 1, 2, 3

sf_polygon( obj = m, polygon_id = 1 )

#>   id         geometry
#> 1  0 0, 0, 1, 0, 1, 1
#> 2  1       2, 2, 2, 3

sf_polygon( obj = m, linestring_id = 1 )

#>   id                     geometry
#> 1  1 0, 0, 1, 0, 1, 1, 2, 2, 2, 3

sf_polygon( obj = m, linestring_id = 1, polygon_id = 1 )

#>   id         geometry
#> 1  0 0, 0, 1, 0, 1, 1
#> 2  1       2, 2, 2, 3

sf_polygon( obj = m, x = 2, y = 3 )

#>   id                     geometry
#> 1  1 0, 0, 1, 2, 2, 0, 1, 1, 2, 3

sf_polygon( obj = m, x = 1, y = 2, z = 3 )

#>   id                                    geometry
#> 1  1 0, 0, 0, 1, 1, 0, 0, 1, 2, 2, 0, 1, 1, 2, 3

sf_polygon( obj = m, x = 2, y = 3, linestring_id = 1, polygon_id = 1 )

#>   id         geometry
#> 1  0 0, 0, 1, 0, 1, 1
#> 2  1       2, 2, 2, 3

df <- data.frame(
  ml_id = c(1,1,1,1,1,1,1,1,2,2,2,2,2)
  , l_id = c(1,1,1,2,2,3,3,3,1,1,1,2,2)
  , x = rnorm(13)
  , y = rnorm(13)
  , z = rnorm(13)
  , m = rnorm(13)
)

sf_polygon( obj = df, x = "x", y = "y")

#>   id
#> 1  1
#>                                                                                                                                                                                                                                                                                                                           geometry
#> 1 -0.20608719, 0.01917759, 0.02956075, 0.54982754, -2.27411486, 2.68255718, -0.36122126, 0.21335575, 1.07434588, -0.66508825, 1.11395242, -0.24589641, -1.17756331, -0.97585062, 1.06505732, 0.13167063, 0.48862881, -1.69945057, -1.47073631, 0.28415034, 1.33732041, 0.23669628, 1.31829338, 0.52390979, 0.60674805, -0.10993567

sf_polygon( obj = df, x = "x", y = "y", z = "z")

#>   id
#> 1  1
#>                                                                                                                                                                                                                                                                                                                                                                                                                                                                                            geometry
#> 1 -0.20608719, 0.01917759, 0.02956075, 0.54982754, -2.27411486, 2.68255718, -0.36122126, 0.21335575, 1.07434588, -0.66508825, 1.11395242, -0.24589641, -1.17756331, -0.97585062, 1.06505732, 0.13167063, 0.48862881, -1.69945057, -1.47073631, 0.28415034, 1.33732041, 0.23669628, 1.31829338, 0.52390979, 0.60674805, -0.10993567, 0.17218172, -0.09032729, 1.92434334, 1.29839276, 0.74879127, 0.55622433, -0.54825726, 1.11053489, -2.61233433, -0.15569378, 0.43388979, -0.38195111, 0.42418757

sf_polygon( obj = df, x = "x", y = "y", z = "z", m = "m")

#>   id
#> 1  1
#>                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                               geometry
#> 1 -0.20608719, 0.01917759, 0.02956075, 0.54982754, -2.27411486, 2.68255718, -0.36122126, 0.21335575, 1.07434588, -0.66508825, 1.11395242, -0.24589641, -1.17756331, -0.97585062, 1.06505732, 0.13167063, 0.48862881, -1.69945057, -1.47073631, 0.28415034, 1.33732041, 0.23669628, 1.31829338, 0.52390979, 0.60674805, -0.10993567, 0.17218172, -0.09032729, 1.92434334, 1.29839276, 0.74879127, 0.55622433, -0.54825726, 1.11053489, -2.61233433, -0.15569378, 0.43388979, -0.38195111, 0.42418757, 1.06310200, 1.04871262, -0.03810289, 0.48614892, 1.67288261, -0.35436116, 0.94634789, 1.31682636, -0.29664002, -0.38721358, -0.78543266, -1.05673687, -0.79554143

sf_polygon( obj = df, x = 2, y = 3)

#>   id
#> 1  1
#>                                                                                                                                                                                                                                                                                                                       geometry
#> 1 1.00000000, 1.00000000, 1.00000000, 2.00000000, 2.00000000, 3.00000000, 3.00000000, 3.00000000, 1.00000000, 1.00000000, 1.00000000, 2.00000000, 2.00000000, -0.20608719, 0.01917759, 0.02956075, 0.54982754, -2.27411486, 2.68255718, -0.36122126, 0.21335575, 1.07434588, -0.66508825, 1.11395242, -0.24589641, -1.17756331

sf_polygon( obj = df, x = 2, y = 3, z = 4)

#>   id
#> 1  1
#>                                                                                                                                                                                                                                                                                                                                                                                                                                                                                       geometry
#> 1 1.00000000, 1.00000000, 1.00000000, 2.00000000, 2.00000000, 3.00000000, 3.00000000, 3.00000000, 1.00000000, 1.00000000, 1.00000000, 2.00000000, 2.00000000, -0.20608719, 0.01917759, 0.02956075, 0.54982754, -2.27411486, 2.68255718, -0.36122126, 0.21335575, 1.07434588, -0.66508825, 1.11395242, -0.24589641, -1.17756331, -0.97585062, 1.06505732, 0.13167063, 0.48862881, -1.69945057, -1.47073631, 0.28415034, 1.33732041, 0.23669628, 1.31829338, 0.52390979, 0.60674805, -0.10993567

sf_polygon( obj = df, x = 2, y = 3, z = 4, m = 5)

#>   id
#> 1  1
#>                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                        geometry
#> 1 1.00000000, 1.00000000, 1.00000000, 2.00000000, 2.00000000, 3.00000000, 3.00000000, 3.00000000, 1.00000000, 1.00000000, 1.00000000, 2.00000000, 2.00000000, -0.20608719, 0.01917759, 0.02956075, 0.54982754, -2.27411486, 2.68255718, -0.36122126, 0.21335575, 1.07434588, -0.66508825, 1.11395242, -0.24589641, -1.17756331, -0.97585062, 1.06505732, 0.13167063, 0.48862881, -1.69945057, -1.47073631, 0.28415034, 1.33732041, 0.23669628, 1.31829338, 0.52390979, 0.60674805, -0.10993567, 0.17218172, -0.09032729, 1.92434334, 1.29839276, 0.74879127, 0.55622433, -0.54825726, 1.11053489, -2.61233433, -0.15569378, 0.43388979, -0.38195111, 0.42418757

sf_polygon( obj = df, polygon_id = "ml_id", linestring_id = "l_id" )

#>   id
#> 1  1
#> 2  2
#>                                                                                                                                                                                                                                                                                                                                                                                                   geometry
#> 1 -0.20608719, 0.01917759, 0.02956075, -0.97585062, 1.06505732, 0.13167063, 0.17218172, -0.09032729, 1.92434334, 1.06310200, 1.04871262, -0.03810289, 0.54982754, -2.27411486, 0.48862881, -1.69945057, 1.29839276, 0.74879127, 0.48614892, 1.67288261, 2.68255718, -0.36122126, 0.21335575, -1.47073631, 0.28415034, 1.33732041, 0.55622433, -0.54825726, 1.11053489, -0.35436116, 0.94634789, 1.31682636
#> 2                                                                                                                                                                   1.0743459, -0.6650882, 1.1139524, 0.2366963, 1.3182934, 0.5239098, -2.6123343, -0.1556938, 0.4338898, -0.2966400, -0.3872136, -0.7854327, -0.2458964, -1.1775633, 0.6067480, -0.1099357, -0.3819511, 0.4241876, -1.0567369, -0.7955414

sf_polygon( obj = df, polygon_id = 1, linestring_id = 2 )

#>   id
#> 1  1
#> 2  2
#>                                                                                                                                                                                                                                                                                                                                                                                                   geometry
#> 1 -0.20608719, 0.01917759, 0.02956075, -0.97585062, 1.06505732, 0.13167063, 0.17218172, -0.09032729, 1.92434334, 1.06310200, 1.04871262, -0.03810289, 0.54982754, -2.27411486, 0.48862881, -1.69945057, 1.29839276, 0.74879127, 0.48614892, 1.67288261, 2.68255718, -0.36122126, 0.21335575, -1.47073631, 0.28415034, 1.33732041, 0.55622433, -0.54825726, 1.11053489, -0.35436116, 0.94634789, 1.31682636
#> 2                                                                                                                                                                   1.0743459, -0.6650882, 1.1139524, 0.2366963, 1.3182934, 0.5239098, -2.6123343, -0.1556938, 0.4338898, -0.2966400, -0.3872136, -0.7854327, -0.2458964, -1.1775633, 0.6067480, -0.1099357, -0.3819511, 0.4241876, -1.0567369, -0.7955414