sfc MULTILINESTRING

constructs an sfc of MULTILINESTRING objects

sfc_multilinestring(obj = NULL, x = NULL, y = NULL, z = NULL,
  m = NULL, multilinestring_id = NULL, linestring_id = NULL)

Arguments

obj
sorted matrix or data.frame

x
x geometry column

y
y geometry column

z
z geometry column

m
m geometry column

multilinestring_id
column of ids for multilinestrings

linestring_id
column of ids for linestrings (within multilinestrings)

notes

sfheaders functions do not perform any validity checks on the geometries. Nor do they set Coordinate Reference Systems, EPSG, PROJ4 or precision attributes.

The data.frame and matrices you send into the sfheader functions must be ordered.

Examples

m <- matrix(c(0,0,0,0,1,1), ncol = 3 )
sfc_multilinestring( m )

#> [[1]]
#> [[1]]
#>      [,1] [,2] [,3]
#> [1,]    0    0    1
#> [2,]    0    0    1
#> 
#> attr(,"class")
#> [1] "XYZ"             "MULTILINESTRING" "sfg"            
#> 
#> attr(,"class")
#> [1] "sfc_MULTILINESTRING" "sfc"                
#> attr(,"crs")
#> $epsg
#> [1] NA
#> 
#> $proj4string
#> [1] NA
#> 
#> attr(,"class")
#> [1] "crs"
#> attr(,"precision")
#> [1] 0
#> attr(,"n_empty")
#> [1] 0
#> attr(,"bbox")
#> xmin ymin xmax ymax 
#>    0    0    0    0 
#> attr(,"class")
#> [1] "bbox"
#> attr(,"z_range")
#> zmin zmax 
#>    1    1 
#> attr(,"class")
#> [1] "z_range"
#> attr(,"m_range")
#> mmin mmax 
#>   NA   NA 
#> attr(,"class")
#> [1] "m_range"

m <- matrix(c(0,0,0,0,0,1,0,1,1,1,2,2,1,2,3), ncol = 3, byrow = TRUE)
sfc_multilinestring( obj = m )

#> [[1]]
#> [[1]]
#>      [,1] [,2] [,3]
#> [1,]    0    0    0
#> [2,]    0    0    1
#> [3,]    0    1    1
#> [4,]    1    2    2
#> [5,]    1    2    3
#> 
#> attr(,"class")
#> [1] "XYZ"             "MULTILINESTRING" "sfg"            
#> 
#> attr(,"class")
#> [1] "sfc_MULTILINESTRING" "sfc"                
#> attr(,"crs")
#> $epsg
#> [1] NA
#> 
#> $proj4string
#> [1] NA
#> 
#> attr(,"class")
#> [1] "crs"
#> attr(,"precision")
#> [1] 0
#> attr(,"n_empty")
#> [1] 0
#> attr(,"bbox")
#> xmin ymin xmax ymax 
#>    0    0    1    2 
#> attr(,"class")
#> [1] "bbox"
#> attr(,"z_range")
#> zmin zmax 
#>    0    3 
#> attr(,"class")
#> [1] "z_range"
#> attr(,"m_range")
#> mmin mmax 
#>   NA   NA 
#> attr(,"class")
#> [1] "m_range"

sfc_multilinestring( obj = m, multilinestring_id = 1 )

#> [[1]]
#> [[1]]
#>      V2 V3
#> [1,]  0  0
#> [2,]  0  1
#> [3,]  1  1
#> 
#> attr(,"class")
#> [1] "XY"              "MULTILINESTRING" "sfg"            
#> 
#> [[2]]
#> [[1]]
#>      V2 V3
#> [1,]  2  2
#> [2,]  2  3
#> 
#> attr(,"class")
#> [1] "XY"              "MULTILINESTRING" "sfg"            
#> 
#> attr(,"class")
#> [1] "sfc_MULTILINESTRING" "sfc"                
#> attr(,"crs")
#> $epsg
#> [1] NA
#> 
#> $proj4string
#> [1] NA
#> 
#> attr(,"class")
#> [1] "crs"
#> attr(,"precision")
#> [1] 0
#> attr(,"n_empty")
#> [1] 0
#> attr(,"bbox")
#> xmin ymin xmax ymax 
#>    0    0    2    3 
#> attr(,"class")
#> [1] "bbox"
#> attr(,"z_range")
#> zmin zmax 
#>   NA   NA 
#> attr(,"class")
#> [1] "z_range"
#> attr(,"m_range")
#> mmin mmax 
#>   NA   NA 
#> attr(,"class")
#> [1] "m_range"

sfc_multilinestring( obj = m, linestring_id = 1 )

#> [[1]]
#> [[1]]
#>      [,1] [,2]
#> [1,]    0    0
#> [2,]    0    1
#> [3,]    1    1
#> 
#> [[2]]
#>      [,1] [,2]
#> [1,]    2    2
#> [2,]    2    3
#> 
#> attr(,"class")
#> [1] "XY"              "MULTILINESTRING" "sfg"            
#> 
#> attr(,"class")
#> [1] "sfc_MULTILINESTRING" "sfc"                
#> attr(,"crs")
#> $epsg
#> [1] NA
#> 
#> $proj4string
#> [1] NA
#> 
#> attr(,"class")
#> [1] "crs"
#> attr(,"precision")
#> [1] 0
#> attr(,"n_empty")
#> [1] 0
#> attr(,"bbox")
#> xmin ymin xmax ymax 
#>    0    0    2    3 
#> attr(,"class")
#> [1] "bbox"
#> attr(,"z_range")
#> zmin zmax 
#>   NA   NA 
#> attr(,"class")
#> [1] "z_range"
#> attr(,"m_range")
#> mmin mmax 
#>   NA   NA 
#> attr(,"class")
#> [1] "m_range"

sfc_multilinestring( obj = m, linestring_id = 1, multilinestring_id = 1 )

#> [[1]]
#> [[1]]
#>      V2 V3
#> [1,]  0  0
#> [2,]  0  1
#> [3,]  1  1
#> 
#> attr(,"class")
#> [1] "XY"              "MULTILINESTRING" "sfg"            
#> 
#> [[2]]
#> [[1]]
#>      V2 V3
#> [1,]  2  2
#> [2,]  2  3
#> 
#> attr(,"class")
#> [1] "XY"              "MULTILINESTRING" "sfg"            
#> 
#> attr(,"class")
#> [1] "sfc_MULTILINESTRING" "sfc"                
#> attr(,"crs")
#> $epsg
#> [1] NA
#> 
#> $proj4string
#> [1] NA
#> 
#> attr(,"class")
#> [1] "crs"
#> attr(,"precision")
#> [1] 0
#> attr(,"n_empty")
#> [1] 0
#> attr(,"bbox")
#> xmin ymin xmax ymax 
#>    0    0    2    3 
#> attr(,"class")
#> [1] "bbox"
#> attr(,"z_range")
#> zmin zmax 
#>   NA   NA 
#> attr(,"class")
#> [1] "z_range"
#> attr(,"m_range")
#> mmin mmax 
#>   NA   NA 
#> attr(,"class")
#> [1] "m_range"

sfc_multilinestring( obj = m, x = 2, y = 3 )

#> [[1]]
#> [[1]]
#>      [,1] [,2]
#> [1,]    0    0
#> [2,]    0    1
#> [3,]    1    1
#> [4,]    2    2
#> [5,]    2    3
#> 
#> attr(,"class")
#> [1] "XY"              "MULTILINESTRING" "sfg"            
#> 
#> attr(,"class")
#> [1] "sfc_MULTILINESTRING" "sfc"                
#> attr(,"crs")
#> $epsg
#> [1] NA
#> 
#> $proj4string
#> [1] NA
#> 
#> attr(,"class")
#> [1] "crs"
#> attr(,"precision")
#> [1] 0
#> attr(,"n_empty")
#> [1] 0
#> attr(,"bbox")
#> xmin ymin xmax ymax 
#>    0    0    2    3 
#> attr(,"class")
#> [1] "bbox"
#> attr(,"z_range")
#> zmin zmax 
#>   NA   NA 
#> attr(,"class")
#> [1] "z_range"
#> attr(,"m_range")
#> mmin mmax 
#>   NA   NA 
#> attr(,"class")
#> [1] "m_range"

sfc_multilinestring( obj = m, x = 1, y = 2, z = 3 )

#> [[1]]
#> [[1]]
#>      [,1] [,2] [,3]
#> [1,]    0    0    0
#> [2,]    0    0    1
#> [3,]    0    1    1
#> [4,]    1    2    2
#> [5,]    1    2    3
#> 
#> attr(,"class")
#> [1] "XYZ"             "MULTILINESTRING" "sfg"            
#> 
#> attr(,"class")
#> [1] "sfc_MULTILINESTRING" "sfc"                
#> attr(,"crs")
#> $epsg
#> [1] NA
#> 
#> $proj4string
#> [1] NA
#> 
#> attr(,"class")
#> [1] "crs"
#> attr(,"precision")
#> [1] 0
#> attr(,"n_empty")
#> [1] 0
#> attr(,"bbox")
#> xmin ymin xmax ymax 
#>    0    0    1    2 
#> attr(,"class")
#> [1] "bbox"
#> attr(,"z_range")
#> zmin zmax 
#>    0    3 
#> attr(,"class")
#> [1] "z_range"
#> attr(,"m_range")
#> mmin mmax 
#>   NA   NA 
#> attr(,"class")
#> [1] "m_range"

sfc_multilinestring( obj = m, x = 2, y = 3, linestring_id = 1, multilinestring_id = 1 )

#> [[1]]
#> [[1]]
#>      V2 V3
#> [1,]  0  0
#> [2,]  0  1
#> [3,]  1  1
#> 
#> attr(,"class")
#> [1] "XY"              "MULTILINESTRING" "sfg"            
#> 
#> [[2]]
#> [[1]]
#>      V2 V3
#> [1,]  2  2
#> [2,]  2  3
#> 
#> attr(,"class")
#> [1] "XY"              "MULTILINESTRING" "sfg"            
#> 
#> attr(,"class")
#> [1] "sfc_MULTILINESTRING" "sfc"                
#> attr(,"crs")
#> $epsg
#> [1] NA
#> 
#> $proj4string
#> [1] NA
#> 
#> attr(,"class")
#> [1] "crs"
#> attr(,"precision")
#> [1] 0
#> attr(,"n_empty")
#> [1] 0
#> attr(,"bbox")
#> xmin ymin xmax ymax 
#>    0    0    2    3 
#> attr(,"class")
#> [1] "bbox"
#> attr(,"z_range")
#> zmin zmax 
#>   NA   NA 
#> attr(,"class")
#> [1] "z_range"
#> attr(,"m_range")
#> mmin mmax 
#>   NA   NA 
#> attr(,"class")
#> [1] "m_range"

df <- data.frame(
  ml_id = c(1,1,1,1,1,1,1,1,2,2,2,2,2)
  , l_id = c(1,1,1,2,2,3,3,3,1,1,1,2,2)
  , x = rnorm(13)
  , y = rnorm(13)
  , z = rnorm(13)
  , m = rnorm(13)
)

sfc_multilinestring( obj = df, x = "x", y = "y")

#> [[1]]
#> [[1]]
#>             [,1]        [,2]
#>  [1,] -1.7562754 -1.68428153
#>  [2,] -0.6905379  0.24940178
#>  [3,] -0.5585420  1.07283825
#>  [4,] -0.5366633  2.03936926
#>  [5,]  0.2271271  0.44945378
#>  [6,]  0.9784549  1.39181405
#>  [7,] -0.2088827  0.42656655
#>  [8,] -1.3994105  0.10758399
#>  [9,]  0.2585373  0.02229473
#> [10,] -0.4417995  0.60361101
#> [11,]  0.5685999 -0.26265057
#> [12,]  2.1268505 -0.52826408
#> [13,]  0.4248584  0.19214942
#> 
#> attr(,"class")
#> [1] "XY"              "MULTILINESTRING" "sfg"            
#> 
#> attr(,"class")
#> [1] "sfc_MULTILINESTRING" "sfc"                
#> attr(,"crs")
#> $epsg
#> [1] NA
#> 
#> $proj4string
#> [1] NA
#> 
#> attr(,"class")
#> [1] "crs"
#> attr(,"precision")
#> [1] 0
#> attr(,"n_empty")
#> [1] 0
#> attr(,"bbox")
#>      xmin      ymin      xmax      ymax 
#> -1.756275 -1.684282  2.126850  2.039369 
#> attr(,"class")
#> [1] "bbox"
#> attr(,"z_range")
#> zmin zmax 
#>   NA   NA 
#> attr(,"class")
#> [1] "z_range"
#> attr(,"m_range")
#> mmin mmax 
#>   NA   NA 
#> attr(,"class")
#> [1] "m_range"

sfc_multilinestring( obj = df, x = "x", y = "y", z = "z")

#> [[1]]
#> [[1]]
#>             [,1]        [,2]        [,3]
#>  [1,] -1.7562754 -1.68428153 -1.14619967
#>  [2,] -0.6905379  0.24940178  0.84618466
#>  [3,] -0.5585420  1.07283825  0.08171963
#>  [4,] -0.5366633  2.03936926 -1.30511701
#>  [5,]  0.2271271  0.44945378 -0.94491206
#>  [6,]  0.9784549  1.39181405  0.45434159
#>  [7,] -0.2088827  0.42656655 -0.85520250
#>  [8,] -1.3994105  0.10758399 -0.28689522
#>  [9,]  0.2585373  0.02229473  0.89496163
#> [10,] -0.4417995  0.60361101  0.06730444
#> [11,]  0.5685999 -0.26265057 -0.16267634
#> [12,]  2.1268505 -0.52826408 -0.82731017
#> [13,]  0.4248584  0.19214942  1.87650562
#> 
#> attr(,"class")
#> [1] "XYZ"             "MULTILINESTRING" "sfg"            
#> 
#> attr(,"class")
#> [1] "sfc_MULTILINESTRING" "sfc"                
#> attr(,"crs")
#> $epsg
#> [1] NA
#> 
#> $proj4string
#> [1] NA
#> 
#> attr(,"class")
#> [1] "crs"
#> attr(,"precision")
#> [1] 0
#> attr(,"n_empty")
#> [1] 0
#> attr(,"bbox")
#>      xmin      ymin      xmax      ymax 
#> -1.756275 -1.684282  2.126850  2.039369 
#> attr(,"class")
#> [1] "bbox"
#> attr(,"z_range")
#> zmin zmax 
#>   NA   NA 
#> attr(,"class")
#> [1] "z_range"
#> attr(,"m_range")
#> mmin mmax 
#>   NA   NA 
#> attr(,"class")
#> [1] "m_range"

sfc_multilinestring( obj = df, x = "x", y = "y", z = "z", m = "m")

#> [[1]]
#> [[1]]
#>             [,1]        [,2]        [,3]       [,4]
#>  [1,] -1.7562754 -1.68428153 -1.14619967  0.7664402
#>  [2,] -0.6905379  0.24940178  0.84618466  0.9799567
#>  [3,] -0.5585420  1.07283825  0.08171963  1.3217810
#>  [4,] -0.5366633  2.03936926 -1.30511701 -1.1197108
#>  [5,]  0.2271271  0.44945378 -0.94491206  0.5145998
#>  [6,]  0.9784549  1.39181405  0.45434159 -1.5090998
#>  [7,] -0.2088827  0.42656655 -0.85520250  1.5327415
#>  [8,] -1.3994105  0.10758399 -0.28689522  0.4291474
#>  [9,]  0.2585373  0.02229473  0.89496163  0.1221034
#> [10,] -0.4417995  0.60361101  0.06730444 -1.1380124
#> [11,]  0.5685999 -0.26265057 -0.16267634 -0.5580151
#> [12,]  2.1268505 -0.52826408 -0.82731017  1.0525385
#> [13,]  0.4248584  0.19214942  1.87650562  0.6776836
#> 
#> attr(,"class")
#> [1] "XYZM"            "MULTILINESTRING" "sfg"            
#> 
#> attr(,"class")
#> [1] "sfc_MULTILINESTRING" "sfc"                
#> attr(,"crs")
#> $epsg
#> [1] NA
#> 
#> $proj4string
#> [1] NA
#> 
#> attr(,"class")
#> [1] "crs"
#> attr(,"precision")
#> [1] 0
#> attr(,"n_empty")
#> [1] 0
#> attr(,"bbox")
#>      xmin      ymin      xmax      ymax 
#> -1.756275 -1.684282  2.126850  2.039369 
#> attr(,"class")
#> [1] "bbox"
#> attr(,"z_range")
#> zmin zmax 
#>   -1    1 
#> attr(,"class")
#> [1] "z_range"
#> attr(,"m_range")
#> mmin mmax 
#>   -1    1 
#> attr(,"class")
#> [1] "m_range"

sfc_multilinestring( obj = df, x = 2, y = 3)

#> [[1]]
#> [[1]]
#>       l_id          x
#>  [1,]    1 -1.7562754
#>  [2,]    1 -0.6905379
#>  [3,]    1 -0.5585420
#>  [4,]    2 -0.5366633
#>  [5,]    2  0.2271271
#>  [6,]    3  0.9784549
#>  [7,]    3 -0.2088827
#>  [8,]    3 -1.3994105
#>  [9,]    1  0.2585373
#> [10,]    1 -0.4417995
#> [11,]    1  0.5685999
#> [12,]    2  2.1268505
#> [13,]    2  0.4248584
#> 
#> attr(,"class")
#> [1] "XY"              "MULTILINESTRING" "sfg"            
#> 
#> attr(,"class")
#> [1] "sfc_MULTILINESTRING" "sfc"                
#> attr(,"crs")
#> $epsg
#> [1] NA
#> 
#> $proj4string
#> [1] NA
#> 
#> attr(,"class")
#> [1] "crs"
#> attr(,"precision")
#> [1] 0
#> attr(,"n_empty")
#> [1] 0
#> attr(,"bbox")
#>      xmin      ymin      xmax      ymax 
#>  1.000000 -1.756275  3.000000  2.126850 
#> attr(,"class")
#> [1] "bbox"
#> attr(,"z_range")
#> zmin zmax 
#>   NA   NA 
#> attr(,"class")
#> [1] "z_range"
#> attr(,"m_range")
#> mmin mmax 
#>   NA   NA 
#> attr(,"class")
#> [1] "m_range"

sfc_multilinestring( obj = df, x = 2, y = 3, z = 4)

#> [[1]]
#> [[1]]
#>       l_id          x           y
#>  [1,]    1 -1.7562754 -1.68428153
#>  [2,]    1 -0.6905379  0.24940178
#>  [3,]    1 -0.5585420  1.07283825
#>  [4,]    2 -0.5366633  2.03936926
#>  [5,]    2  0.2271271  0.44945378
#>  [6,]    3  0.9784549  1.39181405
#>  [7,]    3 -0.2088827  0.42656655
#>  [8,]    3 -1.3994105  0.10758399
#>  [9,]    1  0.2585373  0.02229473
#> [10,]    1 -0.4417995  0.60361101
#> [11,]    1  0.5685999 -0.26265057
#> [12,]    2  2.1268505 -0.52826408
#> [13,]    2  0.4248584  0.19214942
#> 
#> attr(,"class")
#> [1] "XYZ"             "MULTILINESTRING" "sfg"            
#> 
#> attr(,"class")
#> [1] "sfc_MULTILINESTRING" "sfc"                
#> attr(,"crs")
#> $epsg
#> [1] NA
#> 
#> $proj4string
#> [1] NA
#> 
#> attr(,"class")
#> [1] "crs"
#> attr(,"precision")
#> [1] 0
#> attr(,"n_empty")
#> [1] 0
#> attr(,"bbox")
#>      xmin      ymin      xmax      ymax 
#>  1.000000 -1.756275  3.000000  2.126850 
#> attr(,"class")
#> [1] "bbox"
#> attr(,"z_range")
#> zmin zmax 
#>   NA   NA 
#> attr(,"class")
#> [1] "z_range"
#> attr(,"m_range")
#> mmin mmax 
#>   NA   NA 
#> attr(,"class")
#> [1] "m_range"

sfc_multilinestring( obj = df, x = 2, y = 3, z = 4, m = 5)

#> [[1]]
#> [[1]]
#>       l_id          x           y           z
#>  [1,]    1 -1.7562754 -1.68428153 -1.14619967
#>  [2,]    1 -0.6905379  0.24940178  0.84618466
#>  [3,]    1 -0.5585420  1.07283825  0.08171963
#>  [4,]    2 -0.5366633  2.03936926 -1.30511701
#>  [5,]    2  0.2271271  0.44945378 -0.94491206
#>  [6,]    3  0.9784549  1.39181405  0.45434159
#>  [7,]    3 -0.2088827  0.42656655 -0.85520250
#>  [8,]    3 -1.3994105  0.10758399 -0.28689522
#>  [9,]    1  0.2585373  0.02229473  0.89496163
#> [10,]    1 -0.4417995  0.60361101  0.06730444
#> [11,]    1  0.5685999 -0.26265057 -0.16267634
#> [12,]    2  2.1268505 -0.52826408 -0.82731017
#> [13,]    2  0.4248584  0.19214942  1.87650562
#> 
#> attr(,"class")
#> [1] "XYZM"            "MULTILINESTRING" "sfg"            
#> 
#> attr(,"class")
#> [1] "sfc_MULTILINESTRING" "sfc"                
#> attr(,"crs")
#> $epsg
#> [1] NA
#> 
#> $proj4string
#> [1] NA
#> 
#> attr(,"class")
#> [1] "crs"
#> attr(,"precision")
#> [1] 0
#> attr(,"n_empty")
#> [1] 0
#> attr(,"bbox")
#>      xmin      ymin      xmax      ymax 
#>  1.000000 -1.756275  3.000000  2.126850 
#> attr(,"class")
#> [1] "bbox"
#> attr(,"z_range")
#>      zmin      zmax 
#> -1.305117  1.876506 
#> attr(,"class")
#> [1] "z_range"
#> attr(,"m_range")
#>      mmin      mmax 
#> -1.305117  1.876506 
#> attr(,"class")
#> [1] "m_range"

sfc_multilinestring( obj = df, multilinestring_id = "ml_id", linestring_id = "l_id" )

#> [[1]]
#> [[1]]
#>            [,1]       [,2]        [,3]      [,4]
#> [1,] -1.7562754 -1.6842815 -1.14619967 0.7664402
#> [2,] -0.6905379  0.2494018  0.84618466 0.9799567
#> [3,] -0.5585420  1.0728383  0.08171963 1.3217810
#> 
#> [[2]]
#>            [,1]      [,2]       [,3]       [,4]
#> [1,] -0.5366633 2.0393693 -1.3051170 -1.1197108
#> [2,]  0.2271271 0.4494538 -0.9449121  0.5145998
#> 
#> [[3]]
#>            [,1]      [,2]       [,3]       [,4]
#> [1,]  0.9784549 1.3918140  0.4543416 -1.5090998
#> [2,] -0.2088827 0.4265665 -0.8552025  1.5327415
#> [3,] -1.3994105 0.1075840 -0.2868952  0.4291474
#> 
#> attr(,"class")
#> [1] "XYZM"            "MULTILINESTRING" "sfg"            
#> 
#> [[2]]
#> [[1]]
#>            [,1]        [,2]        [,3]       [,4]
#> [1,]  0.2585373  0.02229473  0.89496163  0.1221034
#> [2,] -0.4417995  0.60361101  0.06730444 -1.1380124
#> [3,]  0.5685999 -0.26265057 -0.16267634 -0.5580151
#> 
#> [[2]]
#>           [,1]       [,2]       [,3]      [,4]
#> [1,] 2.1268505 -0.5282641 -0.8273102 1.0525385
#> [2,] 0.4248584  0.1921494  1.8765056 0.6776836
#> 
#> attr(,"class")
#> [1] "XYZM"            "MULTILINESTRING" "sfg"            
#> 
#> attr(,"class")
#> [1] "sfc_MULTILINESTRING" "sfc"                
#> attr(,"crs")
#> $epsg
#> [1] NA
#> 
#> $proj4string
#> [1] NA
#> 
#> attr(,"class")
#> [1] "crs"
#> attr(,"precision")
#> [1] 0
#> attr(,"n_empty")
#> [1] 0
#> attr(,"bbox")
#>      xmin      ymin      xmax      ymax 
#> -1.756275 -1.684282  2.126850  2.039369 
#> attr(,"class")
#> [1] "bbox"
#> attr(,"z_range")
#> zmin zmax 
#>   -1    1 
#> attr(,"class")
#> [1] "z_range"
#> attr(,"m_range")
#> mmin mmax 
#>   -1    1 
#> attr(,"class")
#> [1] "m_range"

sfc_multilinestring( obj = df, multilinestring_id = 1, linestring_id = 2 )

#> [[1]]
#> [[1]]
#>            [,1]       [,2]        [,3]      [,4]
#> [1,] -1.7562754 -1.6842815 -1.14619967 0.7664402
#> [2,] -0.6905379  0.2494018  0.84618466 0.9799567
#> [3,] -0.5585420  1.0728383  0.08171963 1.3217810
#> 
#> [[2]]
#>            [,1]      [,2]       [,3]       [,4]
#> [1,] -0.5366633 2.0393693 -1.3051170 -1.1197108
#> [2,]  0.2271271 0.4494538 -0.9449121  0.5145998
#> 
#> [[3]]
#>            [,1]      [,2]       [,3]       [,4]
#> [1,]  0.9784549 1.3918140  0.4543416 -1.5090998
#> [2,] -0.2088827 0.4265665 -0.8552025  1.5327415
#> [3,] -1.3994105 0.1075840 -0.2868952  0.4291474
#> 
#> attr(,"class")
#> [1] "XYZM"            "MULTILINESTRING" "sfg"            
#> 
#> [[2]]
#> [[1]]
#>            [,1]        [,2]        [,3]       [,4]
#> [1,]  0.2585373  0.02229473  0.89496163  0.1221034
#> [2,] -0.4417995  0.60361101  0.06730444 -1.1380124
#> [3,]  0.5685999 -0.26265057 -0.16267634 -0.5580151
#> 
#> [[2]]
#>           [,1]       [,2]       [,3]      [,4]
#> [1,] 2.1268505 -0.5282641 -0.8273102 1.0525385
#> [2,] 0.4248584  0.1921494  1.8765056 0.6776836
#> 
#> attr(,"class")
#> [1] "XYZM"            "MULTILINESTRING" "sfg"            
#> 
#> attr(,"class")
#> [1] "sfc_MULTILINESTRING" "sfc"                
#> attr(,"crs")
#> $epsg
#> [1] NA
#> 
#> $proj4string
#> [1] NA
#> 
#> attr(,"class")
#> [1] "crs"
#> attr(,"precision")
#> [1] 0
#> attr(,"n_empty")
#> [1] 0
#> attr(,"bbox")
#>      xmin      ymin      xmax      ymax 
#> -1.756275 -1.684282  2.126850  2.039369 
#> attr(,"class")
#> [1] "bbox"
#> attr(,"z_range")
#>      zmin      zmax 
#> -1.509100  1.532741 
#> attr(,"class")
#> [1] "z_range"
#> attr(,"m_range")
#>      mmin      mmax 
#> -1.509100  1.532741 
#> attr(,"class")
#> [1] "m_range"