sfc MULTILINESTRING

constructs an sfc of MULTILINESTRING objects

sfc_multilinestring(
  obj = NULL,
  x = NULL,
  y = NULL,
  z = NULL,
  m = NULL,
  multilinestring_id = NULL,
  linestring_id = NULL
)

Arguments

obj

sorted matrix or data.frame

x

x geometry column

y

y geometry column

z

z geometry column

m

m geometry column

multilinestring_id

column of ids for multilinestrings

linestring_id

column of ids for linestrings (within multilinestrings)

Value

sfc object of MULTILINESTRING geometries

notes

sfheaders functions do not perform any validity checks on the geometries. Nor do they set Coordinate Reference Systems, EPSG, PROJ4 or precision attributes.

The data.frame and matrices you send into the sfheader functions must be ordered.

Examples

m <- matrix(c(0,0,0,0,1,1), ncol = 3 )
sfc_multilinestring( m )
#> [[1]]
#> [[1]]
#>      [,1] [,2] [,3]
#> [1,]    0    0    1
#> [2,]    0    0    1
#> 
#> attr(,"class")
#> [1] "XYZ"             "MULTILINESTRING" "sfg"            
#> 
#> attr(,"n_empty")
#> [1] 0
#> attr(,"crs")
#> $input
#> [1] NA
#> 
#> $wkt
#> [1] NA
#> 
#> attr(,"class")
#> [1] "crs"
#> attr(,"class")
#> [1] "sfc_MULTILINESTRING" "sfc"                
#> attr(,"precision")
#> [1] 0
#> attr(,"bbox")
#> xmin ymin xmax ymax 
#>    0    0    0    0 
#> attr(,"class")
#> [1] "bbox"
#> attr(,"z_range")
#> zmin zmax 
#>    1    1 
#> attr(,"class")
#> [1] "z_range"

m <- matrix(c(0,0,0,0,0,1,0,1,1,1,2,2,1,2,3), ncol = 3, byrow = TRUE)
sfc_multilinestring( obj = m )
#> [[1]]
#> [[1]]
#>      [,1] [,2] [,3]
#> [1,]    0    0    0
#> [2,]    0    0    1
#> [3,]    0    1    1
#> [4,]    1    2    2
#> [5,]    1    2    3
#> 
#> attr(,"class")
#> [1] "XYZ"             "MULTILINESTRING" "sfg"            
#> 
#> attr(,"n_empty")
#> [1] 0
#> attr(,"crs")
#> $input
#> [1] NA
#> 
#> $wkt
#> [1] NA
#> 
#> attr(,"class")
#> [1] "crs"
#> attr(,"class")
#> [1] "sfc_MULTILINESTRING" "sfc"                
#> attr(,"precision")
#> [1] 0
#> attr(,"bbox")
#> xmin ymin xmax ymax 
#>    0    0    1    2 
#> attr(,"class")
#> [1] "bbox"
#> attr(,"z_range")
#> zmin zmax 
#>    0    3 
#> attr(,"class")
#> [1] "z_range"
sfc_multilinestring( obj = m, multilinestring_id = 1 )
#> [[1]]
#> [[1]]
#>      [,1] [,2]
#> [1,]    0    0
#> [2,]    0    1
#> [3,]    1    1
#> 
#> attr(,"class")
#> [1] "XY"              "MULTILINESTRING" "sfg"            
#> 
#> [[2]]
#> [[1]]
#>      [,1] [,2]
#> [1,]    2    2
#> [2,]    2    3
#> 
#> attr(,"class")
#> [1] "XY"              "MULTILINESTRING" "sfg"            
#> 
#> attr(,"n_empty")
#> [1] 0
#> attr(,"crs")
#> $input
#> [1] NA
#> 
#> $wkt
#> [1] NA
#> 
#> attr(,"class")
#> [1] "crs"
#> attr(,"class")
#> [1] "sfc_MULTILINESTRING" "sfc"                
#> attr(,"precision")
#> [1] 0
#> attr(,"bbox")
#> xmin ymin xmax ymax 
#>    0    0    2    3 
#> attr(,"class")
#> [1] "bbox"
sfc_multilinestring( obj = m, linestring_id = 1 )
#> [[1]]
#> [[1]]
#>      [,1] [,2]
#> [1,]    0    0
#> [2,]    0    1
#> [3,]    1    1
#> 
#> [[2]]
#>      [,1] [,2]
#> [1,]    2    2
#> [2,]    2    3
#> 
#> attr(,"class")
#> [1] "XY"              "MULTILINESTRING" "sfg"            
#> 
#> attr(,"n_empty")
#> [1] 0
#> attr(,"crs")
#> $input
#> [1] NA
#> 
#> $wkt
#> [1] NA
#> 
#> attr(,"class")
#> [1] "crs"
#> attr(,"class")
#> [1] "sfc_MULTILINESTRING" "sfc"                
#> attr(,"precision")
#> [1] 0
#> attr(,"bbox")
#> xmin ymin xmax ymax 
#>    0    0    2    3 
#> attr(,"class")
#> [1] "bbox"

sfc_multilinestring( obj = m, linestring_id = 1, multilinestring_id = 1 )
#> [[1]]
#> [[1]]
#>      [,1] [,2]
#> [1,]    0    0
#> [2,]    0    1
#> [3,]    1    1
#> 
#> attr(,"class")
#> [1] "XY"              "MULTILINESTRING" "sfg"            
#> 
#> [[2]]
#> [[1]]
#>      [,1] [,2]
#> [1,]    2    2
#> [2,]    2    3
#> 
#> attr(,"class")
#> [1] "XY"              "MULTILINESTRING" "sfg"            
#> 
#> attr(,"n_empty")
#> [1] 0
#> attr(,"crs")
#> $input
#> [1] NA
#> 
#> $wkt
#> [1] NA
#> 
#> attr(,"class")
#> [1] "crs"
#> attr(,"class")
#> [1] "sfc_MULTILINESTRING" "sfc"                
#> attr(,"precision")
#> [1] 0
#> attr(,"bbox")
#> xmin ymin xmax ymax 
#>    0    0    2    3 
#> attr(,"class")
#> [1] "bbox"

sfc_multilinestring( obj = m, x = 2, y = 3 )
#> [[1]]
#> [[1]]
#>      [,1] [,2]
#> [1,]    0    0
#> [2,]    0    1
#> [3,]    1    1
#> [4,]    2    2
#> [5,]    2    3
#> 
#> attr(,"class")
#> [1] "XY"              "MULTILINESTRING" "sfg"            
#> 
#> attr(,"n_empty")
#> [1] 0
#> attr(,"crs")
#> $input
#> [1] NA
#> 
#> $wkt
#> [1] NA
#> 
#> attr(,"class")
#> [1] "crs"
#> attr(,"class")
#> [1] "sfc_MULTILINESTRING" "sfc"                
#> attr(,"precision")
#> [1] 0
#> attr(,"bbox")
#> xmin ymin xmax ymax 
#>    0    0    2    3 
#> attr(,"class")
#> [1] "bbox"
sfc_multilinestring( obj = m, x = 1, y = 2, z = 3 )
#> [[1]]
#> [[1]]
#>      [,1] [,2] [,3]
#> [1,]    0    0    0
#> [2,]    0    0    1
#> [3,]    0    1    1
#> [4,]    1    2    2
#> [5,]    1    2    3
#> 
#> attr(,"class")
#> [1] "XYZ"             "MULTILINESTRING" "sfg"            
#> 
#> attr(,"n_empty")
#> [1] 0
#> attr(,"crs")
#> $input
#> [1] NA
#> 
#> $wkt
#> [1] NA
#> 
#> attr(,"class")
#> [1] "crs"
#> attr(,"class")
#> [1] "sfc_MULTILINESTRING" "sfc"                
#> attr(,"precision")
#> [1] 0
#> attr(,"bbox")
#> xmin ymin xmax ymax 
#>    0    0    1    2 
#> attr(,"class")
#> [1] "bbox"
#> attr(,"z_range")
#> zmin zmax 
#>    0    3 
#> attr(,"class")
#> [1] "z_range"
sfc_multilinestring( obj = m, x = 2, y = 3, linestring_id = 1, multilinestring_id = 1 )
#> [[1]]
#> [[1]]
#>      [,1] [,2]
#> [1,]    0    0
#> [2,]    0    1
#> [3,]    1    1
#> 
#> attr(,"class")
#> [1] "XY"              "MULTILINESTRING" "sfg"            
#> 
#> [[2]]
#> [[1]]
#>      [,1] [,2]
#> [1,]    2    2
#> [2,]    2    3
#> 
#> attr(,"class")
#> [1] "XY"              "MULTILINESTRING" "sfg"            
#> 
#> attr(,"n_empty")
#> [1] 0
#> attr(,"crs")
#> $input
#> [1] NA
#> 
#> $wkt
#> [1] NA
#> 
#> attr(,"class")
#> [1] "crs"
#> attr(,"class")
#> [1] "sfc_MULTILINESTRING" "sfc"                
#> attr(,"precision")
#> [1] 0
#> attr(,"bbox")
#> xmin ymin xmax ymax 
#>    0    0    2    3 
#> attr(,"class")
#> [1] "bbox"

df <- data.frame(
  ml_id = c(1,1,1,1,1,1,1,1,2,2,2,2,2)
  , l_id = c(1,1,1,2,2,3,3,3,1,1,1,2,2)
  , x = rnorm(13)
  , y = rnorm(13)
  , z = rnorm(13)
  , m = rnorm(13)
)

sfc_multilinestring( obj = df, x = "x", y = "y")
#> [[1]]
#> [[1]]
#>              [,1]       [,2]
#>  [1,] -0.17113218  2.8493675
#>  [2,] -0.05383097 -1.3306830
#>  [3,] -1.36852585 -0.6260212
#>  [4,]  0.06963977 -0.2340877
#>  [5,]  0.93069018  0.1663073
#>  [6,]  0.87617648  1.3289881
#>  [7,]  1.63090671  0.7655132
#>  [8,]  1.28302693 -0.2148770
#>  [9,]  0.76316725 -0.8810034
#> [10,]  0.62404350  0.7613431
#> [11,]  1.46148617 -2.6856065
#> [12,]  0.16623151  1.9254526
#> [13,] -0.12879331  0.8608860
#> 
#> attr(,"class")
#> [1] "XY"              "MULTILINESTRING" "sfg"            
#> 
#> attr(,"n_empty")
#> [1] 0
#> attr(,"crs")
#> $input
#> [1] NA
#> 
#> $wkt
#> [1] NA
#> 
#> attr(,"class")
#> [1] "crs"
#> attr(,"class")
#> [1] "sfc_MULTILINESTRING" "sfc"                
#> attr(,"precision")
#> [1] 0
#> attr(,"bbox")
#>      xmin      ymin      xmax      ymax 
#> -1.368526 -2.685607  1.630907  2.849367 
#> attr(,"class")
#> [1] "bbox"
sfc_multilinestring( obj = df, x = "x", y = "y", z = "z")
#> [[1]]
#> [[1]]
#>              [,1]       [,2]       [,3]
#>  [1,] -0.17113218  2.8493675 -0.1385466
#>  [2,] -0.05383097 -1.3306830  1.6998524
#>  [3,] -1.36852585 -0.6260212  1.0746061
#>  [4,]  0.06963977 -0.2340877  0.5345897
#>  [5,]  0.93069018  0.1663073  0.2454938
#>  [6,]  0.87617648  1.3289881 -1.1947430
#>  [7,]  1.63090671  0.7655132 -0.8113544
#>  [8,]  1.28302693 -0.2148770 -0.7823010
#>  [9,]  0.76316725 -0.8810034 -0.9000242
#> [10,]  0.62404350  0.7613431  1.3621125
#> [11,]  1.46148617 -2.6856065 -1.2295957
#> [12,]  0.16623151  1.9254526 -0.3285105
#> [13,] -0.12879331  0.8608860 -0.8642367
#> 
#> attr(,"class")
#> [1] "XYZ"             "MULTILINESTRING" "sfg"            
#> 
#> attr(,"n_empty")
#> [1] 0
#> attr(,"crs")
#> $input
#> [1] NA
#> 
#> $wkt
#> [1] NA
#> 
#> attr(,"class")
#> [1] "crs"
#> attr(,"class")
#> [1] "sfc_MULTILINESTRING" "sfc"                
#> attr(,"precision")
#> [1] 0
#> attr(,"bbox")
#>      xmin      ymin      xmax      ymax 
#> -1.368526 -2.685607  1.630907  2.849367 
#> attr(,"class")
#> [1] "bbox"
#> attr(,"z_range")
#>      zmin      zmax 
#> -1.229596  1.699852 
#> attr(,"class")
#> [1] "z_range"
sfc_multilinestring( obj = df, x = "x", y = "y", z = "z", m = "m")
#> [[1]]
#> [[1]]
#>              [,1]       [,2]       [,3]        [,4]
#>  [1,] -0.17113218  2.8493675 -0.1385466  1.19885302
#>  [2,] -0.05383097 -1.3306830  1.6998524 -0.54732351
#>  [3,] -1.36852585 -0.6260212  1.0746061  1.29116736
#>  [4,]  0.06963977 -0.2340877  0.5345897  0.36075815
#>  [5,]  0.93069018  0.1663073  0.2454938 -0.14023658
#>  [6,]  0.87617648  1.3289881 -1.1947430 -0.04717296
#>  [7,]  1.63090671  0.7655132 -0.8113544 -0.30494116
#>  [8,]  1.28302693 -0.2148770 -0.7823010 -0.86102542
#>  [9,]  0.76316725 -0.8810034 -0.9000242 -0.20015371
#> [10,]  0.62404350  0.7613431  1.3621125  1.72133965
#> [11,]  1.46148617 -2.6856065 -1.2295957 -0.54465577
#> [12,]  0.16623151  1.9254526 -0.3285105  0.82067843
#> [13,] -0.12879331  0.8608860 -0.8642367  0.39976338
#> 
#> attr(,"class")
#> [1] "XYZM"            "MULTILINESTRING" "sfg"            
#> 
#> attr(,"n_empty")
#> [1] 0
#> attr(,"crs")
#> $input
#> [1] NA
#> 
#> $wkt
#> [1] NA
#> 
#> attr(,"class")
#> [1] "crs"
#> attr(,"class")
#> [1] "sfc_MULTILINESTRING" "sfc"                
#> attr(,"precision")
#> [1] 0
#> attr(,"bbox")
#>      xmin      ymin      xmax      ymax 
#> -1.368526 -2.685607  1.630907  2.849367 
#> attr(,"class")
#> [1] "bbox"
#> attr(,"z_range")
#>      zmin      zmax 
#> -1.229596  1.699852 
#> attr(,"class")
#> [1] "z_range"
#> attr(,"m_range")
#>       mmin       mmax 
#> -0.8610254  1.7213397 
#> attr(,"class")
#> [1] "m_range"

sfc_multilinestring( obj = df, x = 2, y = 3)
#> [[1]]
#> [[1]]
#>       [,1]        [,2]
#>  [1,]    1 -0.17113218
#>  [2,]    1 -0.05383097
#>  [3,]    1 -1.36852585
#>  [4,]    2  0.06963977
#>  [5,]    2  0.93069018
#>  [6,]    3  0.87617648
#>  [7,]    3  1.63090671
#>  [8,]    3  1.28302693
#>  [9,]    1  0.76316725
#> [10,]    1  0.62404350
#> [11,]    1  1.46148617
#> [12,]    2  0.16623151
#> [13,]    2 -0.12879331
#> 
#> attr(,"class")
#> [1] "XY"              "MULTILINESTRING" "sfg"            
#> 
#> attr(,"n_empty")
#> [1] 0
#> attr(,"crs")
#> $input
#> [1] NA
#> 
#> $wkt
#> [1] NA
#> 
#> attr(,"class")
#> [1] "crs"
#> attr(,"class")
#> [1] "sfc_MULTILINESTRING" "sfc"                
#> attr(,"precision")
#> [1] 0
#> attr(,"bbox")
#>      xmin      ymin      xmax      ymax 
#>  1.000000 -1.368526  3.000000  1.630907 
#> attr(,"class")
#> [1] "bbox"
sfc_multilinestring( obj = df, x = 2, y = 3, z = 4)
#> [[1]]
#> [[1]]
#>       [,1]        [,2]       [,3]
#>  [1,]    1 -0.17113218  2.8493675
#>  [2,]    1 -0.05383097 -1.3306830
#>  [3,]    1 -1.36852585 -0.6260212
#>  [4,]    2  0.06963977 -0.2340877
#>  [5,]    2  0.93069018  0.1663073
#>  [6,]    3  0.87617648  1.3289881
#>  [7,]    3  1.63090671  0.7655132
#>  [8,]    3  1.28302693 -0.2148770
#>  [9,]    1  0.76316725 -0.8810034
#> [10,]    1  0.62404350  0.7613431
#> [11,]    1  1.46148617 -2.6856065
#> [12,]    2  0.16623151  1.9254526
#> [13,]    2 -0.12879331  0.8608860
#> 
#> attr(,"class")
#> [1] "XYZ"             "MULTILINESTRING" "sfg"            
#> 
#> attr(,"n_empty")
#> [1] 0
#> attr(,"crs")
#> $input
#> [1] NA
#> 
#> $wkt
#> [1] NA
#> 
#> attr(,"class")
#> [1] "crs"
#> attr(,"class")
#> [1] "sfc_MULTILINESTRING" "sfc"                
#> attr(,"precision")
#> [1] 0
#> attr(,"bbox")
#>      xmin      ymin      xmax      ymax 
#>  1.000000 -1.368526  3.000000  1.630907 
#> attr(,"class")
#> [1] "bbox"
#> attr(,"z_range")
#>      zmin      zmax 
#> -2.685607  2.849367 
#> attr(,"class")
#> [1] "z_range"
sfc_multilinestring( obj = df, x = 2, y = 3, z = 4, m = 5)
#> [[1]]
#> [[1]]
#>       [,1]        [,2]       [,3]       [,4]
#>  [1,]    1 -0.17113218  2.8493675 -0.1385466
#>  [2,]    1 -0.05383097 -1.3306830  1.6998524
#>  [3,]    1 -1.36852585 -0.6260212  1.0746061
#>  [4,]    2  0.06963977 -0.2340877  0.5345897
#>  [5,]    2  0.93069018  0.1663073  0.2454938
#>  [6,]    3  0.87617648  1.3289881 -1.1947430
#>  [7,]    3  1.63090671  0.7655132 -0.8113544
#>  [8,]    3  1.28302693 -0.2148770 -0.7823010
#>  [9,]    1  0.76316725 -0.8810034 -0.9000242
#> [10,]    1  0.62404350  0.7613431  1.3621125
#> [11,]    1  1.46148617 -2.6856065 -1.2295957
#> [12,]    2  0.16623151  1.9254526 -0.3285105
#> [13,]    2 -0.12879331  0.8608860 -0.8642367
#> 
#> attr(,"class")
#> [1] "XYZM"            "MULTILINESTRING" "sfg"            
#> 
#> attr(,"n_empty")
#> [1] 0
#> attr(,"crs")
#> $input
#> [1] NA
#> 
#> $wkt
#> [1] NA
#> 
#> attr(,"class")
#> [1] "crs"
#> attr(,"class")
#> [1] "sfc_MULTILINESTRING" "sfc"                
#> attr(,"precision")
#> [1] 0
#> attr(,"bbox")
#>      xmin      ymin      xmax      ymax 
#>  1.000000 -1.368526  3.000000  1.630907 
#> attr(,"class")
#> [1] "bbox"
#> attr(,"z_range")
#>      zmin      zmax 
#> -2.685607  2.849367 
#> attr(,"class")
#> [1] "z_range"
#> attr(,"m_range")
#>      mmin      mmax 
#> -1.229596  1.699852 
#> attr(,"class")
#> [1] "m_range"

sfc_multilinestring( obj = df, multilinestring_id = "ml_id", linestring_id = "l_id" )
#> [[1]]
#> [[1]]
#>             [,1]       [,2]       [,3]       [,4]
#> [1,] -0.17113218  2.8493675 -0.1385466  1.1988530
#> [2,] -0.05383097 -1.3306830  1.6998524 -0.5473235
#> [3,] -1.36852585 -0.6260212  1.0746061  1.2911674
#> 
#> [[2]]
#>            [,1]       [,2]      [,3]       [,4]
#> [1,] 0.06963977 -0.2340877 0.5345897  0.3607581
#> [2,] 0.93069018  0.1663073 0.2454938 -0.1402366
#> 
#> [[3]]
#>           [,1]       [,2]       [,3]        [,4]
#> [1,] 0.8761765  1.3289881 -1.1947430 -0.04717296
#> [2,] 1.6309067  0.7655132 -0.8113544 -0.30494116
#> [3,] 1.2830269 -0.2148770 -0.7823010 -0.86102542
#> 
#> attr(,"class")
#> [1] "XYZM"            "MULTILINESTRING" "sfg"            
#> 
#> [[2]]
#> [[1]]
#>           [,1]       [,2]       [,3]       [,4]
#> [1,] 0.7631672 -0.8810034 -0.9000242 -0.2001537
#> [2,] 0.6240435  0.7613431  1.3621125  1.7213397
#> [3,] 1.4614862 -2.6856065 -1.2295957 -0.5446558
#> 
#> [[2]]
#>            [,1]     [,2]       [,3]      [,4]
#> [1,]  0.1662315 1.925453 -0.3285105 0.8206784
#> [2,] -0.1287933 0.860886 -0.8642367 0.3997634
#> 
#> attr(,"class")
#> [1] "XYZM"            "MULTILINESTRING" "sfg"            
#> 
#> attr(,"n_empty")
#> [1] 0
#> attr(,"crs")
#> $input
#> [1] NA
#> 
#> $wkt
#> [1] NA
#> 
#> attr(,"class")
#> [1] "crs"
#> attr(,"class")
#> [1] "sfc_MULTILINESTRING" "sfc"                
#> attr(,"precision")
#> [1] 0
#> attr(,"bbox")
#>      xmin      ymin      xmax      ymax 
#> -1.368526 -2.685607  1.630907  2.849367 
#> attr(,"class")
#> [1] "bbox"
#> attr(,"z_range")
#>      zmin      zmax 
#> -1.229596  1.699852 
#> attr(,"class")
#> [1] "z_range"
#> attr(,"m_range")
#>       mmin       mmax 
#> -0.8610254  1.7213397 
#> attr(,"class")
#> [1] "m_range"
sfc_multilinestring( obj = df, multilinestring_id = 1, linestring_id = 2 )
#> [[1]]
#> [[1]]
#>             [,1]       [,2]       [,3]       [,4]
#> [1,] -0.17113218  2.8493675 -0.1385466  1.1988530
#> [2,] -0.05383097 -1.3306830  1.6998524 -0.5473235
#> [3,] -1.36852585 -0.6260212  1.0746061  1.2911674
#> 
#> [[2]]
#>            [,1]       [,2]      [,3]       [,4]
#> [1,] 0.06963977 -0.2340877 0.5345897  0.3607581
#> [2,] 0.93069018  0.1663073 0.2454938 -0.1402366
#> 
#> [[3]]
#>           [,1]       [,2]       [,3]        [,4]
#> [1,] 0.8761765  1.3289881 -1.1947430 -0.04717296
#> [2,] 1.6309067  0.7655132 -0.8113544 -0.30494116
#> [3,] 1.2830269 -0.2148770 -0.7823010 -0.86102542
#> 
#> attr(,"class")
#> [1] "XYZM"            "MULTILINESTRING" "sfg"            
#> 
#> [[2]]
#> [[1]]
#>           [,1]       [,2]       [,3]       [,4]
#> [1,] 0.7631672 -0.8810034 -0.9000242 -0.2001537
#> [2,] 0.6240435  0.7613431  1.3621125  1.7213397
#> [3,] 1.4614862 -2.6856065 -1.2295957 -0.5446558
#> 
#> [[2]]
#>            [,1]     [,2]       [,3]      [,4]
#> [1,]  0.1662315 1.925453 -0.3285105 0.8206784
#> [2,] -0.1287933 0.860886 -0.8642367 0.3997634
#> 
#> attr(,"class")
#> [1] "XYZM"            "MULTILINESTRING" "sfg"            
#> 
#> attr(,"n_empty")
#> [1] 0
#> attr(,"crs")
#> $input
#> [1] NA
#> 
#> $wkt
#> [1] NA
#> 
#> attr(,"class")
#> [1] "crs"
#> attr(,"class")
#> [1] "sfc_MULTILINESTRING" "sfc"                
#> attr(,"precision")
#> [1] 0
#> attr(,"bbox")
#>      xmin      ymin      xmax      ymax 
#> -1.368526 -2.685607  1.630907  2.849367 
#> attr(,"class")
#> [1] "bbox"
#> attr(,"z_range")
#>      zmin      zmax 
#> -1.229596  1.699852 
#> attr(,"class")
#> [1] "z_range"
#> attr(,"m_range")
#>       mmin       mmax 
#> -0.8610254  1.7213397 
#> attr(,"class")
#> [1] "m_range"