sfc MULTILINESTRING

constructs an sfc of MULTILINESTRING objects

sfc_multilinestring(
  obj = NULL,
  x = NULL,
  y = NULL,
  z = NULL,
  m = NULL,
  multilinestring_id = NULL,
  linestring_id = NULL
)

Arguments

obj

sorted matrix or data.frame

x

x geometry column

y

y geometry column

z

z geometry column

m

m geometry column

multilinestring_id

column of ids for multilinestrings

linestring_id

column of ids for linestrings (within multilinestrings)

Value

sfc object of MULTILINESTRING geometries

notes

sfheaders functions do not perform any validity checks on the geometries. Nor do they set Coordinate Reference Systems, EPSG, PROJ4 or precision attributes.

The data.frame and matrices you send into the sfheader functions must be ordered.

Examples

m <- matrix(c(0,0,0,0,1,1), ncol = 3 )
sfc_multilinestring( m )
#> [[1]]
#> [[1]]
#>      [,1] [,2] [,3]
#> [1,]    0    0    1
#> [2,]    0    0    1
#> 
#> attr(,"class")
#> [1] "XYZ"             "MULTILINESTRING" "sfg"            
#> 
#> attr(,"n_empty")
#> [1] 0
#> attr(,"crs")
#> $input
#> [1] NA
#> 
#> $wkt
#> [1] NA
#> 
#> attr(,"class")
#> [1] "crs"
#> attr(,"class")
#> [1] "sfc_MULTILINESTRING" "sfc"                
#> attr(,"precision")
#> [1] 0
#> attr(,"bbox")
#> xmin ymin xmax ymax 
#>    0    0    0    0 
#> attr(,"class")
#> [1] "bbox"
#> attr(,"z_range")
#> zmin zmax 
#>    1    1 
#> attr(,"class")
#> [1] "z_range"

m <- matrix(c(0,0,0,0,0,1,0,1,1,1,2,2,1,2,3), ncol = 3, byrow = TRUE)
sfc_multilinestring( obj = m )
#> [[1]]
#> [[1]]
#>      [,1] [,2] [,3]
#> [1,]    0    0    0
#> [2,]    0    0    1
#> [3,]    0    1    1
#> [4,]    1    2    2
#> [5,]    1    2    3
#> 
#> attr(,"class")
#> [1] "XYZ"             "MULTILINESTRING" "sfg"            
#> 
#> attr(,"n_empty")
#> [1] 0
#> attr(,"crs")
#> $input
#> [1] NA
#> 
#> $wkt
#> [1] NA
#> 
#> attr(,"class")
#> [1] "crs"
#> attr(,"class")
#> [1] "sfc_MULTILINESTRING" "sfc"                
#> attr(,"precision")
#> [1] 0
#> attr(,"bbox")
#> xmin ymin xmax ymax 
#>    0    0    1    2 
#> attr(,"class")
#> [1] "bbox"
#> attr(,"z_range")
#> zmin zmax 
#>    0    3 
#> attr(,"class")
#> [1] "z_range"
sfc_multilinestring( obj = m, multilinestring_id = 1 )
#> [[1]]
#> [[1]]
#>      [,1] [,2]
#> [1,]    0    0
#> [2,]    0    1
#> [3,]    1    1
#> 
#> attr(,"class")
#> [1] "XY"              "MULTILINESTRING" "sfg"            
#> 
#> [[2]]
#> [[1]]
#>      [,1] [,2]
#> [1,]    2    2
#> [2,]    2    3
#> 
#> attr(,"class")
#> [1] "XY"              "MULTILINESTRING" "sfg"            
#> 
#> attr(,"n_empty")
#> [1] 0
#> attr(,"crs")
#> $input
#> [1] NA
#> 
#> $wkt
#> [1] NA
#> 
#> attr(,"class")
#> [1] "crs"
#> attr(,"class")
#> [1] "sfc_MULTILINESTRING" "sfc"                
#> attr(,"precision")
#> [1] 0
#> attr(,"bbox")
#> xmin ymin xmax ymax 
#>    0    0    2    3 
#> attr(,"class")
#> [1] "bbox"
sfc_multilinestring( obj = m, linestring_id = 1 )
#> [[1]]
#> [[1]]
#>      [,1] [,2]
#> [1,]    0    0
#> [2,]    0    1
#> [3,]    1    1
#> 
#> [[2]]
#>      [,1] [,2]
#> [1,]    2    2
#> [2,]    2    3
#> 
#> attr(,"class")
#> [1] "XY"              "MULTILINESTRING" "sfg"            
#> 
#> attr(,"n_empty")
#> [1] 0
#> attr(,"crs")
#> $input
#> [1] NA
#> 
#> $wkt
#> [1] NA
#> 
#> attr(,"class")
#> [1] "crs"
#> attr(,"class")
#> [1] "sfc_MULTILINESTRING" "sfc"                
#> attr(,"precision")
#> [1] 0
#> attr(,"bbox")
#> xmin ymin xmax ymax 
#>    0    0    2    3 
#> attr(,"class")
#> [1] "bbox"

sfc_multilinestring( obj = m, linestring_id = 1, multilinestring_id = 1 )
#> [[1]]
#> [[1]]
#>      [,1] [,2]
#> [1,]    0    0
#> [2,]    0    1
#> [3,]    1    1
#> 
#> attr(,"class")
#> [1] "XY"              "MULTILINESTRING" "sfg"            
#> 
#> [[2]]
#> [[1]]
#>      [,1] [,2]
#> [1,]    2    2
#> [2,]    2    3
#> 
#> attr(,"class")
#> [1] "XY"              "MULTILINESTRING" "sfg"            
#> 
#> attr(,"n_empty")
#> [1] 0
#> attr(,"crs")
#> $input
#> [1] NA
#> 
#> $wkt
#> [1] NA
#> 
#> attr(,"class")
#> [1] "crs"
#> attr(,"class")
#> [1] "sfc_MULTILINESTRING" "sfc"                
#> attr(,"precision")
#> [1] 0
#> attr(,"bbox")
#> xmin ymin xmax ymax 
#>    0    0    2    3 
#> attr(,"class")
#> [1] "bbox"

sfc_multilinestring( obj = m, x = 2, y = 3 )
#> [[1]]
#> [[1]]
#>      [,1] [,2]
#> [1,]    0    0
#> [2,]    0    1
#> [3,]    1    1
#> [4,]    2    2
#> [5,]    2    3
#> 
#> attr(,"class")
#> [1] "XY"              "MULTILINESTRING" "sfg"            
#> 
#> attr(,"n_empty")
#> [1] 0
#> attr(,"crs")
#> $input
#> [1] NA
#> 
#> $wkt
#> [1] NA
#> 
#> attr(,"class")
#> [1] "crs"
#> attr(,"class")
#> [1] "sfc_MULTILINESTRING" "sfc"                
#> attr(,"precision")
#> [1] 0
#> attr(,"bbox")
#> xmin ymin xmax ymax 
#>    0    0    2    3 
#> attr(,"class")
#> [1] "bbox"
sfc_multilinestring( obj = m, x = 1, y = 2, z = 3 )
#> [[1]]
#> [[1]]
#>      [,1] [,2] [,3]
#> [1,]    0    0    0
#> [2,]    0    0    1
#> [3,]    0    1    1
#> [4,]    1    2    2
#> [5,]    1    2    3
#> 
#> attr(,"class")
#> [1] "XYZ"             "MULTILINESTRING" "sfg"            
#> 
#> attr(,"n_empty")
#> [1] 0
#> attr(,"crs")
#> $input
#> [1] NA
#> 
#> $wkt
#> [1] NA
#> 
#> attr(,"class")
#> [1] "crs"
#> attr(,"class")
#> [1] "sfc_MULTILINESTRING" "sfc"                
#> attr(,"precision")
#> [1] 0
#> attr(,"bbox")
#> xmin ymin xmax ymax 
#>    0    0    1    2 
#> attr(,"class")
#> [1] "bbox"
#> attr(,"z_range")
#> zmin zmax 
#>    0    3 
#> attr(,"class")
#> [1] "z_range"
sfc_multilinestring( obj = m, x = 2, y = 3, linestring_id = 1, multilinestring_id = 1 )
#> [[1]]
#> [[1]]
#>      [,1] [,2]
#> [1,]    0    0
#> [2,]    0    1
#> [3,]    1    1
#> 
#> attr(,"class")
#> [1] "XY"              "MULTILINESTRING" "sfg"            
#> 
#> [[2]]
#> [[1]]
#>      [,1] [,2]
#> [1,]    2    2
#> [2,]    2    3
#> 
#> attr(,"class")
#> [1] "XY"              "MULTILINESTRING" "sfg"            
#> 
#> attr(,"n_empty")
#> [1] 0
#> attr(,"crs")
#> $input
#> [1] NA
#> 
#> $wkt
#> [1] NA
#> 
#> attr(,"class")
#> [1] "crs"
#> attr(,"class")
#> [1] "sfc_MULTILINESTRING" "sfc"                
#> attr(,"precision")
#> [1] 0
#> attr(,"bbox")
#> xmin ymin xmax ymax 
#>    0    0    2    3 
#> attr(,"class")
#> [1] "bbox"

df <- data.frame(
  ml_id = c(1,1,1,1,1,1,1,1,2,2,2,2,2)
  , l_id = c(1,1,1,2,2,3,3,3,1,1,1,2,2)
  , x = rnorm(13)
  , y = rnorm(13)
  , z = rnorm(13)
  , m = rnorm(13)
)

sfc_multilinestring( obj = df, x = "x", y = "y")
#> [[1]]
#> [[1]]
#>              [,1]        [,2]
#>  [1,] -0.72477040  0.73419583
#>  [2,] -0.35745209  0.65115182
#>  [3,]  0.48089085 -0.61471396
#>  [4,] -1.16053843 -0.86069844
#>  [5,]  0.63335358 -0.01859121
#>  [6,]  0.69714383  0.34643868
#>  [7,]  0.76230125  1.55795821
#>  [8,]  0.44560194  1.08564328
#>  [9,]  0.13838995  0.13707904
#> [10,] -0.46347664  0.47264776
#> [11,] -0.04192427 -1.60368661
#> [12,]  0.23126611  1.01782745
#> [13,] -0.19218543 -0.94927441
#> 
#> attr(,"class")
#> [1] "XY"              "MULTILINESTRING" "sfg"            
#> 
#> attr(,"n_empty")
#> [1] 0
#> attr(,"crs")
#> $input
#> [1] NA
#> 
#> $wkt
#> [1] NA
#> 
#> attr(,"class")
#> [1] "crs"
#> attr(,"class")
#> [1] "sfc_MULTILINESTRING" "sfc"                
#> attr(,"precision")
#> [1] 0
#> attr(,"bbox")
#>       xmin       ymin       xmax       ymax 
#> -1.1605384 -1.6036866  0.7623013  1.5579582 
#> attr(,"class")
#> [1] "bbox"
sfc_multilinestring( obj = df, x = "x", y = "y", z = "z")
#> [[1]]
#> [[1]]
#>              [,1]        [,2]       [,3]
#>  [1,] -0.72477040  0.73419583 -0.1572917
#>  [2,] -0.35745209  0.65115182 -0.6446505
#>  [3,]  0.48089085 -0.61471396  1.0675388
#>  [4,] -1.16053843 -0.86069844  1.2046122
#>  [5,]  0.63335358 -0.01859121 -0.8560544
#>  [6,]  0.69714383  0.34643868 -0.8785738
#>  [7,]  0.76230125  1.55795821  1.3227462
#>  [8,]  0.44560194  1.08564328  0.3928975
#>  [9,]  0.13838995  0.13707904  0.8802446
#> [10,] -0.46347664  0.47264776  0.4254564
#> [11,] -0.04192427 -1.60368661 -0.3381054
#> [12,]  0.23126611  1.01782745  0.6835632
#> [13,] -0.19218543 -0.94927441  1.1083145
#> 
#> attr(,"class")
#> [1] "XYZ"             "MULTILINESTRING" "sfg"            
#> 
#> attr(,"n_empty")
#> [1] 0
#> attr(,"crs")
#> $input
#> [1] NA
#> 
#> $wkt
#> [1] NA
#> 
#> attr(,"class")
#> [1] "crs"
#> attr(,"class")
#> [1] "sfc_MULTILINESTRING" "sfc"                
#> attr(,"precision")
#> [1] 0
#> attr(,"bbox")
#>       xmin       ymin       xmax       ymax 
#> -1.1605384 -1.6036866  0.7623013  1.5579582 
#> attr(,"class")
#> [1] "bbox"
#> attr(,"z_range")
#>       zmin       zmax 
#> -0.8785738  1.3227462 
#> attr(,"class")
#> [1] "z_range"
sfc_multilinestring( obj = df, x = "x", y = "y", z = "z", m = "m")
#> [[1]]
#> [[1]]
#>              [,1]        [,2]       [,3]        [,4]
#>  [1,] -0.72477040  0.73419583 -0.1572917  0.38689140
#>  [2,] -0.35745209  0.65115182 -0.6446505  2.05053551
#>  [3,]  0.48089085 -0.61471396  1.0675388 -1.08432784
#>  [4,] -1.16053843 -0.86069844  1.2046122 -0.39754138
#>  [5,]  0.63335358 -0.01859121 -0.8560544 -0.95495375
#>  [6,]  0.69714383  0.34643868 -0.8785738 -0.81342668
#>  [7,]  0.76230125  1.55795821  1.3227462  0.02926002
#>  [8,]  0.44560194  1.08564328  0.3928975  2.31431295
#>  [9,]  0.13838995  0.13707904  0.8802446  0.58178274
#> [10,] -0.46347664  0.47264776  0.4254564  1.44410324
#> [11,] -0.04192427 -1.60368661 -0.3381054  0.60879732
#> [12,]  0.23126611  1.01782745  0.6835632 -1.29623695
#> [13,] -0.19218543 -0.94927441  1.1083145  0.53945763
#> 
#> attr(,"class")
#> [1] "XYZM"            "MULTILINESTRING" "sfg"            
#> 
#> attr(,"n_empty")
#> [1] 0
#> attr(,"crs")
#> $input
#> [1] NA
#> 
#> $wkt
#> [1] NA
#> 
#> attr(,"class")
#> [1] "crs"
#> attr(,"class")
#> [1] "sfc_MULTILINESTRING" "sfc"                
#> attr(,"precision")
#> [1] 0
#> attr(,"bbox")
#>       xmin       ymin       xmax       ymax 
#> -1.1605384 -1.6036866  0.7623013  1.5579582 
#> attr(,"class")
#> [1] "bbox"
#> attr(,"z_range")
#>       zmin       zmax 
#> -0.8785738  1.3227462 
#> attr(,"class")
#> [1] "z_range"
#> attr(,"m_range")
#>      mmin      mmax 
#> -1.296237  2.314313 
#> attr(,"class")
#> [1] "m_range"

sfc_multilinestring( obj = df, x = 2, y = 3)
#> [[1]]
#> [[1]]
#>       [,1]        [,2]
#>  [1,]    1 -0.72477040
#>  [2,]    1 -0.35745209
#>  [3,]    1  0.48089085
#>  [4,]    2 -1.16053843
#>  [5,]    2  0.63335358
#>  [6,]    3  0.69714383
#>  [7,]    3  0.76230125
#>  [8,]    3  0.44560194
#>  [9,]    1  0.13838995
#> [10,]    1 -0.46347664
#> [11,]    1 -0.04192427
#> [12,]    2  0.23126611
#> [13,]    2 -0.19218543
#> 
#> attr(,"class")
#> [1] "XY"              "MULTILINESTRING" "sfg"            
#> 
#> attr(,"n_empty")
#> [1] 0
#> attr(,"crs")
#> $input
#> [1] NA
#> 
#> $wkt
#> [1] NA
#> 
#> attr(,"class")
#> [1] "crs"
#> attr(,"class")
#> [1] "sfc_MULTILINESTRING" "sfc"                
#> attr(,"precision")
#> [1] 0
#> attr(,"bbox")
#>       xmin       ymin       xmax       ymax 
#>  1.0000000 -1.1605384  3.0000000  0.7623013 
#> attr(,"class")
#> [1] "bbox"
sfc_multilinestring( obj = df, x = 2, y = 3, z = 4)
#> [[1]]
#> [[1]]
#>       [,1]        [,2]        [,3]
#>  [1,]    1 -0.72477040  0.73419583
#>  [2,]    1 -0.35745209  0.65115182
#>  [3,]    1  0.48089085 -0.61471396
#>  [4,]    2 -1.16053843 -0.86069844
#>  [5,]    2  0.63335358 -0.01859121
#>  [6,]    3  0.69714383  0.34643868
#>  [7,]    3  0.76230125  1.55795821
#>  [8,]    3  0.44560194  1.08564328
#>  [9,]    1  0.13838995  0.13707904
#> [10,]    1 -0.46347664  0.47264776
#> [11,]    1 -0.04192427 -1.60368661
#> [12,]    2  0.23126611  1.01782745
#> [13,]    2 -0.19218543 -0.94927441
#> 
#> attr(,"class")
#> [1] "XYZ"             "MULTILINESTRING" "sfg"            
#> 
#> attr(,"n_empty")
#> [1] 0
#> attr(,"crs")
#> $input
#> [1] NA
#> 
#> $wkt
#> [1] NA
#> 
#> attr(,"class")
#> [1] "crs"
#> attr(,"class")
#> [1] "sfc_MULTILINESTRING" "sfc"                
#> attr(,"precision")
#> [1] 0
#> attr(,"bbox")
#>       xmin       ymin       xmax       ymax 
#>  1.0000000 -1.1605384  3.0000000  0.7623013 
#> attr(,"class")
#> [1] "bbox"
#> attr(,"z_range")
#>      zmin      zmax 
#> -1.603687  1.557958 
#> attr(,"class")
#> [1] "z_range"
sfc_multilinestring( obj = df, x = 2, y = 3, z = 4, m = 5)
#> [[1]]
#> [[1]]
#>       [,1]        [,2]        [,3]       [,4]
#>  [1,]    1 -0.72477040  0.73419583 -0.1572917
#>  [2,]    1 -0.35745209  0.65115182 -0.6446505
#>  [3,]    1  0.48089085 -0.61471396  1.0675388
#>  [4,]    2 -1.16053843 -0.86069844  1.2046122
#>  [5,]    2  0.63335358 -0.01859121 -0.8560544
#>  [6,]    3  0.69714383  0.34643868 -0.8785738
#>  [7,]    3  0.76230125  1.55795821  1.3227462
#>  [8,]    3  0.44560194  1.08564328  0.3928975
#>  [9,]    1  0.13838995  0.13707904  0.8802446
#> [10,]    1 -0.46347664  0.47264776  0.4254564
#> [11,]    1 -0.04192427 -1.60368661 -0.3381054
#> [12,]    2  0.23126611  1.01782745  0.6835632
#> [13,]    2 -0.19218543 -0.94927441  1.1083145
#> 
#> attr(,"class")
#> [1] "XYZM"            "MULTILINESTRING" "sfg"            
#> 
#> attr(,"n_empty")
#> [1] 0
#> attr(,"crs")
#> $input
#> [1] NA
#> 
#> $wkt
#> [1] NA
#> 
#> attr(,"class")
#> [1] "crs"
#> attr(,"class")
#> [1] "sfc_MULTILINESTRING" "sfc"                
#> attr(,"precision")
#> [1] 0
#> attr(,"bbox")
#>       xmin       ymin       xmax       ymax 
#>  1.0000000 -1.1605384  3.0000000  0.7623013 
#> attr(,"class")
#> [1] "bbox"
#> attr(,"z_range")
#>      zmin      zmax 
#> -1.603687  1.557958 
#> attr(,"class")
#> [1] "z_range"
#> attr(,"m_range")
#>       mmin       mmax 
#> -0.8785738  1.3227462 
#> attr(,"class")
#> [1] "m_range"

sfc_multilinestring( obj = df, multilinestring_id = "ml_id", linestring_id = "l_id" )
#> [[1]]
#> [[1]]
#>            [,1]       [,2]       [,3]       [,4]
#> [1,] -0.7247704  0.7341958 -0.1572917  0.3868914
#> [2,] -0.3574521  0.6511518 -0.6446505  2.0505355
#> [3,]  0.4808909 -0.6147140  1.0675388 -1.0843278
#> 
#> [[2]]
#>            [,1]        [,2]       [,3]       [,4]
#> [1,] -1.1605384 -0.86069844  1.2046122 -0.3975414
#> [2,]  0.6333536 -0.01859121 -0.8560544 -0.9549538
#> 
#> [[3]]
#>           [,1]      [,2]       [,3]        [,4]
#> [1,] 0.6971438 0.3464387 -0.8785738 -0.81342668
#> [2,] 0.7623013 1.5579582  1.3227462  0.02926002
#> [3,] 0.4456019 1.0856433  0.3928975  2.31431295
#> 
#> attr(,"class")
#> [1] "XYZM"            "MULTILINESTRING" "sfg"            
#> 
#> [[2]]
#> [[1]]
#>             [,1]       [,2]       [,3]      [,4]
#> [1,]  0.13838995  0.1370790  0.8802446 0.5817827
#> [2,] -0.46347664  0.4726478  0.4254564 1.4441032
#> [3,] -0.04192427 -1.6036866 -0.3381054 0.6087973
#> 
#> [[2]]
#>            [,1]       [,2]      [,3]       [,4]
#> [1,]  0.2312661  1.0178275 0.6835632 -1.2962370
#> [2,] -0.1921854 -0.9492744 1.1083145  0.5394576
#> 
#> attr(,"class")
#> [1] "XYZM"            "MULTILINESTRING" "sfg"            
#> 
#> attr(,"n_empty")
#> [1] 0
#> attr(,"crs")
#> $input
#> [1] NA
#> 
#> $wkt
#> [1] NA
#> 
#> attr(,"class")
#> [1] "crs"
#> attr(,"class")
#> [1] "sfc_MULTILINESTRING" "sfc"                
#> attr(,"precision")
#> [1] 0
#> attr(,"bbox")
#>       xmin       ymin       xmax       ymax 
#> -1.1605384 -1.6036866  0.7623013  1.5579582 
#> attr(,"class")
#> [1] "bbox"
#> attr(,"z_range")
#>       zmin       zmax 
#> -0.8785738  1.3227462 
#> attr(,"class")
#> [1] "z_range"
#> attr(,"m_range")
#>      mmin      mmax 
#> -1.296237  2.314313 
#> attr(,"class")
#> [1] "m_range"
sfc_multilinestring( obj = df, multilinestring_id = 1, linestring_id = 2 )
#> [[1]]
#> [[1]]
#>            [,1]       [,2]       [,3]       [,4]
#> [1,] -0.7247704  0.7341958 -0.1572917  0.3868914
#> [2,] -0.3574521  0.6511518 -0.6446505  2.0505355
#> [3,]  0.4808909 -0.6147140  1.0675388 -1.0843278
#> 
#> [[2]]
#>            [,1]        [,2]       [,3]       [,4]
#> [1,] -1.1605384 -0.86069844  1.2046122 -0.3975414
#> [2,]  0.6333536 -0.01859121 -0.8560544 -0.9549538
#> 
#> [[3]]
#>           [,1]      [,2]       [,3]        [,4]
#> [1,] 0.6971438 0.3464387 -0.8785738 -0.81342668
#> [2,] 0.7623013 1.5579582  1.3227462  0.02926002
#> [3,] 0.4456019 1.0856433  0.3928975  2.31431295
#> 
#> attr(,"class")
#> [1] "XYZM"            "MULTILINESTRING" "sfg"            
#> 
#> [[2]]
#> [[1]]
#>             [,1]       [,2]       [,3]      [,4]
#> [1,]  0.13838995  0.1370790  0.8802446 0.5817827
#> [2,] -0.46347664  0.4726478  0.4254564 1.4441032
#> [3,] -0.04192427 -1.6036866 -0.3381054 0.6087973
#> 
#> [[2]]
#>            [,1]       [,2]      [,3]       [,4]
#> [1,]  0.2312661  1.0178275 0.6835632 -1.2962370
#> [2,] -0.1921854 -0.9492744 1.1083145  0.5394576
#> 
#> attr(,"class")
#> [1] "XYZM"            "MULTILINESTRING" "sfg"            
#> 
#> attr(,"n_empty")
#> [1] 0
#> attr(,"crs")
#> $input
#> [1] NA
#> 
#> $wkt
#> [1] NA
#> 
#> attr(,"class")
#> [1] "crs"
#> attr(,"class")
#> [1] "sfc_MULTILINESTRING" "sfc"                
#> attr(,"precision")
#> [1] 0
#> attr(,"bbox")
#>       xmin       ymin       xmax       ymax 
#> -1.1605384 -1.6036866  0.7623013  1.5579582 
#> attr(,"class")
#> [1] "bbox"
#> attr(,"z_range")
#>       zmin       zmax 
#> -0.8785738  1.3227462 
#> attr(,"class")
#> [1] "z_range"
#> attr(,"m_range")
#>      mmin      mmax 
#> -1.296237  2.314313 
#> attr(,"class")
#> [1] "m_range"