sfc MULTILINESTRING

constructs an sfc of MULTILINESTRING objects

sfc_multilinestring(
  obj = NULL,
  x = NULL,
  y = NULL,
  z = NULL,
  m = NULL,
  multilinestring_id = NULL,
  linestring_id = NULL
)

Arguments

obj

sorted matrix or data.frame

x

x geometry column

y

y geometry column

z

z geometry column

m

m geometry column

multilinestring_id

column of ids for multilinestrings

linestring_id

column of ids for linestrings (within multilinestrings)

Value

sfc object of MULTILINESTRING geometries

notes

sfheaders functions do not perform any validity checks on the geometries. Nor do they set Coordinate Reference Systems, EPSG, PROJ4 or precision attributes.

The data.frame and matrices you send into the sfheader functions must be ordered.

Examples

m <- matrix(c(0,0,0,0,1,1), ncol = 3 )
sfc_multilinestring( m )
#> [[1]]
#> [[1]]
#>      [,1] [,2] [,3]
#> [1,]    0    0    1
#> [2,]    0    0    1
#> 
#> attr(,"class")
#> [1] "XYZ"             "MULTILINESTRING" "sfg"            
#> 
#> attr(,"n_empty")
#> [1] 0
#> attr(,"crs")
#> $input
#> [1] NA
#> 
#> $wkt
#> [1] NA
#> 
#> attr(,"class")
#> [1] "crs"
#> attr(,"class")
#> [1] "sfc_MULTILINESTRING" "sfc"                
#> attr(,"precision")
#> [1] 0
#> attr(,"bbox")
#> xmin ymin xmax ymax 
#>    0    0    0    0 
#> attr(,"class")
#> [1] "bbox"
#> attr(,"z_range")
#> zmin zmax 
#>    1    1 
#> attr(,"class")
#> [1] "z_range"

m <- matrix(c(0,0,0,0,0,1,0,1,1,1,2,2,1,2,3), ncol = 3, byrow = TRUE)
sfc_multilinestring( obj = m )
#> [[1]]
#> [[1]]
#>      [,1] [,2] [,3]
#> [1,]    0    0    0
#> [2,]    0    0    1
#> [3,]    0    1    1
#> [4,]    1    2    2
#> [5,]    1    2    3
#> 
#> attr(,"class")
#> [1] "XYZ"             "MULTILINESTRING" "sfg"            
#> 
#> attr(,"n_empty")
#> [1] 0
#> attr(,"crs")
#> $input
#> [1] NA
#> 
#> $wkt
#> [1] NA
#> 
#> attr(,"class")
#> [1] "crs"
#> attr(,"class")
#> [1] "sfc_MULTILINESTRING" "sfc"                
#> attr(,"precision")
#> [1] 0
#> attr(,"bbox")
#> xmin ymin xmax ymax 
#>    0    0    1    2 
#> attr(,"class")
#> [1] "bbox"
#> attr(,"z_range")
#> zmin zmax 
#>    0    3 
#> attr(,"class")
#> [1] "z_range"
sfc_multilinestring( obj = m, multilinestring_id = 1 )
#> [[1]]
#> [[1]]
#>      [,1] [,2]
#> [1,]    0    0
#> [2,]    0    1
#> [3,]    1    1
#> 
#> attr(,"class")
#> [1] "XY"              "MULTILINESTRING" "sfg"            
#> 
#> [[2]]
#> [[1]]
#>      [,1] [,2]
#> [1,]    2    2
#> [2,]    2    3
#> 
#> attr(,"class")
#> [1] "XY"              "MULTILINESTRING" "sfg"            
#> 
#> attr(,"n_empty")
#> [1] 0
#> attr(,"crs")
#> $input
#> [1] NA
#> 
#> $wkt
#> [1] NA
#> 
#> attr(,"class")
#> [1] "crs"
#> attr(,"class")
#> [1] "sfc_MULTILINESTRING" "sfc"                
#> attr(,"precision")
#> [1] 0
#> attr(,"bbox")
#> xmin ymin xmax ymax 
#>    0    0    2    3 
#> attr(,"class")
#> [1] "bbox"
sfc_multilinestring( obj = m, linestring_id = 1 )
#> [[1]]
#> [[1]]
#>      [,1] [,2]
#> [1,]    0    0
#> [2,]    0    1
#> [3,]    1    1
#> 
#> [[2]]
#>      [,1] [,2]
#> [1,]    2    2
#> [2,]    2    3
#> 
#> attr(,"class")
#> [1] "XY"              "MULTILINESTRING" "sfg"            
#> 
#> attr(,"n_empty")
#> [1] 0
#> attr(,"crs")
#> $input
#> [1] NA
#> 
#> $wkt
#> [1] NA
#> 
#> attr(,"class")
#> [1] "crs"
#> attr(,"class")
#> [1] "sfc_MULTILINESTRING" "sfc"                
#> attr(,"precision")
#> [1] 0
#> attr(,"bbox")
#> xmin ymin xmax ymax 
#>    0    0    2    3 
#> attr(,"class")
#> [1] "bbox"

sfc_multilinestring( obj = m, linestring_id = 1, multilinestring_id = 1 )
#> [[1]]
#> [[1]]
#>      [,1] [,2]
#> [1,]    0    0
#> [2,]    0    1
#> [3,]    1    1
#> 
#> attr(,"class")
#> [1] "XY"              "MULTILINESTRING" "sfg"            
#> 
#> [[2]]
#> [[1]]
#>      [,1] [,2]
#> [1,]    2    2
#> [2,]    2    3
#> 
#> attr(,"class")
#> [1] "XY"              "MULTILINESTRING" "sfg"            
#> 
#> attr(,"n_empty")
#> [1] 0
#> attr(,"crs")
#> $input
#> [1] NA
#> 
#> $wkt
#> [1] NA
#> 
#> attr(,"class")
#> [1] "crs"
#> attr(,"class")
#> [1] "sfc_MULTILINESTRING" "sfc"                
#> attr(,"precision")
#> [1] 0
#> attr(,"bbox")
#> xmin ymin xmax ymax 
#>    0    0    2    3 
#> attr(,"class")
#> [1] "bbox"

sfc_multilinestring( obj = m, x = 2, y = 3 )
#> [[1]]
#> [[1]]
#>      [,1] [,2]
#> [1,]    0    0
#> [2,]    0    1
#> [3,]    1    1
#> [4,]    2    2
#> [5,]    2    3
#> 
#> attr(,"class")
#> [1] "XY"              "MULTILINESTRING" "sfg"            
#> 
#> attr(,"n_empty")
#> [1] 0
#> attr(,"crs")
#> $input
#> [1] NA
#> 
#> $wkt
#> [1] NA
#> 
#> attr(,"class")
#> [1] "crs"
#> attr(,"class")
#> [1] "sfc_MULTILINESTRING" "sfc"                
#> attr(,"precision")
#> [1] 0
#> attr(,"bbox")
#> xmin ymin xmax ymax 
#>    0    0    2    3 
#> attr(,"class")
#> [1] "bbox"
sfc_multilinestring( obj = m, x = 1, y = 2, z = 3 )
#> [[1]]
#> [[1]]
#>      [,1] [,2] [,3]
#> [1,]    0    0    0
#> [2,]    0    0    1
#> [3,]    0    1    1
#> [4,]    1    2    2
#> [5,]    1    2    3
#> 
#> attr(,"class")
#> [1] "XYZ"             "MULTILINESTRING" "sfg"            
#> 
#> attr(,"n_empty")
#> [1] 0
#> attr(,"crs")
#> $input
#> [1] NA
#> 
#> $wkt
#> [1] NA
#> 
#> attr(,"class")
#> [1] "crs"
#> attr(,"class")
#> [1] "sfc_MULTILINESTRING" "sfc"                
#> attr(,"precision")
#> [1] 0
#> attr(,"bbox")
#> xmin ymin xmax ymax 
#>    0    0    1    2 
#> attr(,"class")
#> [1] "bbox"
#> attr(,"z_range")
#> zmin zmax 
#>    0    3 
#> attr(,"class")
#> [1] "z_range"
sfc_multilinestring( obj = m, x = 2, y = 3, linestring_id = 1, multilinestring_id = 1 )
#> [[1]]
#> [[1]]
#>      [,1] [,2]
#> [1,]    0    0
#> [2,]    0    1
#> [3,]    1    1
#> 
#> attr(,"class")
#> [1] "XY"              "MULTILINESTRING" "sfg"            
#> 
#> [[2]]
#> [[1]]
#>      [,1] [,2]
#> [1,]    2    2
#> [2,]    2    3
#> 
#> attr(,"class")
#> [1] "XY"              "MULTILINESTRING" "sfg"            
#> 
#> attr(,"n_empty")
#> [1] 0
#> attr(,"crs")
#> $input
#> [1] NA
#> 
#> $wkt
#> [1] NA
#> 
#> attr(,"class")
#> [1] "crs"
#> attr(,"class")
#> [1] "sfc_MULTILINESTRING" "sfc"                
#> attr(,"precision")
#> [1] 0
#> attr(,"bbox")
#> xmin ymin xmax ymax 
#>    0    0    2    3 
#> attr(,"class")
#> [1] "bbox"

df <- data.frame(
  ml_id = c(1,1,1,1,1,1,1,1,2,2,2,2,2)
  , l_id = c(1,1,1,2,2,3,3,3,1,1,1,2,2)
  , x = rnorm(13)
  , y = rnorm(13)
  , z = rnorm(13)
  , m = rnorm(13)
)

sfc_multilinestring( obj = df, x = "x", y = "y")
#> [[1]]
#> [[1]]
#>             [,1]        [,2]
#>  [1,]  1.5565905  0.37874013
#>  [2,]  0.7273139 -0.46758878
#>  [3,] -0.7345878 -0.40129287
#>  [4,]  0.2989703 -0.08646814
#>  [5,] -0.2816043 -0.33930114
#>  [6,]  0.6925730 -0.26515810
#>  [7,]  1.0841661 -0.55260430
#>  [8,]  1.5133805 -0.97862347
#>  [9,]  0.2279996 -0.32844705
#> [10,] -0.8608110  0.64227222
#> [11,] -1.4473680 -0.01022913
#> [12,]  0.0493561 -1.27802330
#> [13,] -0.5136593 -0.64248246
#> 
#> attr(,"class")
#> [1] "XY"              "MULTILINESTRING" "sfg"            
#> 
#> attr(,"n_empty")
#> [1] 0
#> attr(,"crs")
#> $input
#> [1] NA
#> 
#> $wkt
#> [1] NA
#> 
#> attr(,"class")
#> [1] "crs"
#> attr(,"class")
#> [1] "sfc_MULTILINESTRING" "sfc"                
#> attr(,"precision")
#> [1] 0
#> attr(,"bbox")
#>       xmin       ymin       xmax       ymax 
#> -1.4473680 -1.2780233  1.5565905  0.6422722 
#> attr(,"class")
#> [1] "bbox"
sfc_multilinestring( obj = df, x = "x", y = "y", z = "z")
#> [[1]]
#> [[1]]
#>             [,1]        [,2]       [,3]
#>  [1,]  1.5565905  0.37874013 -0.2168756
#>  [2,]  0.7273139 -0.46758878  0.9272685
#>  [3,] -0.7345878 -0.40129287 -0.6469326
#>  [4,]  0.2989703 -0.08646814  1.1259606
#>  [5,] -0.2816043 -0.33930114  0.2231915
#>  [6,]  0.6925730 -0.26515810  0.7382979
#>  [7,]  1.0841661 -0.55260430 -0.9838014
#>  [8,]  1.5133805 -0.97862347  0.4718368
#>  [9,]  0.2279996 -0.32844705  0.6001309
#> [10,] -0.8608110  0.64227222 -1.3206307
#> [11,] -1.4473680 -0.01022913  0.1497284
#> [12,]  0.0493561 -1.27802330 -1.2417915
#> [13,] -0.5136593 -0.64248246 -2.1119756
#> 
#> attr(,"class")
#> [1] "XYZ"             "MULTILINESTRING" "sfg"            
#> 
#> attr(,"n_empty")
#> [1] 0
#> attr(,"crs")
#> $input
#> [1] NA
#> 
#> $wkt
#> [1] NA
#> 
#> attr(,"class")
#> [1] "crs"
#> attr(,"class")
#> [1] "sfc_MULTILINESTRING" "sfc"                
#> attr(,"precision")
#> [1] 0
#> attr(,"bbox")
#>       xmin       ymin       xmax       ymax 
#> -1.4473680 -1.2780233  1.5565905  0.6422722 
#> attr(,"class")
#> [1] "bbox"
#> attr(,"z_range")
#>      zmin      zmax 
#> -2.111976  1.125961 
#> attr(,"class")
#> [1] "z_range"
sfc_multilinestring( obj = df, x = "x", y = "y", z = "z", m = "m")
#> [[1]]
#> [[1]]
#>             [,1]        [,2]       [,3]       [,4]
#>  [1,]  1.5565905  0.37874013 -0.2168756 -1.1283149
#>  [2,]  0.7273139 -0.46758878  0.9272685  2.4594926
#>  [3,] -0.7345878 -0.40129287 -0.6469326  0.1363568
#>  [4,]  0.2989703 -0.08646814  1.1259606 -0.8121986
#>  [5,] -0.2816043 -0.33930114  0.2231915  0.7029741
#>  [6,]  0.6925730 -0.26515810  0.7382979 -2.1926348
#>  [7,]  1.0841661 -0.55260430 -0.9838014 -1.3183680
#>  [8,]  1.5133805 -0.97862347  0.4718368 -0.8808006
#>  [9,]  0.2279996 -0.32844705  0.6001309  1.3050548
#> [10,] -0.8608110  0.64227222 -1.3206307 -1.0624314
#> [11,] -1.4473680 -0.01022913  0.1497284 -0.8600658
#> [12,]  0.0493561 -1.27802330 -1.2417915  0.5393462
#> [13,] -0.5136593 -0.64248246 -2.1119756 -0.5249988
#> 
#> attr(,"class")
#> [1] "XYZM"            "MULTILINESTRING" "sfg"            
#> 
#> attr(,"n_empty")
#> [1] 0
#> attr(,"crs")
#> $input
#> [1] NA
#> 
#> $wkt
#> [1] NA
#> 
#> attr(,"class")
#> [1] "crs"
#> attr(,"class")
#> [1] "sfc_MULTILINESTRING" "sfc"                
#> attr(,"precision")
#> [1] 0
#> attr(,"bbox")
#>       xmin       ymin       xmax       ymax 
#> -1.4473680 -1.2780233  1.5565905  0.6422722 
#> attr(,"class")
#> [1] "bbox"
#> attr(,"z_range")
#>      zmin      zmax 
#> -2.111976  1.125961 
#> attr(,"class")
#> [1] "z_range"
#> attr(,"m_range")
#>      mmin      mmax 
#> -2.192635  2.459493 
#> attr(,"class")
#> [1] "m_range"

sfc_multilinestring( obj = df, x = 2, y = 3)
#> [[1]]
#> [[1]]
#>       [,1]       [,2]
#>  [1,]    1  1.5565905
#>  [2,]    1  0.7273139
#>  [3,]    1 -0.7345878
#>  [4,]    2  0.2989703
#>  [5,]    2 -0.2816043
#>  [6,]    3  0.6925730
#>  [7,]    3  1.0841661
#>  [8,]    3  1.5133805
#>  [9,]    1  0.2279996
#> [10,]    1 -0.8608110
#> [11,]    1 -1.4473680
#> [12,]    2  0.0493561
#> [13,]    2 -0.5136593
#> 
#> attr(,"class")
#> [1] "XY"              "MULTILINESTRING" "sfg"            
#> 
#> attr(,"n_empty")
#> [1] 0
#> attr(,"crs")
#> $input
#> [1] NA
#> 
#> $wkt
#> [1] NA
#> 
#> attr(,"class")
#> [1] "crs"
#> attr(,"class")
#> [1] "sfc_MULTILINESTRING" "sfc"                
#> attr(,"precision")
#> [1] 0
#> attr(,"bbox")
#>      xmin      ymin      xmax      ymax 
#>  1.000000 -1.447368  3.000000  1.556590 
#> attr(,"class")
#> [1] "bbox"
sfc_multilinestring( obj = df, x = 2, y = 3, z = 4)
#> [[1]]
#> [[1]]
#>       [,1]       [,2]        [,3]
#>  [1,]    1  1.5565905  0.37874013
#>  [2,]    1  0.7273139 -0.46758878
#>  [3,]    1 -0.7345878 -0.40129287
#>  [4,]    2  0.2989703 -0.08646814
#>  [5,]    2 -0.2816043 -0.33930114
#>  [6,]    3  0.6925730 -0.26515810
#>  [7,]    3  1.0841661 -0.55260430
#>  [8,]    3  1.5133805 -0.97862347
#>  [9,]    1  0.2279996 -0.32844705
#> [10,]    1 -0.8608110  0.64227222
#> [11,]    1 -1.4473680 -0.01022913
#> [12,]    2  0.0493561 -1.27802330
#> [13,]    2 -0.5136593 -0.64248246
#> 
#> attr(,"class")
#> [1] "XYZ"             "MULTILINESTRING" "sfg"            
#> 
#> attr(,"n_empty")
#> [1] 0
#> attr(,"crs")
#> $input
#> [1] NA
#> 
#> $wkt
#> [1] NA
#> 
#> attr(,"class")
#> [1] "crs"
#> attr(,"class")
#> [1] "sfc_MULTILINESTRING" "sfc"                
#> attr(,"precision")
#> [1] 0
#> attr(,"bbox")
#>      xmin      ymin      xmax      ymax 
#>  1.000000 -1.447368  3.000000  1.556590 
#> attr(,"class")
#> [1] "bbox"
#> attr(,"z_range")
#>       zmin       zmax 
#> -1.2780233  0.6422722 
#> attr(,"class")
#> [1] "z_range"
sfc_multilinestring( obj = df, x = 2, y = 3, z = 4, m = 5)
#> [[1]]
#> [[1]]
#>       [,1]       [,2]        [,3]       [,4]
#>  [1,]    1  1.5565905  0.37874013 -0.2168756
#>  [2,]    1  0.7273139 -0.46758878  0.9272685
#>  [3,]    1 -0.7345878 -0.40129287 -0.6469326
#>  [4,]    2  0.2989703 -0.08646814  1.1259606
#>  [5,]    2 -0.2816043 -0.33930114  0.2231915
#>  [6,]    3  0.6925730 -0.26515810  0.7382979
#>  [7,]    3  1.0841661 -0.55260430 -0.9838014
#>  [8,]    3  1.5133805 -0.97862347  0.4718368
#>  [9,]    1  0.2279996 -0.32844705  0.6001309
#> [10,]    1 -0.8608110  0.64227222 -1.3206307
#> [11,]    1 -1.4473680 -0.01022913  0.1497284
#> [12,]    2  0.0493561 -1.27802330 -1.2417915
#> [13,]    2 -0.5136593 -0.64248246 -2.1119756
#> 
#> attr(,"class")
#> [1] "XYZM"            "MULTILINESTRING" "sfg"            
#> 
#> attr(,"n_empty")
#> [1] 0
#> attr(,"crs")
#> $input
#> [1] NA
#> 
#> $wkt
#> [1] NA
#> 
#> attr(,"class")
#> [1] "crs"
#> attr(,"class")
#> [1] "sfc_MULTILINESTRING" "sfc"                
#> attr(,"precision")
#> [1] 0
#> attr(,"bbox")
#>      xmin      ymin      xmax      ymax 
#>  1.000000 -1.447368  3.000000  1.556590 
#> attr(,"class")
#> [1] "bbox"
#> attr(,"z_range")
#>       zmin       zmax 
#> -1.2780233  0.6422722 
#> attr(,"class")
#> [1] "z_range"
#> attr(,"m_range")
#>      mmin      mmax 
#> -2.111976  1.125961 
#> attr(,"class")
#> [1] "m_range"

sfc_multilinestring( obj = df, multilinestring_id = "ml_id", linestring_id = "l_id" )
#> [[1]]
#> [[1]]
#>            [,1]       [,2]       [,3]       [,4]
#> [1,]  1.5565905  0.3787401 -0.2168756 -1.1283149
#> [2,]  0.7273139 -0.4675888  0.9272685  2.4594926
#> [3,] -0.7345878 -0.4012929 -0.6469326  0.1363568
#> 
#> [[2]]
#>            [,1]        [,2]      [,3]       [,4]
#> [1,]  0.2989703 -0.08646814 1.1259606 -0.8121986
#> [2,] -0.2816043 -0.33930114 0.2231915  0.7029741
#> 
#> [[3]]
#>          [,1]       [,2]       [,3]       [,4]
#> [1,] 0.692573 -0.2651581  0.7382979 -2.1926348
#> [2,] 1.084166 -0.5526043 -0.9838014 -1.3183680
#> [3,] 1.513380 -0.9786235  0.4718368 -0.8808006
#> 
#> attr(,"class")
#> [1] "XYZM"            "MULTILINESTRING" "sfg"            
#> 
#> [[2]]
#> [[1]]
#>            [,1]        [,2]       [,3]       [,4]
#> [1,]  0.2279996 -0.32844705  0.6001309  1.3050548
#> [2,] -0.8608110  0.64227222 -1.3206307 -1.0624314
#> [3,] -1.4473680 -0.01022913  0.1497284 -0.8600658
#> 
#> [[2]]
#>            [,1]       [,2]      [,3]       [,4]
#> [1,]  0.0493561 -1.2780233 -1.241791  0.5393462
#> [2,] -0.5136593 -0.6424825 -2.111976 -0.5249988
#> 
#> attr(,"class")
#> [1] "XYZM"            "MULTILINESTRING" "sfg"            
#> 
#> attr(,"n_empty")
#> [1] 0
#> attr(,"crs")
#> $input
#> [1] NA
#> 
#> $wkt
#> [1] NA
#> 
#> attr(,"class")
#> [1] "crs"
#> attr(,"class")
#> [1] "sfc_MULTILINESTRING" "sfc"                
#> attr(,"precision")
#> [1] 0
#> attr(,"bbox")
#>       xmin       ymin       xmax       ymax 
#> -1.4473680 -1.2780233  1.5565905  0.6422722 
#> attr(,"class")
#> [1] "bbox"
#> attr(,"z_range")
#>      zmin      zmax 
#> -2.111976  1.125961 
#> attr(,"class")
#> [1] "z_range"
#> attr(,"m_range")
#>      mmin      mmax 
#> -2.192635  2.459493 
#> attr(,"class")
#> [1] "m_range"
sfc_multilinestring( obj = df, multilinestring_id = 1, linestring_id = 2 )
#> [[1]]
#> [[1]]
#>            [,1]       [,2]       [,3]       [,4]
#> [1,]  1.5565905  0.3787401 -0.2168756 -1.1283149
#> [2,]  0.7273139 -0.4675888  0.9272685  2.4594926
#> [3,] -0.7345878 -0.4012929 -0.6469326  0.1363568
#> 
#> [[2]]
#>            [,1]        [,2]      [,3]       [,4]
#> [1,]  0.2989703 -0.08646814 1.1259606 -0.8121986
#> [2,] -0.2816043 -0.33930114 0.2231915  0.7029741
#> 
#> [[3]]
#>          [,1]       [,2]       [,3]       [,4]
#> [1,] 0.692573 -0.2651581  0.7382979 -2.1926348
#> [2,] 1.084166 -0.5526043 -0.9838014 -1.3183680
#> [3,] 1.513380 -0.9786235  0.4718368 -0.8808006
#> 
#> attr(,"class")
#> [1] "XYZM"            "MULTILINESTRING" "sfg"            
#> 
#> [[2]]
#> [[1]]
#>            [,1]        [,2]       [,3]       [,4]
#> [1,]  0.2279996 -0.32844705  0.6001309  1.3050548
#> [2,] -0.8608110  0.64227222 -1.3206307 -1.0624314
#> [3,] -1.4473680 -0.01022913  0.1497284 -0.8600658
#> 
#> [[2]]
#>            [,1]       [,2]      [,3]       [,4]
#> [1,]  0.0493561 -1.2780233 -1.241791  0.5393462
#> [2,] -0.5136593 -0.6424825 -2.111976 -0.5249988
#> 
#> attr(,"class")
#> [1] "XYZM"            "MULTILINESTRING" "sfg"            
#> 
#> attr(,"n_empty")
#> [1] 0
#> attr(,"crs")
#> $input
#> [1] NA
#> 
#> $wkt
#> [1] NA
#> 
#> attr(,"class")
#> [1] "crs"
#> attr(,"class")
#> [1] "sfc_MULTILINESTRING" "sfc"                
#> attr(,"precision")
#> [1] 0
#> attr(,"bbox")
#>       xmin       ymin       xmax       ymax 
#> -1.4473680 -1.2780233  1.5565905  0.6422722 
#> attr(,"class")
#> [1] "bbox"
#> attr(,"z_range")
#>      zmin      zmax 
#> -2.111976  1.125961 
#> attr(,"class")
#> [1] "z_range"
#> attr(,"m_range")
#>      mmin      mmax 
#> -2.192635  2.459493 
#> attr(,"class")
#> [1] "m_range"