sfc MULTILINESTRING

constructs an sfc of MULTILINESTRING objects

sfc_multilinestring(
  obj = NULL,
  x = NULL,
  y = NULL,
  z = NULL,
  m = NULL,
  multilinestring_id = NULL,
  linestring_id = NULL
)

Arguments

obj

sorted matrix or data.frame

x

x geometry column

y

y geometry column

z

z geometry column

m

m geometry column

multilinestring_id

column of ids for multilinestrings

linestring_id

column of ids for linestrings (within multilinestrings)

Value

sfc object of MULTILINESTRING geometries

notes

sfheaders functions do not perform any validity checks on the geometries. Nor do they set Coordinate Reference Systems, EPSG, PROJ4 or precision attributes.

The data.frame and matrices you send into the sfheader functions must be ordered.

Examples

m <- matrix(c(0,0,0,0,1,1), ncol = 3 )
sfc_multilinestring( m )
#> [[1]]
#> [[1]]
#>      [,1] [,2] [,3]
#> [1,]    0    0    1
#> [2,]    0    0    1
#> 
#> attr(,"class")
#> [1] "XYZ"             "MULTILINESTRING" "sfg"            
#> 
#> attr(,"n_empty")
#> [1] 0
#> attr(,"crs")
#> $input
#> [1] NA
#> 
#> $wkt
#> [1] NA
#> 
#> attr(,"class")
#> [1] "crs"
#> attr(,"class")
#> [1] "sfc_MULTILINESTRING" "sfc"                
#> attr(,"precision")
#> [1] 0
#> attr(,"bbox")
#> xmin ymin xmax ymax 
#>    0    0    0    0 
#> attr(,"class")
#> [1] "bbox"
#> attr(,"z_range")
#> zmin zmax 
#>    1    1 
#> attr(,"class")
#> [1] "z_range"

m <- matrix(c(0,0,0,0,0,1,0,1,1,1,2,2,1,2,3), ncol = 3, byrow = TRUE)
sfc_multilinestring( obj = m )
#> [[1]]
#> [[1]]
#>      [,1] [,2] [,3]
#> [1,]    0    0    0
#> [2,]    0    0    1
#> [3,]    0    1    1
#> [4,]    1    2    2
#> [5,]    1    2    3
#> 
#> attr(,"class")
#> [1] "XYZ"             "MULTILINESTRING" "sfg"            
#> 
#> attr(,"n_empty")
#> [1] 0
#> attr(,"crs")
#> $input
#> [1] NA
#> 
#> $wkt
#> [1] NA
#> 
#> attr(,"class")
#> [1] "crs"
#> attr(,"class")
#> [1] "sfc_MULTILINESTRING" "sfc"                
#> attr(,"precision")
#> [1] 0
#> attr(,"bbox")
#> xmin ymin xmax ymax 
#>    0    0    1    2 
#> attr(,"class")
#> [1] "bbox"
#> attr(,"z_range")
#> zmin zmax 
#>    0    3 
#> attr(,"class")
#> [1] "z_range"
sfc_multilinestring( obj = m, multilinestring_id = 1 )
#> [[1]]
#> [[1]]
#>      [,1] [,2]
#> [1,]    0    0
#> [2,]    0    1
#> [3,]    1    1
#> 
#> attr(,"class")
#> [1] "XY"              "MULTILINESTRING" "sfg"            
#> 
#> [[2]]
#> [[1]]
#>      [,1] [,2]
#> [1,]    2    2
#> [2,]    2    3
#> 
#> attr(,"class")
#> [1] "XY"              "MULTILINESTRING" "sfg"            
#> 
#> attr(,"n_empty")
#> [1] 0
#> attr(,"crs")
#> $input
#> [1] NA
#> 
#> $wkt
#> [1] NA
#> 
#> attr(,"class")
#> [1] "crs"
#> attr(,"class")
#> [1] "sfc_MULTILINESTRING" "sfc"                
#> attr(,"precision")
#> [1] 0
#> attr(,"bbox")
#> xmin ymin xmax ymax 
#>    0    0    2    3 
#> attr(,"class")
#> [1] "bbox"
sfc_multilinestring( obj = m, linestring_id = 1 )
#> [[1]]
#> [[1]]
#>      [,1] [,2]
#> [1,]    0    0
#> [2,]    0    1
#> [3,]    1    1
#> 
#> [[2]]
#>      [,1] [,2]
#> [1,]    2    2
#> [2,]    2    3
#> 
#> attr(,"class")
#> [1] "XY"              "MULTILINESTRING" "sfg"            
#> 
#> attr(,"n_empty")
#> [1] 0
#> attr(,"crs")
#> $input
#> [1] NA
#> 
#> $wkt
#> [1] NA
#> 
#> attr(,"class")
#> [1] "crs"
#> attr(,"class")
#> [1] "sfc_MULTILINESTRING" "sfc"                
#> attr(,"precision")
#> [1] 0
#> attr(,"bbox")
#> xmin ymin xmax ymax 
#>    0    0    2    3 
#> attr(,"class")
#> [1] "bbox"

sfc_multilinestring( obj = m, linestring_id = 1, multilinestring_id = 1 )
#> [[1]]
#> [[1]]
#>      [,1] [,2]
#> [1,]    0    0
#> [2,]    0    1
#> [3,]    1    1
#> 
#> attr(,"class")
#> [1] "XY"              "MULTILINESTRING" "sfg"            
#> 
#> [[2]]
#> [[1]]
#>      [,1] [,2]
#> [1,]    2    2
#> [2,]    2    3
#> 
#> attr(,"class")
#> [1] "XY"              "MULTILINESTRING" "sfg"            
#> 
#> attr(,"n_empty")
#> [1] 0
#> attr(,"crs")
#> $input
#> [1] NA
#> 
#> $wkt
#> [1] NA
#> 
#> attr(,"class")
#> [1] "crs"
#> attr(,"class")
#> [1] "sfc_MULTILINESTRING" "sfc"                
#> attr(,"precision")
#> [1] 0
#> attr(,"bbox")
#> xmin ymin xmax ymax 
#>    0    0    2    3 
#> attr(,"class")
#> [1] "bbox"

sfc_multilinestring( obj = m, x = 2, y = 3 )
#> [[1]]
#> [[1]]
#>      [,1] [,2]
#> [1,]    0    0
#> [2,]    0    1
#> [3,]    1    1
#> [4,]    2    2
#> [5,]    2    3
#> 
#> attr(,"class")
#> [1] "XY"              "MULTILINESTRING" "sfg"            
#> 
#> attr(,"n_empty")
#> [1] 0
#> attr(,"crs")
#> $input
#> [1] NA
#> 
#> $wkt
#> [1] NA
#> 
#> attr(,"class")
#> [1] "crs"
#> attr(,"class")
#> [1] "sfc_MULTILINESTRING" "sfc"                
#> attr(,"precision")
#> [1] 0
#> attr(,"bbox")
#> xmin ymin xmax ymax 
#>    0    0    2    3 
#> attr(,"class")
#> [1] "bbox"
sfc_multilinestring( obj = m, x = 1, y = 2, z = 3 )
#> [[1]]
#> [[1]]
#>      [,1] [,2] [,3]
#> [1,]    0    0    0
#> [2,]    0    0    1
#> [3,]    0    1    1
#> [4,]    1    2    2
#> [5,]    1    2    3
#> 
#> attr(,"class")
#> [1] "XYZ"             "MULTILINESTRING" "sfg"            
#> 
#> attr(,"n_empty")
#> [1] 0
#> attr(,"crs")
#> $input
#> [1] NA
#> 
#> $wkt
#> [1] NA
#> 
#> attr(,"class")
#> [1] "crs"
#> attr(,"class")
#> [1] "sfc_MULTILINESTRING" "sfc"                
#> attr(,"precision")
#> [1] 0
#> attr(,"bbox")
#> xmin ymin xmax ymax 
#>    0    0    1    2 
#> attr(,"class")
#> [1] "bbox"
#> attr(,"z_range")
#> zmin zmax 
#>    0    3 
#> attr(,"class")
#> [1] "z_range"
sfc_multilinestring( obj = m, x = 2, y = 3, linestring_id = 1, multilinestring_id = 1 )
#> [[1]]
#> [[1]]
#>      [,1] [,2]
#> [1,]    0    0
#> [2,]    0    1
#> [3,]    1    1
#> 
#> attr(,"class")
#> [1] "XY"              "MULTILINESTRING" "sfg"            
#> 
#> [[2]]
#> [[1]]
#>      [,1] [,2]
#> [1,]    2    2
#> [2,]    2    3
#> 
#> attr(,"class")
#> [1] "XY"              "MULTILINESTRING" "sfg"            
#> 
#> attr(,"n_empty")
#> [1] 0
#> attr(,"crs")
#> $input
#> [1] NA
#> 
#> $wkt
#> [1] NA
#> 
#> attr(,"class")
#> [1] "crs"
#> attr(,"class")
#> [1] "sfc_MULTILINESTRING" "sfc"                
#> attr(,"precision")
#> [1] 0
#> attr(,"bbox")
#> xmin ymin xmax ymax 
#>    0    0    2    3 
#> attr(,"class")
#> [1] "bbox"

df <- data.frame(
  ml_id = c(1,1,1,1,1,1,1,1,2,2,2,2,2)
  , l_id = c(1,1,1,2,2,3,3,3,1,1,1,2,2)
  , x = rnorm(13)
  , y = rnorm(13)
  , z = rnorm(13)
  , m = rnorm(13)
)

sfc_multilinestring( obj = df, x = "x", y = "y")
#> [[1]]
#> [[1]]
#>             [,1]        [,2]
#>  [1,]  0.6766680 -0.19532194
#>  [2,]  0.2225271 -0.94273223
#>  [3,] -0.8614526  0.02132701
#>  [4,] -1.9043411 -1.79178299
#>  [5,] -0.4022813  1.47243418
#>  [6,]  0.6102084  0.94482499
#>  [7,]  0.9325165  0.11343139
#>  [8,]  0.2306830  0.23620205
#>  [9,]  0.1066735  2.03365652
#> [10,] -0.2610999 -0.66589325
#> [11,]  0.9306085  0.80852346
#> [12,]  0.7688857  1.25823974
#> [13,] -1.9413764 -1.14196850
#> 
#> attr(,"class")
#> [1] "XY"              "MULTILINESTRING" "sfg"            
#> 
#> attr(,"n_empty")
#> [1] 0
#> attr(,"crs")
#> $input
#> [1] NA
#> 
#> $wkt
#> [1] NA
#> 
#> attr(,"class")
#> [1] "crs"
#> attr(,"class")
#> [1] "sfc_MULTILINESTRING" "sfc"                
#> attr(,"precision")
#> [1] 0
#> attr(,"bbox")
#>       xmin       ymin       xmax       ymax 
#> -1.9413764 -1.7917830  0.9325165  2.0336565 
#> attr(,"class")
#> [1] "bbox"
sfc_multilinestring( obj = df, x = "x", y = "y", z = "z")
#> [[1]]
#> [[1]]
#>             [,1]        [,2]         [,3]
#>  [1,]  0.6766680 -0.19532194  0.555104361
#>  [2,]  0.2225271 -0.94273223 -0.995235120
#>  [3,] -0.8614526  0.02132701  0.293345957
#>  [4,] -1.9043411 -1.79178299 -0.713616085
#>  [5,] -0.4022813  1.47243418  0.349705508
#>  [6,]  0.6102084  0.94482499 -0.516070547
#>  [7,]  0.9325165  0.11343139 -1.495227945
#>  [8,]  0.2306830  0.23620205 -0.125046614
#>  [9,]  0.1066735  2.03365652 -0.005682587
#> [10,] -0.2610999 -0.66589325  0.754338375
#> [11,]  0.9306085  0.80852346  0.833877542
#> [12,]  0.7688857  1.25823974 -0.945494973
#> [13,] -1.9413764 -1.14196850  0.848893776
#> 
#> attr(,"class")
#> [1] "XYZ"             "MULTILINESTRING" "sfg"            
#> 
#> attr(,"n_empty")
#> [1] 0
#> attr(,"crs")
#> $input
#> [1] NA
#> 
#> $wkt
#> [1] NA
#> 
#> attr(,"class")
#> [1] "crs"
#> attr(,"class")
#> [1] "sfc_MULTILINESTRING" "sfc"                
#> attr(,"precision")
#> [1] 0
#> attr(,"bbox")
#>       xmin       ymin       xmax       ymax 
#> -1.9413764 -1.7917830  0.9325165  2.0336565 
#> attr(,"class")
#> [1] "bbox"
#> attr(,"z_range")
#>       zmin       zmax 
#> -1.4952279  0.8488938 
#> attr(,"class")
#> [1] "z_range"
sfc_multilinestring( obj = df, x = "x", y = "y", z = "z", m = "m")
#> [[1]]
#> [[1]]
#>             [,1]        [,2]         [,3]       [,4]
#>  [1,]  0.6766680 -0.19532194  0.555104361 -0.5854161
#>  [2,]  0.2225271 -0.94273223 -0.995235120  0.5190141
#>  [3,] -0.8614526  0.02132701  0.293345957 -0.2763932
#>  [4,] -1.9043411 -1.79178299 -0.713616085  0.9859690
#>  [5,] -0.4022813  1.47243418  0.349705508  0.4926951
#>  [6,]  0.6102084  0.94482499 -0.516070547 -1.4969965
#>  [7,]  0.9325165  0.11343139 -1.495227945 -0.7684963
#>  [8,]  0.2306830  0.23620205 -0.125046614 -1.4504372
#>  [9,]  0.1066735  2.03365652 -0.005682587 -0.6160634
#> [10,] -0.2610999 -0.66589325  0.754338375 -0.6213298
#> [11,]  0.9306085  0.80852346  0.833877542  1.8308418
#> [12,]  0.7688857  1.25823974 -0.945494973  1.3157485
#> [13,] -1.9413764 -1.14196850  0.848893776 -0.5869711
#> 
#> attr(,"class")
#> [1] "XYZM"            "MULTILINESTRING" "sfg"            
#> 
#> attr(,"n_empty")
#> [1] 0
#> attr(,"crs")
#> $input
#> [1] NA
#> 
#> $wkt
#> [1] NA
#> 
#> attr(,"class")
#> [1] "crs"
#> attr(,"class")
#> [1] "sfc_MULTILINESTRING" "sfc"                
#> attr(,"precision")
#> [1] 0
#> attr(,"bbox")
#>       xmin       ymin       xmax       ymax 
#> -1.9413764 -1.7917830  0.9325165  2.0336565 
#> attr(,"class")
#> [1] "bbox"
#> attr(,"z_range")
#>       zmin       zmax 
#> -1.4952279  0.8488938 
#> attr(,"class")
#> [1] "z_range"
#> attr(,"m_range")
#>      mmin      mmax 
#> -1.496997  1.830842 
#> attr(,"class")
#> [1] "m_range"

sfc_multilinestring( obj = df, x = 2, y = 3)
#> [[1]]
#> [[1]]
#>       [,1]       [,2]
#>  [1,]    1  0.6766680
#>  [2,]    1  0.2225271
#>  [3,]    1 -0.8614526
#>  [4,]    2 -1.9043411
#>  [5,]    2 -0.4022813
#>  [6,]    3  0.6102084
#>  [7,]    3  0.9325165
#>  [8,]    3  0.2306830
#>  [9,]    1  0.1066735
#> [10,]    1 -0.2610999
#> [11,]    1  0.9306085
#> [12,]    2  0.7688857
#> [13,]    2 -1.9413764
#> 
#> attr(,"class")
#> [1] "XY"              "MULTILINESTRING" "sfg"            
#> 
#> attr(,"n_empty")
#> [1] 0
#> attr(,"crs")
#> $input
#> [1] NA
#> 
#> $wkt
#> [1] NA
#> 
#> attr(,"class")
#> [1] "crs"
#> attr(,"class")
#> [1] "sfc_MULTILINESTRING" "sfc"                
#> attr(,"precision")
#> [1] 0
#> attr(,"bbox")
#>       xmin       ymin       xmax       ymax 
#>  1.0000000 -1.9413764  3.0000000  0.9325165 
#> attr(,"class")
#> [1] "bbox"
sfc_multilinestring( obj = df, x = 2, y = 3, z = 4)
#> [[1]]
#> [[1]]
#>       [,1]       [,2]        [,3]
#>  [1,]    1  0.6766680 -0.19532194
#>  [2,]    1  0.2225271 -0.94273223
#>  [3,]    1 -0.8614526  0.02132701
#>  [4,]    2 -1.9043411 -1.79178299
#>  [5,]    2 -0.4022813  1.47243418
#>  [6,]    3  0.6102084  0.94482499
#>  [7,]    3  0.9325165  0.11343139
#>  [8,]    3  0.2306830  0.23620205
#>  [9,]    1  0.1066735  2.03365652
#> [10,]    1 -0.2610999 -0.66589325
#> [11,]    1  0.9306085  0.80852346
#> [12,]    2  0.7688857  1.25823974
#> [13,]    2 -1.9413764 -1.14196850
#> 
#> attr(,"class")
#> [1] "XYZ"             "MULTILINESTRING" "sfg"            
#> 
#> attr(,"n_empty")
#> [1] 0
#> attr(,"crs")
#> $input
#> [1] NA
#> 
#> $wkt
#> [1] NA
#> 
#> attr(,"class")
#> [1] "crs"
#> attr(,"class")
#> [1] "sfc_MULTILINESTRING" "sfc"                
#> attr(,"precision")
#> [1] 0
#> attr(,"bbox")
#>       xmin       ymin       xmax       ymax 
#>  1.0000000 -1.9413764  3.0000000  0.9325165 
#> attr(,"class")
#> [1] "bbox"
#> attr(,"z_range")
#>      zmin      zmax 
#> -1.791783  2.033657 
#> attr(,"class")
#> [1] "z_range"
sfc_multilinestring( obj = df, x = 2, y = 3, z = 4, m = 5)
#> [[1]]
#> [[1]]
#>       [,1]       [,2]        [,3]         [,4]
#>  [1,]    1  0.6766680 -0.19532194  0.555104361
#>  [2,]    1  0.2225271 -0.94273223 -0.995235120
#>  [3,]    1 -0.8614526  0.02132701  0.293345957
#>  [4,]    2 -1.9043411 -1.79178299 -0.713616085
#>  [5,]    2 -0.4022813  1.47243418  0.349705508
#>  [6,]    3  0.6102084  0.94482499 -0.516070547
#>  [7,]    3  0.9325165  0.11343139 -1.495227945
#>  [8,]    3  0.2306830  0.23620205 -0.125046614
#>  [9,]    1  0.1066735  2.03365652 -0.005682587
#> [10,]    1 -0.2610999 -0.66589325  0.754338375
#> [11,]    1  0.9306085  0.80852346  0.833877542
#> [12,]    2  0.7688857  1.25823974 -0.945494973
#> [13,]    2 -1.9413764 -1.14196850  0.848893776
#> 
#> attr(,"class")
#> [1] "XYZM"            "MULTILINESTRING" "sfg"            
#> 
#> attr(,"n_empty")
#> [1] 0
#> attr(,"crs")
#> $input
#> [1] NA
#> 
#> $wkt
#> [1] NA
#> 
#> attr(,"class")
#> [1] "crs"
#> attr(,"class")
#> [1] "sfc_MULTILINESTRING" "sfc"                
#> attr(,"precision")
#> [1] 0
#> attr(,"bbox")
#>       xmin       ymin       xmax       ymax 
#>  1.0000000 -1.9413764  3.0000000  0.9325165 
#> attr(,"class")
#> [1] "bbox"
#> attr(,"z_range")
#>      zmin      zmax 
#> -1.791783  2.033657 
#> attr(,"class")
#> [1] "z_range"
#> attr(,"m_range")
#>       mmin       mmax 
#> -1.4952279  0.8488938 
#> attr(,"class")
#> [1] "m_range"

sfc_multilinestring( obj = df, multilinestring_id = "ml_id", linestring_id = "l_id" )
#> [[1]]
#> [[1]]
#>            [,1]        [,2]       [,3]       [,4]
#> [1,]  0.6766680 -0.19532194  0.5551044 -0.5854161
#> [2,]  0.2225271 -0.94273223 -0.9952351  0.5190141
#> [3,] -0.8614526  0.02132701  0.2933460 -0.2763932
#> 
#> [[2]]
#>            [,1]      [,2]       [,3]      [,4]
#> [1,] -1.9043411 -1.791783 -0.7136161 0.9859690
#> [2,] -0.4022813  1.472434  0.3497055 0.4926951
#> 
#> [[3]]
#>           [,1]      [,2]       [,3]       [,4]
#> [1,] 0.6102084 0.9448250 -0.5160705 -1.4969965
#> [2,] 0.9325165 0.1134314 -1.4952279 -0.7684963
#> [3,] 0.2306830 0.2362020 -0.1250466 -1.4504372
#> 
#> attr(,"class")
#> [1] "XYZM"            "MULTILINESTRING" "sfg"            
#> 
#> [[2]]
#> [[1]]
#>            [,1]       [,2]         [,3]       [,4]
#> [1,]  0.1066735  2.0336565 -0.005682587 -0.6160634
#> [2,] -0.2610999 -0.6658933  0.754338375 -0.6213298
#> [3,]  0.9306085  0.8085235  0.833877542  1.8308418
#> 
#> [[2]]
#>            [,1]      [,2]       [,3]       [,4]
#> [1,]  0.7688857  1.258240 -0.9454950  1.3157485
#> [2,] -1.9413764 -1.141969  0.8488938 -0.5869711
#> 
#> attr(,"class")
#> [1] "XYZM"            "MULTILINESTRING" "sfg"            
#> 
#> attr(,"n_empty")
#> [1] 0
#> attr(,"crs")
#> $input
#> [1] NA
#> 
#> $wkt
#> [1] NA
#> 
#> attr(,"class")
#> [1] "crs"
#> attr(,"class")
#> [1] "sfc_MULTILINESTRING" "sfc"                
#> attr(,"precision")
#> [1] 0
#> attr(,"bbox")
#>       xmin       ymin       xmax       ymax 
#> -1.9413764 -1.7917830  0.9325165  2.0336565 
#> attr(,"class")
#> [1] "bbox"
#> attr(,"z_range")
#>       zmin       zmax 
#> -1.4952279  0.8488938 
#> attr(,"class")
#> [1] "z_range"
#> attr(,"m_range")
#>      mmin      mmax 
#> -1.496997  1.830842 
#> attr(,"class")
#> [1] "m_range"
sfc_multilinestring( obj = df, multilinestring_id = 1, linestring_id = 2 )
#> [[1]]
#> [[1]]
#>            [,1]        [,2]       [,3]       [,4]
#> [1,]  0.6766680 -0.19532194  0.5551044 -0.5854161
#> [2,]  0.2225271 -0.94273223 -0.9952351  0.5190141
#> [3,] -0.8614526  0.02132701  0.2933460 -0.2763932
#> 
#> [[2]]
#>            [,1]      [,2]       [,3]      [,4]
#> [1,] -1.9043411 -1.791783 -0.7136161 0.9859690
#> [2,] -0.4022813  1.472434  0.3497055 0.4926951
#> 
#> [[3]]
#>           [,1]      [,2]       [,3]       [,4]
#> [1,] 0.6102084 0.9448250 -0.5160705 -1.4969965
#> [2,] 0.9325165 0.1134314 -1.4952279 -0.7684963
#> [3,] 0.2306830 0.2362020 -0.1250466 -1.4504372
#> 
#> attr(,"class")
#> [1] "XYZM"            "MULTILINESTRING" "sfg"            
#> 
#> [[2]]
#> [[1]]
#>            [,1]       [,2]         [,3]       [,4]
#> [1,]  0.1066735  2.0336565 -0.005682587 -0.6160634
#> [2,] -0.2610999 -0.6658933  0.754338375 -0.6213298
#> [3,]  0.9306085  0.8085235  0.833877542  1.8308418
#> 
#> [[2]]
#>            [,1]      [,2]       [,3]       [,4]
#> [1,]  0.7688857  1.258240 -0.9454950  1.3157485
#> [2,] -1.9413764 -1.141969  0.8488938 -0.5869711
#> 
#> attr(,"class")
#> [1] "XYZM"            "MULTILINESTRING" "sfg"            
#> 
#> attr(,"n_empty")
#> [1] 0
#> attr(,"crs")
#> $input
#> [1] NA
#> 
#> $wkt
#> [1] NA
#> 
#> attr(,"class")
#> [1] "crs"
#> attr(,"class")
#> [1] "sfc_MULTILINESTRING" "sfc"                
#> attr(,"precision")
#> [1] 0
#> attr(,"bbox")
#>       xmin       ymin       xmax       ymax 
#> -1.9413764 -1.7917830  0.9325165  2.0336565 
#> attr(,"class")
#> [1] "bbox"
#> attr(,"z_range")
#>       zmin       zmax 
#> -1.4952279  0.8488938 
#> attr(,"class")
#> [1] "z_range"
#> attr(,"m_range")
#>      mmin      mmax 
#> -1.496997  1.830842 
#> attr(,"class")
#> [1] "m_range"