sfc POLYGON

constructs an sfc of POLYGON objects

sfc_polygon(
  obj = NULL,
  x = NULL,
  y = NULL,
  z = NULL,
  m = NULL,
  polygon_id = NULL,
  linestring_id = NULL,
  close = TRUE
)

Arguments

obj

sorted matrix or data.frame

x

x geometry column

y

y geometry column

z

z geometry column

m

m geometry column

polygon_id

column of ids for polygons

linestring_id

column of ids for lines (within polygons)

close

logical indicating whether polygons should be closed. If TRUE, all polygons will be checked and force closed if possible

Value

sfc object of POLYGON geometries

notes

sfheaders functions do not perform any validity checks on the geometries. Nor do they set Coordinate Reference Systems, EPSG, PROJ4 or precision attributes.

The data.frame and matrices you send into the sfheader functions must be ordered.

Examples

m <- matrix(c(0,0,0,0,1,1), ncol = 2 )
sfc_polygon( m )
#> [[1]]
#> [[1]]
#>      [,1] [,2]
#> [1,]    0    0
#> [2,]    0    1
#> [3,]    0    1
#> [4,]    0    0
#> 
#> attr(,"class")
#> [1] "XY"      "POLYGON" "sfg"    
#> 
#> attr(,"n_empty")
#> [1] 0
#> attr(,"crs")
#> $input
#> [1] NA
#> 
#> $wkt
#> [1] NA
#> 
#> attr(,"class")
#> [1] "crs"
#> attr(,"class")
#> [1] "sfc_POLYGON" "sfc"        
#> attr(,"precision")
#> [1] 0
#> attr(,"bbox")
#> xmin ymin xmax ymax 
#>    0    0    0    1 
#> attr(,"class")
#> [1] "bbox"

m <- matrix(c(0,0,0,0,0,1,0,1,1,1,2,2,1,2,3,1,3,2), ncol = 3, byrow = TRUE)
sfc_polygon( obj = m )
#> [[1]]
#> [[1]]
#>      [,1] [,2] [,3]
#> [1,]    0    0    0
#> [2,]    0    0    1
#> [3,]    0    1    1
#> [4,]    1    2    2
#> [5,]    1    2    3
#> [6,]    1    3    2
#> [7,]    0    0    0
#> 
#> attr(,"class")
#> [1] "XYZ"     "POLYGON" "sfg"    
#> 
#> attr(,"n_empty")
#> [1] 0
#> attr(,"crs")
#> $input
#> [1] NA
#> 
#> $wkt
#> [1] NA
#> 
#> attr(,"class")
#> [1] "crs"
#> attr(,"class")
#> [1] "sfc_POLYGON" "sfc"        
#> attr(,"precision")
#> [1] 0
#> attr(,"bbox")
#> xmin ymin xmax ymax 
#>    0    0    1    3 
#> attr(,"class")
#> [1] "bbox"
#> attr(,"z_range")
#> zmin zmax 
#>    0    3 
#> attr(,"class")
#> [1] "z_range"
sfc_polygon( obj = m, polygon_id = 1 )
#> [[1]]
#> [[1]]
#>      [,1] [,2]
#> [1,]    0    0
#> [2,]    0    1
#> [3,]    1    1
#> [4,]    0    0
#> 
#> attr(,"class")
#> [1] "XY"      "POLYGON" "sfg"    
#> 
#> [[2]]
#> [[1]]
#>      [,1] [,2]
#> [1,]    2    2
#> [2,]    2    3
#> [3,]    3    2
#> [4,]    2    2
#> 
#> attr(,"class")
#> [1] "XY"      "POLYGON" "sfg"    
#> 
#> attr(,"n_empty")
#> [1] 0
#> attr(,"crs")
#> $input
#> [1] NA
#> 
#> $wkt
#> [1] NA
#> 
#> attr(,"class")
#> [1] "crs"
#> attr(,"class")
#> [1] "sfc_POLYGON" "sfc"        
#> attr(,"precision")
#> [1] 0
#> attr(,"bbox")
#> xmin ymin xmax ymax 
#>    0    0    3    3 
#> attr(,"class")
#> [1] "bbox"
sfc_polygon( obj = m, linestring_id = 1 )
#> [[1]]
#> [[1]]
#>      [,1] [,2]
#> [1,]    0    0
#> [2,]    0    1
#> [3,]    1    1
#> [4,]    0    0
#> 
#> [[2]]
#>      [,1] [,2]
#> [1,]    2    2
#> [2,]    2    3
#> [3,]    3    2
#> [4,]    2    2
#> 
#> attr(,"class")
#> [1] "XY"      "POLYGON" "sfg"    
#> 
#> attr(,"n_empty")
#> [1] 0
#> attr(,"crs")
#> $input
#> [1] NA
#> 
#> $wkt
#> [1] NA
#> 
#> attr(,"class")
#> [1] "crs"
#> attr(,"class")
#> [1] "sfc_POLYGON" "sfc"        
#> attr(,"precision")
#> [1] 0
#> attr(,"bbox")
#> xmin ymin xmax ymax 
#>    0    0    3    3 
#> attr(,"class")
#> [1] "bbox"

sfc_polygon( obj = m, linestring_id = 1, polygon_id = 1 )
#> [[1]]
#> [[1]]
#>      [,1] [,2]
#> [1,]    0    0
#> [2,]    0    1
#> [3,]    1    1
#> [4,]    0    0
#> 
#> attr(,"class")
#> [1] "XY"      "POLYGON" "sfg"    
#> 
#> [[2]]
#> [[1]]
#>      [,1] [,2]
#> [1,]    2    2
#> [2,]    2    3
#> [3,]    3    2
#> [4,]    2    2
#> 
#> attr(,"class")
#> [1] "XY"      "POLYGON" "sfg"    
#> 
#> attr(,"n_empty")
#> [1] 0
#> attr(,"crs")
#> $input
#> [1] NA
#> 
#> $wkt
#> [1] NA
#> 
#> attr(,"class")
#> [1] "crs"
#> attr(,"class")
#> [1] "sfc_POLYGON" "sfc"        
#> attr(,"precision")
#> [1] 0
#> attr(,"bbox")
#> xmin ymin xmax ymax 
#>    0    0    3    3 
#> attr(,"class")
#> [1] "bbox"

sfc_polygon( obj = m, x = 2, y = 3 )
#> [[1]]
#> [[1]]
#>      [,1] [,2]
#> [1,]    0    0
#> [2,]    0    1
#> [3,]    1    1
#> [4,]    2    2
#> [5,]    2    3
#> [6,]    3    2
#> [7,]    0    0
#> 
#> attr(,"class")
#> [1] "XY"      "POLYGON" "sfg"    
#> 
#> attr(,"n_empty")
#> [1] 0
#> attr(,"crs")
#> $input
#> [1] NA
#> 
#> $wkt
#> [1] NA
#> 
#> attr(,"class")
#> [1] "crs"
#> attr(,"class")
#> [1] "sfc_POLYGON" "sfc"        
#> attr(,"precision")
#> [1] 0
#> attr(,"bbox")
#> xmin ymin xmax ymax 
#>    0    0    3    3 
#> attr(,"class")
#> [1] "bbox"
sfc_polygon( obj = m, x = 1, y = 2, z = 3 )
#> [[1]]
#> [[1]]
#>      [,1] [,2] [,3]
#> [1,]    0    0    0
#> [2,]    0    0    1
#> [3,]    0    1    1
#> [4,]    1    2    2
#> [5,]    1    2    3
#> [6,]    1    3    2
#> [7,]    0    0    0
#> 
#> attr(,"class")
#> [1] "XYZ"     "POLYGON" "sfg"    
#> 
#> attr(,"n_empty")
#> [1] 0
#> attr(,"crs")
#> $input
#> [1] NA
#> 
#> $wkt
#> [1] NA
#> 
#> attr(,"class")
#> [1] "crs"
#> attr(,"class")
#> [1] "sfc_POLYGON" "sfc"        
#> attr(,"precision")
#> [1] 0
#> attr(,"bbox")
#> xmin ymin xmax ymax 
#>    0    0    1    3 
#> attr(,"class")
#> [1] "bbox"
#> attr(,"z_range")
#> zmin zmax 
#>    0    3 
#> attr(,"class")
#> [1] "z_range"
sfc_polygon( obj = m, x = 2, y = 3, linestring_id = 1, polygon_id = 1 )
#> [[1]]
#> [[1]]
#>      [,1] [,2]
#> [1,]    0    0
#> [2,]    0    1
#> [3,]    1    1
#> [4,]    0    0
#> 
#> attr(,"class")
#> [1] "XY"      "POLYGON" "sfg"    
#> 
#> [[2]]
#> [[1]]
#>      [,1] [,2]
#> [1,]    2    2
#> [2,]    2    3
#> [3,]    3    2
#> [4,]    2    2
#> 
#> attr(,"class")
#> [1] "XY"      "POLYGON" "sfg"    
#> 
#> attr(,"n_empty")
#> [1] 0
#> attr(,"crs")
#> $input
#> [1] NA
#> 
#> $wkt
#> [1] NA
#> 
#> attr(,"class")
#> [1] "crs"
#> attr(,"class")
#> [1] "sfc_POLYGON" "sfc"        
#> attr(,"precision")
#> [1] 0
#> attr(,"bbox")
#> xmin ymin xmax ymax 
#>    0    0    3    3 
#> attr(,"class")
#> [1] "bbox"

df <- data.frame(
  ml_id = c(1,1,1,1,1,1,1,1,1,2,2,2,2,2,2)
  , l_id = c(1,1,1,2,2,2,3,3,3,1,1,1,2,2,2)
  , x = rnorm(15)
  , y = rnorm(15)
  , z = rnorm(15)
  , m = rnorm(15)
)

sfc_polygon( obj = df, x = "x", y = "y")
#> [[1]]
#> [[1]]
#>              [,1]        [,2]
#>  [1,] -1.04340816 -0.20935732
#>  [2,] -0.78939368  0.49171330
#>  [3,]  0.20904375  0.36401471
#>  [4,]  1.24449973 -0.65465370
#>  [5,] -0.84405786 -0.39324829
#>  [6,]  2.01401399  0.44001959
#>  [7,] -0.80038264  0.11345993
#>  [8,] -0.24952537 -0.05702265
#>  [9,] -0.38532758 -0.04214736
#> [10,] -0.02314503 -0.66277432
#> [11,]  0.45628543  1.53638278
#> [12,]  0.44377112 -0.75038216
#> [13,] -0.25686579 -0.76870465
#> [14,]  0.57913432 -1.00920333
#> [15,]  0.20056236 -1.91242558
#> [16,] -1.04340816 -0.20935732
#> 
#> attr(,"class")
#> [1] "XY"      "POLYGON" "sfg"    
#> 
#> attr(,"n_empty")
#> [1] 0
#> attr(,"crs")
#> $input
#> [1] NA
#> 
#> $wkt
#> [1] NA
#> 
#> attr(,"class")
#> [1] "crs"
#> attr(,"class")
#> [1] "sfc_POLYGON" "sfc"        
#> attr(,"precision")
#> [1] 0
#> attr(,"bbox")
#>      xmin      ymin      xmax      ymax 
#> -1.043408 -1.912426  2.014014  1.536383 
#> attr(,"class")
#> [1] "bbox"
sfc_polygon( obj = df, x = "x", y = "y", z = "z")
#> [[1]]
#> [[1]]
#>              [,1]        [,2]       [,3]
#>  [1,] -1.04340816 -0.20935732 -0.2161252
#>  [2,] -0.78939368  0.49171330  0.1145929
#>  [3,]  0.20904375  0.36401471  0.3914653
#>  [4,]  1.24449973 -0.65465370  0.5557399
#>  [5,] -0.84405786 -0.39324829 -0.4478459
#>  [6,]  2.01401399  0.44001959  1.0288817
#>  [7,] -0.80038264  0.11345993 -0.8429896
#>  [8,] -0.24952537 -0.05702265 -0.6984916
#>  [9,] -0.38532758 -0.04214736  0.2140222
#> [10,] -0.02314503 -0.66277432  2.4654550
#> [11,]  0.45628543  1.53638278 -1.6048626
#> [12,]  0.44377112 -0.75038216  0.6222578
#> [13,] -0.25686579 -0.76870465 -0.5368717
#> [14,]  0.57913432 -1.00920333 -1.3588827
#> [15,]  0.20056236 -1.91242558 -0.7343261
#> [16,] -1.04340816 -0.20935732 -0.2161252
#> 
#> attr(,"class")
#> [1] "XYZ"     "POLYGON" "sfg"    
#> 
#> attr(,"n_empty")
#> [1] 0
#> attr(,"crs")
#> $input
#> [1] NA
#> 
#> $wkt
#> [1] NA
#> 
#> attr(,"class")
#> [1] "crs"
#> attr(,"class")
#> [1] "sfc_POLYGON" "sfc"        
#> attr(,"precision")
#> [1] 0
#> attr(,"bbox")
#>      xmin      ymin      xmax      ymax 
#> -1.043408 -1.912426  2.014014  1.536383 
#> attr(,"class")
#> [1] "bbox"
#> attr(,"z_range")
#>      zmin      zmax 
#> -1.604863  2.465455 
#> attr(,"class")
#> [1] "z_range"
sfc_polygon( obj = df, x = "x", y = "y", z = "z", m = "m")
#> [[1]]
#> [[1]]
#>              [,1]        [,2]       [,3]        [,4]
#>  [1,] -1.04340816 -0.20935732 -0.2161252  1.81843635
#>  [2,] -0.78939368  0.49171330  0.1145929  0.02695604
#>  [3,]  0.20904375  0.36401471  0.3914653  1.04446107
#>  [4,]  1.24449973 -0.65465370  0.5557399 -0.29510755
#>  [5,] -0.84405786 -0.39324829 -0.4478459 -0.30593257
#>  [6,]  2.01401399  0.44001959  1.0288817 -1.43320967
#>  [7,] -0.80038264  0.11345993 -0.8429896  0.46293913
#>  [8,] -0.24952537 -0.05702265 -0.6984916 -0.64976312
#>  [9,] -0.38532758 -0.04214736  0.2140222 -0.90313326
#> [10,] -0.02314503 -0.66277432  2.4654550 -1.28394440
#> [11,]  0.45628543  1.53638278 -1.6048626  0.75708906
#> [12,]  0.44377112 -0.75038216  0.6222578 -0.92541512
#> [13,] -0.25686579 -0.76870465 -0.5368717 -1.78747115
#> [14,]  0.57913432 -1.00920333 -1.3588827  1.08379085
#> [15,]  0.20056236 -1.91242558 -0.7343261  0.92671508
#> [16,] -1.04340816 -0.20935732 -0.2161252  1.81843635
#> 
#> attr(,"class")
#> [1] "XYZM"    "POLYGON" "sfg"    
#> 
#> attr(,"n_empty")
#> [1] 0
#> attr(,"crs")
#> $input
#> [1] NA
#> 
#> $wkt
#> [1] NA
#> 
#> attr(,"class")
#> [1] "crs"
#> attr(,"class")
#> [1] "sfc_POLYGON" "sfc"        
#> attr(,"precision")
#> [1] 0
#> attr(,"bbox")
#>      xmin      ymin      xmax      ymax 
#> -1.043408 -1.912426  2.014014  1.536383 
#> attr(,"class")
#> [1] "bbox"
#> attr(,"z_range")
#>      zmin      zmax 
#> -1.604863  2.465455 
#> attr(,"class")
#> [1] "z_range"
#> attr(,"m_range")
#>      mmin      mmax 
#> -1.787471  1.818436 
#> attr(,"class")
#> [1] "m_range"

sfc_polygon( obj = df, x = 2, y = 3)
#> [[1]]
#> [[1]]
#>       [,1]        [,2]
#>  [1,]    1 -1.04340816
#>  [2,]    1 -0.78939368
#>  [3,]    1  0.20904375
#>  [4,]    2  1.24449973
#>  [5,]    2 -0.84405786
#>  [6,]    2  2.01401399
#>  [7,]    3 -0.80038264
#>  [8,]    3 -0.24952537
#>  [9,]    3 -0.38532758
#> [10,]    1 -0.02314503
#> [11,]    1  0.45628543
#> [12,]    1  0.44377112
#> [13,]    2 -0.25686579
#> [14,]    2  0.57913432
#> [15,]    2  0.20056236
#> [16,]    1 -1.04340816
#> 
#> attr(,"class")
#> [1] "XY"      "POLYGON" "sfg"    
#> 
#> attr(,"n_empty")
#> [1] 0
#> attr(,"crs")
#> $input
#> [1] NA
#> 
#> $wkt
#> [1] NA
#> 
#> attr(,"class")
#> [1] "crs"
#> attr(,"class")
#> [1] "sfc_POLYGON" "sfc"        
#> attr(,"precision")
#> [1] 0
#> attr(,"bbox")
#>      xmin      ymin      xmax      ymax 
#>  1.000000 -1.043408  3.000000  2.014014 
#> attr(,"class")
#> [1] "bbox"
sfc_polygon( obj = df, x = 2, y = 3, z = 4)
#> [[1]]
#> [[1]]
#>       [,1]        [,2]        [,3]
#>  [1,]    1 -1.04340816 -0.20935732
#>  [2,]    1 -0.78939368  0.49171330
#>  [3,]    1  0.20904375  0.36401471
#>  [4,]    2  1.24449973 -0.65465370
#>  [5,]    2 -0.84405786 -0.39324829
#>  [6,]    2  2.01401399  0.44001959
#>  [7,]    3 -0.80038264  0.11345993
#>  [8,]    3 -0.24952537 -0.05702265
#>  [9,]    3 -0.38532758 -0.04214736
#> [10,]    1 -0.02314503 -0.66277432
#> [11,]    1  0.45628543  1.53638278
#> [12,]    1  0.44377112 -0.75038216
#> [13,]    2 -0.25686579 -0.76870465
#> [14,]    2  0.57913432 -1.00920333
#> [15,]    2  0.20056236 -1.91242558
#> [16,]    1 -1.04340816 -0.20935732
#> 
#> attr(,"class")
#> [1] "XYZ"     "POLYGON" "sfg"    
#> 
#> attr(,"n_empty")
#> [1] 0
#> attr(,"crs")
#> $input
#> [1] NA
#> 
#> $wkt
#> [1] NA
#> 
#> attr(,"class")
#> [1] "crs"
#> attr(,"class")
#> [1] "sfc_POLYGON" "sfc"        
#> attr(,"precision")
#> [1] 0
#> attr(,"bbox")
#>      xmin      ymin      xmax      ymax 
#>  1.000000 -1.043408  3.000000  2.014014 
#> attr(,"class")
#> [1] "bbox"
#> attr(,"z_range")
#>      zmin      zmax 
#> -1.912426  1.536383 
#> attr(,"class")
#> [1] "z_range"
sfc_polygon( obj = df, x = 2, y = 3, z = 4, m = 5)
#> [[1]]
#> [[1]]
#>       [,1]        [,2]        [,3]       [,4]
#>  [1,]    1 -1.04340816 -0.20935732 -0.2161252
#>  [2,]    1 -0.78939368  0.49171330  0.1145929
#>  [3,]    1  0.20904375  0.36401471  0.3914653
#>  [4,]    2  1.24449973 -0.65465370  0.5557399
#>  [5,]    2 -0.84405786 -0.39324829 -0.4478459
#>  [6,]    2  2.01401399  0.44001959  1.0288817
#>  [7,]    3 -0.80038264  0.11345993 -0.8429896
#>  [8,]    3 -0.24952537 -0.05702265 -0.6984916
#>  [9,]    3 -0.38532758 -0.04214736  0.2140222
#> [10,]    1 -0.02314503 -0.66277432  2.4654550
#> [11,]    1  0.45628543  1.53638278 -1.6048626
#> [12,]    1  0.44377112 -0.75038216  0.6222578
#> [13,]    2 -0.25686579 -0.76870465 -0.5368717
#> [14,]    2  0.57913432 -1.00920333 -1.3588827
#> [15,]    2  0.20056236 -1.91242558 -0.7343261
#> [16,]    1 -1.04340816 -0.20935732 -0.2161252
#> 
#> attr(,"class")
#> [1] "XYZM"    "POLYGON" "sfg"    
#> 
#> attr(,"n_empty")
#> [1] 0
#> attr(,"crs")
#> $input
#> [1] NA
#> 
#> $wkt
#> [1] NA
#> 
#> attr(,"class")
#> [1] "crs"
#> attr(,"class")
#> [1] "sfc_POLYGON" "sfc"        
#> attr(,"precision")
#> [1] 0
#> attr(,"bbox")
#>      xmin      ymin      xmax      ymax 
#>  1.000000 -1.043408  3.000000  2.014014 
#> attr(,"class")
#> [1] "bbox"
#> attr(,"z_range")
#>      zmin      zmax 
#> -1.912426  1.536383 
#> attr(,"class")
#> [1] "z_range"
#> attr(,"m_range")
#>      mmin      mmax 
#> -1.604863  2.465455 
#> attr(,"class")
#> [1] "m_range"

sfc_polygon( obj = df, polygon_id = "ml_id", linestring_id = "l_id" )
#> [[1]]
#> [[1]]
#>            [,1]       [,2]       [,3]       [,4]
#> [1,] -1.0434082 -0.2093573 -0.2161252 1.81843635
#> [2,] -0.7893937  0.4917133  0.1145929 0.02695604
#> [3,]  0.2090437  0.3640147  0.3914653 1.04446107
#> [4,] -1.0434082 -0.2093573 -0.2161252 1.81843635
#> 
#> [[2]]
#>            [,1]       [,2]       [,3]       [,4]
#> [1,]  1.2444997 -0.6546537  0.5557399 -0.2951076
#> [2,] -0.8440579 -0.3932483 -0.4478459 -0.3059326
#> [3,]  2.0140140  0.4400196  1.0288817 -1.4332097
#> [4,]  1.2444997 -0.6546537  0.5557399 -0.2951076
#> 
#> [[3]]
#>            [,1]        [,2]       [,3]       [,4]
#> [1,] -0.8003826  0.11345993 -0.8429896  0.4629391
#> [2,] -0.2495254 -0.05702265 -0.6984916 -0.6497631
#> [3,] -0.3853276 -0.04214736  0.2140222 -0.9031333
#> [4,] -0.8003826  0.11345993 -0.8429896  0.4629391
#> 
#> attr(,"class")
#> [1] "XYZM"    "POLYGON" "sfg"    
#> 
#> [[2]]
#> [[1]]
#>             [,1]       [,2]       [,3]       [,4]
#> [1,] -0.02314503 -0.6627743  2.4654550 -1.2839444
#> [2,]  0.45628543  1.5363828 -1.6048626  0.7570891
#> [3,]  0.44377112 -0.7503822  0.6222578 -0.9254151
#> [4,] -0.02314503 -0.6627743  2.4654550 -1.2839444
#> 
#> [[2]]
#>            [,1]       [,2]       [,3]       [,4]
#> [1,] -0.2568658 -0.7687047 -0.5368717 -1.7874712
#> [2,]  0.5791343 -1.0092033 -1.3588827  1.0837908
#> [3,]  0.2005624 -1.9124256 -0.7343261  0.9267151
#> [4,] -0.2568658 -0.7687047 -0.5368717 -1.7874712
#> 
#> attr(,"class")
#> [1] "XYZM"    "POLYGON" "sfg"    
#> 
#> attr(,"n_empty")
#> [1] 0
#> attr(,"crs")
#> $input
#> [1] NA
#> 
#> $wkt
#> [1] NA
#> 
#> attr(,"class")
#> [1] "crs"
#> attr(,"class")
#> [1] "sfc_POLYGON" "sfc"        
#> attr(,"precision")
#> [1] 0
#> attr(,"bbox")
#>      xmin      ymin      xmax      ymax 
#> -1.043408 -1.912426  2.014014  1.536383 
#> attr(,"class")
#> [1] "bbox"
#> attr(,"z_range")
#>      zmin      zmax 
#> -1.604863  2.465455 
#> attr(,"class")
#> [1] "z_range"
#> attr(,"m_range")
#>      mmin      mmax 
#> -1.787471  1.818436 
#> attr(,"class")
#> [1] "m_range"
sfc_polygon( obj = df, polygon_id = 1, linestring_id = 2 )
#> [[1]]
#> [[1]]
#>            [,1]       [,2]       [,3]       [,4]
#> [1,] -1.0434082 -0.2093573 -0.2161252 1.81843635
#> [2,] -0.7893937  0.4917133  0.1145929 0.02695604
#> [3,]  0.2090437  0.3640147  0.3914653 1.04446107
#> [4,] -1.0434082 -0.2093573 -0.2161252 1.81843635
#> 
#> [[2]]
#>            [,1]       [,2]       [,3]       [,4]
#> [1,]  1.2444997 -0.6546537  0.5557399 -0.2951076
#> [2,] -0.8440579 -0.3932483 -0.4478459 -0.3059326
#> [3,]  2.0140140  0.4400196  1.0288817 -1.4332097
#> [4,]  1.2444997 -0.6546537  0.5557399 -0.2951076
#> 
#> [[3]]
#>            [,1]        [,2]       [,3]       [,4]
#> [1,] -0.8003826  0.11345993 -0.8429896  0.4629391
#> [2,] -0.2495254 -0.05702265 -0.6984916 -0.6497631
#> [3,] -0.3853276 -0.04214736  0.2140222 -0.9031333
#> [4,] -0.8003826  0.11345993 -0.8429896  0.4629391
#> 
#> attr(,"class")
#> [1] "XYZM"    "POLYGON" "sfg"    
#> 
#> [[2]]
#> [[1]]
#>             [,1]       [,2]       [,3]       [,4]
#> [1,] -0.02314503 -0.6627743  2.4654550 -1.2839444
#> [2,]  0.45628543  1.5363828 -1.6048626  0.7570891
#> [3,]  0.44377112 -0.7503822  0.6222578 -0.9254151
#> [4,] -0.02314503 -0.6627743  2.4654550 -1.2839444
#> 
#> [[2]]
#>            [,1]       [,2]       [,3]       [,4]
#> [1,] -0.2568658 -0.7687047 -0.5368717 -1.7874712
#> [2,]  0.5791343 -1.0092033 -1.3588827  1.0837908
#> [3,]  0.2005624 -1.9124256 -0.7343261  0.9267151
#> [4,] -0.2568658 -0.7687047 -0.5368717 -1.7874712
#> 
#> attr(,"class")
#> [1] "XYZM"    "POLYGON" "sfg"    
#> 
#> attr(,"n_empty")
#> [1] 0
#> attr(,"crs")
#> $input
#> [1] NA
#> 
#> $wkt
#> [1] NA
#> 
#> attr(,"class")
#> [1] "crs"
#> attr(,"class")
#> [1] "sfc_POLYGON" "sfc"        
#> attr(,"precision")
#> [1] 0
#> attr(,"bbox")
#>      xmin      ymin      xmax      ymax 
#> -1.043408 -1.912426  2.014014  1.536383 
#> attr(,"class")
#> [1] "bbox"
#> attr(,"z_range")
#>      zmin      zmax 
#> -1.604863  2.465455 
#> attr(,"class")
#> [1] "z_range"
#> attr(,"m_range")
#>      mmin      mmax 
#> -1.787471  1.818436 
#> attr(,"class")
#> [1] "m_range"