sfc POLYGON

constructs an sfc of POLYGON objects

sfc_polygon(
  obj = NULL,
  x = NULL,
  y = NULL,
  z = NULL,
  m = NULL,
  polygon_id = NULL,
  linestring_id = NULL,
  close = TRUE
)

Arguments

obj

sorted matrix or data.frame

x

x geometry column

y

y geometry column

z

z geometry column

m

m geometry column

polygon_id

column of ids for polygons

linestring_id

column of ids for lines (within polygons)

close

logical indicating whether polygons should be closed. If TRUE, all polygons will be checked and force closed if possible

Value

sfc object of POLYGON geometries

notes

sfheaders functions do not perform any validity checks on the geometries. Nor do they set Coordinate Reference Systems, EPSG, PROJ4 or precision attributes.

The data.frame and matrices you send into the sfheader functions must be ordered.

Examples

m <- matrix(c(0,0,0,0,1,1), ncol = 2 )
sfc_polygon( m )
#> [[1]]
#> [[1]]
#>      [,1] [,2]
#> [1,]    0    0
#> [2,]    0    1
#> [3,]    0    1
#> [4,]    0    0
#> 
#> attr(,"class")
#> [1] "XY"      "POLYGON" "sfg"    
#> 
#> attr(,"n_empty")
#> [1] 0
#> attr(,"crs")
#> $input
#> [1] NA
#> 
#> $wkt
#> [1] NA
#> 
#> attr(,"class")
#> [1] "crs"
#> attr(,"class")
#> [1] "sfc_POLYGON" "sfc"        
#> attr(,"precision")
#> [1] 0
#> attr(,"bbox")
#> xmin ymin xmax ymax 
#>    0    0    0    1 
#> attr(,"class")
#> [1] "bbox"

m <- matrix(c(0,0,0,0,0,1,0,1,1,1,2,2,1,2,3,1,3,2), ncol = 3, byrow = TRUE)
sfc_polygon( obj = m )
#> [[1]]
#> [[1]]
#>      [,1] [,2] [,3]
#> [1,]    0    0    0
#> [2,]    0    0    1
#> [3,]    0    1    1
#> [4,]    1    2    2
#> [5,]    1    2    3
#> [6,]    1    3    2
#> [7,]    0    0    0
#> 
#> attr(,"class")
#> [1] "XYZ"     "POLYGON" "sfg"    
#> 
#> attr(,"n_empty")
#> [1] 0
#> attr(,"crs")
#> $input
#> [1] NA
#> 
#> $wkt
#> [1] NA
#> 
#> attr(,"class")
#> [1] "crs"
#> attr(,"class")
#> [1] "sfc_POLYGON" "sfc"        
#> attr(,"precision")
#> [1] 0
#> attr(,"bbox")
#> xmin ymin xmax ymax 
#>    0    0    1    3 
#> attr(,"class")
#> [1] "bbox"
#> attr(,"z_range")
#> zmin zmax 
#>    0    3 
#> attr(,"class")
#> [1] "z_range"
sfc_polygon( obj = m, polygon_id = 1 )
#> [[1]]
#> [[1]]
#>      [,1] [,2]
#> [1,]    0    0
#> [2,]    0    1
#> [3,]    1    1
#> [4,]    0    0
#> 
#> attr(,"class")
#> [1] "XY"      "POLYGON" "sfg"    
#> 
#> [[2]]
#> [[1]]
#>      [,1] [,2]
#> [1,]    2    2
#> [2,]    2    3
#> [3,]    3    2
#> [4,]    2    2
#> 
#> attr(,"class")
#> [1] "XY"      "POLYGON" "sfg"    
#> 
#> attr(,"n_empty")
#> [1] 0
#> attr(,"crs")
#> $input
#> [1] NA
#> 
#> $wkt
#> [1] NA
#> 
#> attr(,"class")
#> [1] "crs"
#> attr(,"class")
#> [1] "sfc_POLYGON" "sfc"        
#> attr(,"precision")
#> [1] 0
#> attr(,"bbox")
#> xmin ymin xmax ymax 
#>    0    0    3    3 
#> attr(,"class")
#> [1] "bbox"
sfc_polygon( obj = m, linestring_id = 1 )
#> [[1]]
#> [[1]]
#>      [,1] [,2]
#> [1,]    0    0
#> [2,]    0    1
#> [3,]    1    1
#> [4,]    0    0
#> 
#> [[2]]
#>      [,1] [,2]
#> [1,]    2    2
#> [2,]    2    3
#> [3,]    3    2
#> [4,]    2    2
#> 
#> attr(,"class")
#> [1] "XY"      "POLYGON" "sfg"    
#> 
#> attr(,"n_empty")
#> [1] 0
#> attr(,"crs")
#> $input
#> [1] NA
#> 
#> $wkt
#> [1] NA
#> 
#> attr(,"class")
#> [1] "crs"
#> attr(,"class")
#> [1] "sfc_POLYGON" "sfc"        
#> attr(,"precision")
#> [1] 0
#> attr(,"bbox")
#> xmin ymin xmax ymax 
#>    0    0    3    3 
#> attr(,"class")
#> [1] "bbox"

sfc_polygon( obj = m, linestring_id = 1, polygon_id = 1 )
#> [[1]]
#> [[1]]
#>      [,1] [,2]
#> [1,]    0    0
#> [2,]    0    1
#> [3,]    1    1
#> [4,]    0    0
#> 
#> attr(,"class")
#> [1] "XY"      "POLYGON" "sfg"    
#> 
#> [[2]]
#> [[1]]
#>      [,1] [,2]
#> [1,]    2    2
#> [2,]    2    3
#> [3,]    3    2
#> [4,]    2    2
#> 
#> attr(,"class")
#> [1] "XY"      "POLYGON" "sfg"    
#> 
#> attr(,"n_empty")
#> [1] 0
#> attr(,"crs")
#> $input
#> [1] NA
#> 
#> $wkt
#> [1] NA
#> 
#> attr(,"class")
#> [1] "crs"
#> attr(,"class")
#> [1] "sfc_POLYGON" "sfc"        
#> attr(,"precision")
#> [1] 0
#> attr(,"bbox")
#> xmin ymin xmax ymax 
#>    0    0    3    3 
#> attr(,"class")
#> [1] "bbox"

sfc_polygon( obj = m, x = 2, y = 3 )
#> [[1]]
#> [[1]]
#>      [,1] [,2]
#> [1,]    0    0
#> [2,]    0    1
#> [3,]    1    1
#> [4,]    2    2
#> [5,]    2    3
#> [6,]    3    2
#> [7,]    0    0
#> 
#> attr(,"class")
#> [1] "XY"      "POLYGON" "sfg"    
#> 
#> attr(,"n_empty")
#> [1] 0
#> attr(,"crs")
#> $input
#> [1] NA
#> 
#> $wkt
#> [1] NA
#> 
#> attr(,"class")
#> [1] "crs"
#> attr(,"class")
#> [1] "sfc_POLYGON" "sfc"        
#> attr(,"precision")
#> [1] 0
#> attr(,"bbox")
#> xmin ymin xmax ymax 
#>    0    0    3    3 
#> attr(,"class")
#> [1] "bbox"
sfc_polygon( obj = m, x = 1, y = 2, z = 3 )
#> [[1]]
#> [[1]]
#>      [,1] [,2] [,3]
#> [1,]    0    0    0
#> [2,]    0    0    1
#> [3,]    0    1    1
#> [4,]    1    2    2
#> [5,]    1    2    3
#> [6,]    1    3    2
#> [7,]    0    0    0
#> 
#> attr(,"class")
#> [1] "XYZ"     "POLYGON" "sfg"    
#> 
#> attr(,"n_empty")
#> [1] 0
#> attr(,"crs")
#> $input
#> [1] NA
#> 
#> $wkt
#> [1] NA
#> 
#> attr(,"class")
#> [1] "crs"
#> attr(,"class")
#> [1] "sfc_POLYGON" "sfc"        
#> attr(,"precision")
#> [1] 0
#> attr(,"bbox")
#> xmin ymin xmax ymax 
#>    0    0    1    3 
#> attr(,"class")
#> [1] "bbox"
#> attr(,"z_range")
#> zmin zmax 
#>    0    3 
#> attr(,"class")
#> [1] "z_range"
sfc_polygon( obj = m, x = 2, y = 3, linestring_id = 1, polygon_id = 1 )
#> [[1]]
#> [[1]]
#>      [,1] [,2]
#> [1,]    0    0
#> [2,]    0    1
#> [3,]    1    1
#> [4,]    0    0
#> 
#> attr(,"class")
#> [1] "XY"      "POLYGON" "sfg"    
#> 
#> [[2]]
#> [[1]]
#>      [,1] [,2]
#> [1,]    2    2
#> [2,]    2    3
#> [3,]    3    2
#> [4,]    2    2
#> 
#> attr(,"class")
#> [1] "XY"      "POLYGON" "sfg"    
#> 
#> attr(,"n_empty")
#> [1] 0
#> attr(,"crs")
#> $input
#> [1] NA
#> 
#> $wkt
#> [1] NA
#> 
#> attr(,"class")
#> [1] "crs"
#> attr(,"class")
#> [1] "sfc_POLYGON" "sfc"        
#> attr(,"precision")
#> [1] 0
#> attr(,"bbox")
#> xmin ymin xmax ymax 
#>    0    0    3    3 
#> attr(,"class")
#> [1] "bbox"

df <- data.frame(
  ml_id = c(1,1,1,1,1,1,1,1,1,2,2,2,2,2,2)
  , l_id = c(1,1,1,2,2,2,3,3,3,1,1,1,2,2,2)
  , x = rnorm(15)
  , y = rnorm(15)
  , z = rnorm(15)
  , m = rnorm(15)
)

sfc_polygon( obj = df, x = "x", y = "y")
#> [[1]]
#> [[1]]
#>              [,1]       [,2]
#>  [1,]  0.97404216 -1.0163435
#>  [2,]  0.59200126 -2.0335729
#>  [3,]  0.69265123  1.9513629
#>  [4,]  1.68365211  0.5986378
#>  [5,]  0.09022128 -0.5432657
#>  [6,] -0.18453604  0.3313667
#>  [7,]  1.34200044  0.6343972
#>  [8,]  1.99842837  1.0873430
#>  [9,]  0.46969500 -0.7862765
#> [10,]  0.07450954  0.7361263
#> [11,]  1.39107390 -0.9106159
#> [12,]  0.28240707  1.3501987
#> [13,] -0.81951994 -0.1662701
#> [14,] -0.35337679  1.0275787
#> [15,] -0.85241000 -0.7337706
#> [16,]  0.97404216 -1.0163435
#> 
#> attr(,"class")
#> [1] "XY"      "POLYGON" "sfg"    
#> 
#> attr(,"n_empty")
#> [1] 0
#> attr(,"crs")
#> $input
#> [1] NA
#> 
#> $wkt
#> [1] NA
#> 
#> attr(,"class")
#> [1] "crs"
#> attr(,"class")
#> [1] "sfc_POLYGON" "sfc"        
#> attr(,"precision")
#> [1] 0
#> attr(,"bbox")
#>      xmin      ymin      xmax      ymax 
#> -0.852410 -2.033573  1.998428  1.951363 
#> attr(,"class")
#> [1] "bbox"
sfc_polygon( obj = df, x = "x", y = "y", z = "z")
#> [[1]]
#> [[1]]
#>              [,1]       [,2]        [,3]
#>  [1,]  0.97404216 -1.0163435  1.40382905
#>  [2,]  0.59200126 -2.0335729 -0.16899458
#>  [3,]  0.69265123  1.9513629 -0.02408085
#>  [4,]  1.68365211  0.5986378  0.37595235
#>  [5,]  0.09022128 -0.5432657 -0.81337402
#>  [6,] -0.18453604  0.3313667 -0.09841843
#>  [7,]  1.34200044  0.6343972 -1.17695429
#>  [8,]  1.99842837  1.0873430 -1.31011516
#>  [9,]  0.46969500 -0.7862765 -0.52482463
#> [10,]  0.07450954  0.7361263 -0.23654602
#> [11,]  1.39107390 -0.9106159  0.69456312
#> [12,]  0.28240707  1.3501987  0.17220298
#> [13,] -0.81951994 -0.1662701 -0.13432332
#> [14,] -0.35337679  1.0275787  0.94221169
#> [15,] -0.85241000 -0.7337706  0.83809310
#> [16,]  0.97404216 -1.0163435  1.40382905
#> 
#> attr(,"class")
#> [1] "XYZ"     "POLYGON" "sfg"    
#> 
#> attr(,"n_empty")
#> [1] 0
#> attr(,"crs")
#> $input
#> [1] NA
#> 
#> $wkt
#> [1] NA
#> 
#> attr(,"class")
#> [1] "crs"
#> attr(,"class")
#> [1] "sfc_POLYGON" "sfc"        
#> attr(,"precision")
#> [1] 0
#> attr(,"bbox")
#>      xmin      ymin      xmax      ymax 
#> -0.852410 -2.033573  1.998428  1.951363 
#> attr(,"class")
#> [1] "bbox"
#> attr(,"z_range")
#>      zmin      zmax 
#> -1.310115  1.403829 
#> attr(,"class")
#> [1] "z_range"
sfc_polygon( obj = df, x = "x", y = "y", z = "z", m = "m")
#> [[1]]
#> [[1]]
#>              [,1]       [,2]        [,3]        [,4]
#>  [1,]  0.97404216 -1.0163435  1.40382905 -0.45614827
#>  [2,]  0.59200126 -2.0335729 -0.16899458 -0.41937274
#>  [3,]  0.69265123  1.9513629 -0.02408085 -1.00036072
#>  [4,]  1.68365211  0.5986378  0.37595235 -0.59279923
#>  [5,]  0.09022128 -0.5432657 -0.81337402 -1.15050026
#>  [6,] -0.18453604  0.3313667 -0.09841843 -0.23775783
#>  [7,]  1.34200044  0.6343972 -1.17695429  0.61859922
#>  [8,]  1.99842837  1.0873430 -1.31011516  0.02170148
#>  [9,]  0.46969500 -0.7862765 -0.52482463 -0.21883655
#> [10,]  0.07450954  0.7361263 -0.23654602  0.07283977
#> [11,]  1.39107390 -0.9106159  0.69456312  1.14849860
#> [12,]  0.28240707  1.3501987  0.17220298 -1.02215064
#> [13,] -0.81951994 -0.1662701 -0.13432332  0.60401171
#> [14,] -0.35337679  1.0275787  0.94221169  2.86740940
#> [15,] -0.85241000 -0.7337706  0.83809310 -0.99049843
#> [16,]  0.97404216 -1.0163435  1.40382905 -0.45614827
#> 
#> attr(,"class")
#> [1] "XYZM"    "POLYGON" "sfg"    
#> 
#> attr(,"n_empty")
#> [1] 0
#> attr(,"crs")
#> $input
#> [1] NA
#> 
#> $wkt
#> [1] NA
#> 
#> attr(,"class")
#> [1] "crs"
#> attr(,"class")
#> [1] "sfc_POLYGON" "sfc"        
#> attr(,"precision")
#> [1] 0
#> attr(,"bbox")
#>      xmin      ymin      xmax      ymax 
#> -0.852410 -2.033573  1.998428  1.951363 
#> attr(,"class")
#> [1] "bbox"
#> attr(,"z_range")
#>      zmin      zmax 
#> -1.310115  1.403829 
#> attr(,"class")
#> [1] "z_range"
#> attr(,"m_range")
#>      mmin      mmax 
#> -1.150500  2.867409 
#> attr(,"class")
#> [1] "m_range"

sfc_polygon( obj = df, x = 2, y = 3)
#> [[1]]
#> [[1]]
#>       [,1]        [,2]
#>  [1,]    1  0.97404216
#>  [2,]    1  0.59200126
#>  [3,]    1  0.69265123
#>  [4,]    2  1.68365211
#>  [5,]    2  0.09022128
#>  [6,]    2 -0.18453604
#>  [7,]    3  1.34200044
#>  [8,]    3  1.99842837
#>  [9,]    3  0.46969500
#> [10,]    1  0.07450954
#> [11,]    1  1.39107390
#> [12,]    1  0.28240707
#> [13,]    2 -0.81951994
#> [14,]    2 -0.35337679
#> [15,]    2 -0.85241000
#> [16,]    1  0.97404216
#> 
#> attr(,"class")
#> [1] "XY"      "POLYGON" "sfg"    
#> 
#> attr(,"n_empty")
#> [1] 0
#> attr(,"crs")
#> $input
#> [1] NA
#> 
#> $wkt
#> [1] NA
#> 
#> attr(,"class")
#> [1] "crs"
#> attr(,"class")
#> [1] "sfc_POLYGON" "sfc"        
#> attr(,"precision")
#> [1] 0
#> attr(,"bbox")
#>      xmin      ymin      xmax      ymax 
#>  1.000000 -0.852410  3.000000  1.998428 
#> attr(,"class")
#> [1] "bbox"
sfc_polygon( obj = df, x = 2, y = 3, z = 4)
#> [[1]]
#> [[1]]
#>       [,1]        [,2]       [,3]
#>  [1,]    1  0.97404216 -1.0163435
#>  [2,]    1  0.59200126 -2.0335729
#>  [3,]    1  0.69265123  1.9513629
#>  [4,]    2  1.68365211  0.5986378
#>  [5,]    2  0.09022128 -0.5432657
#>  [6,]    2 -0.18453604  0.3313667
#>  [7,]    3  1.34200044  0.6343972
#>  [8,]    3  1.99842837  1.0873430
#>  [9,]    3  0.46969500 -0.7862765
#> [10,]    1  0.07450954  0.7361263
#> [11,]    1  1.39107390 -0.9106159
#> [12,]    1  0.28240707  1.3501987
#> [13,]    2 -0.81951994 -0.1662701
#> [14,]    2 -0.35337679  1.0275787
#> [15,]    2 -0.85241000 -0.7337706
#> [16,]    1  0.97404216 -1.0163435
#> 
#> attr(,"class")
#> [1] "XYZ"     "POLYGON" "sfg"    
#> 
#> attr(,"n_empty")
#> [1] 0
#> attr(,"crs")
#> $input
#> [1] NA
#> 
#> $wkt
#> [1] NA
#> 
#> attr(,"class")
#> [1] "crs"
#> attr(,"class")
#> [1] "sfc_POLYGON" "sfc"        
#> attr(,"precision")
#> [1] 0
#> attr(,"bbox")
#>      xmin      ymin      xmax      ymax 
#>  1.000000 -0.852410  3.000000  1.998428 
#> attr(,"class")
#> [1] "bbox"
#> attr(,"z_range")
#>      zmin      zmax 
#> -2.033573  1.951363 
#> attr(,"class")
#> [1] "z_range"
sfc_polygon( obj = df, x = 2, y = 3, z = 4, m = 5)
#> [[1]]
#> [[1]]
#>       [,1]        [,2]       [,3]        [,4]
#>  [1,]    1  0.97404216 -1.0163435  1.40382905
#>  [2,]    1  0.59200126 -2.0335729 -0.16899458
#>  [3,]    1  0.69265123  1.9513629 -0.02408085
#>  [4,]    2  1.68365211  0.5986378  0.37595235
#>  [5,]    2  0.09022128 -0.5432657 -0.81337402
#>  [6,]    2 -0.18453604  0.3313667 -0.09841843
#>  [7,]    3  1.34200044  0.6343972 -1.17695429
#>  [8,]    3  1.99842837  1.0873430 -1.31011516
#>  [9,]    3  0.46969500 -0.7862765 -0.52482463
#> [10,]    1  0.07450954  0.7361263 -0.23654602
#> [11,]    1  1.39107390 -0.9106159  0.69456312
#> [12,]    1  0.28240707  1.3501987  0.17220298
#> [13,]    2 -0.81951994 -0.1662701 -0.13432332
#> [14,]    2 -0.35337679  1.0275787  0.94221169
#> [15,]    2 -0.85241000 -0.7337706  0.83809310
#> [16,]    1  0.97404216 -1.0163435  1.40382905
#> 
#> attr(,"class")
#> [1] "XYZM"    "POLYGON" "sfg"    
#> 
#> attr(,"n_empty")
#> [1] 0
#> attr(,"crs")
#> $input
#> [1] NA
#> 
#> $wkt
#> [1] NA
#> 
#> attr(,"class")
#> [1] "crs"
#> attr(,"class")
#> [1] "sfc_POLYGON" "sfc"        
#> attr(,"precision")
#> [1] 0
#> attr(,"bbox")
#>      xmin      ymin      xmax      ymax 
#>  1.000000 -0.852410  3.000000  1.998428 
#> attr(,"class")
#> [1] "bbox"
#> attr(,"z_range")
#>      zmin      zmax 
#> -2.033573  1.951363 
#> attr(,"class")
#> [1] "z_range"
#> attr(,"m_range")
#>      mmin      mmax 
#> -1.310115  1.403829 
#> attr(,"class")
#> [1] "m_range"

sfc_polygon( obj = df, polygon_id = "ml_id", linestring_id = "l_id" )
#> [[1]]
#> [[1]]
#>           [,1]      [,2]        [,3]       [,4]
#> [1,] 0.9740422 -1.016343  1.40382905 -0.4561483
#> [2,] 0.5920013 -2.033573 -0.16899458 -0.4193727
#> [3,] 0.6926512  1.951363 -0.02408085 -1.0003607
#> [4,] 0.9740422 -1.016343  1.40382905 -0.4561483
#> 
#> [[2]]
#>             [,1]       [,2]        [,3]       [,4]
#> [1,]  1.68365211  0.5986378  0.37595235 -0.5927992
#> [2,]  0.09022128 -0.5432657 -0.81337402 -1.1505003
#> [3,] -0.18453604  0.3313667 -0.09841843 -0.2377578
#> [4,]  1.68365211  0.5986378  0.37595235 -0.5927992
#> 
#> [[3]]
#>          [,1]       [,2]       [,3]        [,4]
#> [1,] 1.342000  0.6343972 -1.1769543  0.61859922
#> [2,] 1.998428  1.0873430 -1.3101152  0.02170148
#> [3,] 0.469695 -0.7862765 -0.5248246 -0.21883655
#> [4,] 1.342000  0.6343972 -1.1769543  0.61859922
#> 
#> attr(,"class")
#> [1] "XYZM"    "POLYGON" "sfg"    
#> 
#> [[2]]
#> [[1]]
#>            [,1]       [,2]       [,3]        [,4]
#> [1,] 0.07450954  0.7361263 -0.2365460  0.07283977
#> [2,] 1.39107390 -0.9106159  0.6945631  1.14849860
#> [3,] 0.28240707  1.3501987  0.1722030 -1.02215064
#> [4,] 0.07450954  0.7361263 -0.2365460  0.07283977
#> 
#> [[2]]
#>            [,1]       [,2]       [,3]       [,4]
#> [1,] -0.8195199 -0.1662701 -0.1343233  0.6040117
#> [2,] -0.3533768  1.0275787  0.9422117  2.8674094
#> [3,] -0.8524100 -0.7337706  0.8380931 -0.9904984
#> [4,] -0.8195199 -0.1662701 -0.1343233  0.6040117
#> 
#> attr(,"class")
#> [1] "XYZM"    "POLYGON" "sfg"    
#> 
#> attr(,"n_empty")
#> [1] 0
#> attr(,"crs")
#> $input
#> [1] NA
#> 
#> $wkt
#> [1] NA
#> 
#> attr(,"class")
#> [1] "crs"
#> attr(,"class")
#> [1] "sfc_POLYGON" "sfc"        
#> attr(,"precision")
#> [1] 0
#> attr(,"bbox")
#>      xmin      ymin      xmax      ymax 
#> -0.852410 -2.033573  1.998428  1.951363 
#> attr(,"class")
#> [1] "bbox"
#> attr(,"z_range")
#>      zmin      zmax 
#> -1.310115  1.403829 
#> attr(,"class")
#> [1] "z_range"
#> attr(,"m_range")
#>      mmin      mmax 
#> -1.150500  2.867409 
#> attr(,"class")
#> [1] "m_range"
sfc_polygon( obj = df, polygon_id = 1, linestring_id = 2 )
#> [[1]]
#> [[1]]
#>           [,1]      [,2]        [,3]       [,4]
#> [1,] 0.9740422 -1.016343  1.40382905 -0.4561483
#> [2,] 0.5920013 -2.033573 -0.16899458 -0.4193727
#> [3,] 0.6926512  1.951363 -0.02408085 -1.0003607
#> [4,] 0.9740422 -1.016343  1.40382905 -0.4561483
#> 
#> [[2]]
#>             [,1]       [,2]        [,3]       [,4]
#> [1,]  1.68365211  0.5986378  0.37595235 -0.5927992
#> [2,]  0.09022128 -0.5432657 -0.81337402 -1.1505003
#> [3,] -0.18453604  0.3313667 -0.09841843 -0.2377578
#> [4,]  1.68365211  0.5986378  0.37595235 -0.5927992
#> 
#> [[3]]
#>          [,1]       [,2]       [,3]        [,4]
#> [1,] 1.342000  0.6343972 -1.1769543  0.61859922
#> [2,] 1.998428  1.0873430 -1.3101152  0.02170148
#> [3,] 0.469695 -0.7862765 -0.5248246 -0.21883655
#> [4,] 1.342000  0.6343972 -1.1769543  0.61859922
#> 
#> attr(,"class")
#> [1] "XYZM"    "POLYGON" "sfg"    
#> 
#> [[2]]
#> [[1]]
#>            [,1]       [,2]       [,3]        [,4]
#> [1,] 0.07450954  0.7361263 -0.2365460  0.07283977
#> [2,] 1.39107390 -0.9106159  0.6945631  1.14849860
#> [3,] 0.28240707  1.3501987  0.1722030 -1.02215064
#> [4,] 0.07450954  0.7361263 -0.2365460  0.07283977
#> 
#> [[2]]
#>            [,1]       [,2]       [,3]       [,4]
#> [1,] -0.8195199 -0.1662701 -0.1343233  0.6040117
#> [2,] -0.3533768  1.0275787  0.9422117  2.8674094
#> [3,] -0.8524100 -0.7337706  0.8380931 -0.9904984
#> [4,] -0.8195199 -0.1662701 -0.1343233  0.6040117
#> 
#> attr(,"class")
#> [1] "XYZM"    "POLYGON" "sfg"    
#> 
#> attr(,"n_empty")
#> [1] 0
#> attr(,"crs")
#> $input
#> [1] NA
#> 
#> $wkt
#> [1] NA
#> 
#> attr(,"class")
#> [1] "crs"
#> attr(,"class")
#> [1] "sfc_POLYGON" "sfc"        
#> attr(,"precision")
#> [1] 0
#> attr(,"bbox")
#>      xmin      ymin      xmax      ymax 
#> -0.852410 -2.033573  1.998428  1.951363 
#> attr(,"class")
#> [1] "bbox"
#> attr(,"z_range")
#>      zmin      zmax 
#> -1.310115  1.403829 
#> attr(,"class")
#> [1] "z_range"
#> attr(,"m_range")
#>      mmin      mmax 
#> -1.150500  2.867409 
#> attr(,"class")
#> [1] "m_range"