sfc POLYGON

constructs an sfc of POLYGON objects

sfc_polygon(
  obj = NULL,
  x = NULL,
  y = NULL,
  z = NULL,
  m = NULL,
  polygon_id = NULL,
  linestring_id = NULL,
  close = TRUE
)

Arguments

obj

sorted matrix or data.frame

x

x geometry column

y

y geometry column

z

z geometry column

m

m geometry column

polygon_id

column of ids for polygons

linestring_id

column of ids for lines (within polygons)

close

logical indicating whether polygons should be closed. If TRUE, all polygons will be checked and force closed if possible

Value

sfc object of POLYGON geometries

notes

sfheaders functions do not perform any validity checks on the geometries. Nor do they set Coordinate Reference Systems, EPSG, PROJ4 or precision attributes.

The data.frame and matrices you send into the sfheader functions must be ordered.

Examples

m <- matrix(c(0,0,0,0,1,1), ncol = 2 )
sfc_polygon( m )
#> [[1]]
#> [[1]]
#>      [,1] [,2]
#> [1,]    0    0
#> [2,]    0    1
#> [3,]    0    1
#> [4,]    0    0
#> 
#> attr(,"class")
#> [1] "XY"      "POLYGON" "sfg"    
#> 
#> attr(,"n_empty")
#> [1] 0
#> attr(,"crs")
#> $input
#> [1] NA
#> 
#> $wkt
#> [1] NA
#> 
#> attr(,"class")
#> [1] "crs"
#> attr(,"class")
#> [1] "sfc_POLYGON" "sfc"        
#> attr(,"precision")
#> [1] 0
#> attr(,"bbox")
#> xmin ymin xmax ymax 
#>    0    0    0    1 
#> attr(,"class")
#> [1] "bbox"

m <- matrix(c(0,0,0,0,0,1,0,1,1,1,2,2,1,2,3,1,3,2), ncol = 3, byrow = TRUE)
sfc_polygon( obj = m )
#> [[1]]
#> [[1]]
#>      [,1] [,2] [,3]
#> [1,]    0    0    0
#> [2,]    0    0    1
#> [3,]    0    1    1
#> [4,]    1    2    2
#> [5,]    1    2    3
#> [6,]    1    3    2
#> [7,]    0    0    0
#> 
#> attr(,"class")
#> [1] "XYZ"     "POLYGON" "sfg"    
#> 
#> attr(,"n_empty")
#> [1] 0
#> attr(,"crs")
#> $input
#> [1] NA
#> 
#> $wkt
#> [1] NA
#> 
#> attr(,"class")
#> [1] "crs"
#> attr(,"class")
#> [1] "sfc_POLYGON" "sfc"        
#> attr(,"precision")
#> [1] 0
#> attr(,"bbox")
#> xmin ymin xmax ymax 
#>    0    0    1    3 
#> attr(,"class")
#> [1] "bbox"
#> attr(,"z_range")
#> zmin zmax 
#>    0    3 
#> attr(,"class")
#> [1] "z_range"
sfc_polygon( obj = m, polygon_id = 1 )
#> [[1]]
#> [[1]]
#>      [,1] [,2]
#> [1,]    0    0
#> [2,]    0    1
#> [3,]    1    1
#> [4,]    0    0
#> 
#> attr(,"class")
#> [1] "XY"      "POLYGON" "sfg"    
#> 
#> [[2]]
#> [[1]]
#>      [,1] [,2]
#> [1,]    2    2
#> [2,]    2    3
#> [3,]    3    2
#> [4,]    2    2
#> 
#> attr(,"class")
#> [1] "XY"      "POLYGON" "sfg"    
#> 
#> attr(,"n_empty")
#> [1] 0
#> attr(,"crs")
#> $input
#> [1] NA
#> 
#> $wkt
#> [1] NA
#> 
#> attr(,"class")
#> [1] "crs"
#> attr(,"class")
#> [1] "sfc_POLYGON" "sfc"        
#> attr(,"precision")
#> [1] 0
#> attr(,"bbox")
#> xmin ymin xmax ymax 
#>    0    0    3    3 
#> attr(,"class")
#> [1] "bbox"
sfc_polygon( obj = m, linestring_id = 1 )
#> [[1]]
#> [[1]]
#>      [,1] [,2]
#> [1,]    0    0
#> [2,]    0    1
#> [3,]    1    1
#> [4,]    0    0
#> 
#> [[2]]
#>      [,1] [,2]
#> [1,]    2    2
#> [2,]    2    3
#> [3,]    3    2
#> [4,]    2    2
#> 
#> attr(,"class")
#> [1] "XY"      "POLYGON" "sfg"    
#> 
#> attr(,"n_empty")
#> [1] 0
#> attr(,"crs")
#> $input
#> [1] NA
#> 
#> $wkt
#> [1] NA
#> 
#> attr(,"class")
#> [1] "crs"
#> attr(,"class")
#> [1] "sfc_POLYGON" "sfc"        
#> attr(,"precision")
#> [1] 0
#> attr(,"bbox")
#> xmin ymin xmax ymax 
#>    0    0    3    3 
#> attr(,"class")
#> [1] "bbox"

sfc_polygon( obj = m, linestring_id = 1, polygon_id = 1 )
#> [[1]]
#> [[1]]
#>      [,1] [,2]
#> [1,]    0    0
#> [2,]    0    1
#> [3,]    1    1
#> [4,]    0    0
#> 
#> attr(,"class")
#> [1] "XY"      "POLYGON" "sfg"    
#> 
#> [[2]]
#> [[1]]
#>      [,1] [,2]
#> [1,]    2    2
#> [2,]    2    3
#> [3,]    3    2
#> [4,]    2    2
#> 
#> attr(,"class")
#> [1] "XY"      "POLYGON" "sfg"    
#> 
#> attr(,"n_empty")
#> [1] 0
#> attr(,"crs")
#> $input
#> [1] NA
#> 
#> $wkt
#> [1] NA
#> 
#> attr(,"class")
#> [1] "crs"
#> attr(,"class")
#> [1] "sfc_POLYGON" "sfc"        
#> attr(,"precision")
#> [1] 0
#> attr(,"bbox")
#> xmin ymin xmax ymax 
#>    0    0    3    3 
#> attr(,"class")
#> [1] "bbox"

sfc_polygon( obj = m, x = 2, y = 3 )
#> [[1]]
#> [[1]]
#>      [,1] [,2]
#> [1,]    0    0
#> [2,]    0    1
#> [3,]    1    1
#> [4,]    2    2
#> [5,]    2    3
#> [6,]    3    2
#> [7,]    0    0
#> 
#> attr(,"class")
#> [1] "XY"      "POLYGON" "sfg"    
#> 
#> attr(,"n_empty")
#> [1] 0
#> attr(,"crs")
#> $input
#> [1] NA
#> 
#> $wkt
#> [1] NA
#> 
#> attr(,"class")
#> [1] "crs"
#> attr(,"class")
#> [1] "sfc_POLYGON" "sfc"        
#> attr(,"precision")
#> [1] 0
#> attr(,"bbox")
#> xmin ymin xmax ymax 
#>    0    0    3    3 
#> attr(,"class")
#> [1] "bbox"
sfc_polygon( obj = m, x = 1, y = 2, z = 3 )
#> [[1]]
#> [[1]]
#>      [,1] [,2] [,3]
#> [1,]    0    0    0
#> [2,]    0    0    1
#> [3,]    0    1    1
#> [4,]    1    2    2
#> [5,]    1    2    3
#> [6,]    1    3    2
#> [7,]    0    0    0
#> 
#> attr(,"class")
#> [1] "XYZ"     "POLYGON" "sfg"    
#> 
#> attr(,"n_empty")
#> [1] 0
#> attr(,"crs")
#> $input
#> [1] NA
#> 
#> $wkt
#> [1] NA
#> 
#> attr(,"class")
#> [1] "crs"
#> attr(,"class")
#> [1] "sfc_POLYGON" "sfc"        
#> attr(,"precision")
#> [1] 0
#> attr(,"bbox")
#> xmin ymin xmax ymax 
#>    0    0    1    3 
#> attr(,"class")
#> [1] "bbox"
#> attr(,"z_range")
#> zmin zmax 
#>    0    3 
#> attr(,"class")
#> [1] "z_range"
sfc_polygon( obj = m, x = 2, y = 3, linestring_id = 1, polygon_id = 1 )
#> [[1]]
#> [[1]]
#>      [,1] [,2]
#> [1,]    0    0
#> [2,]    0    1
#> [3,]    1    1
#> [4,]    0    0
#> 
#> attr(,"class")
#> [1] "XY"      "POLYGON" "sfg"    
#> 
#> [[2]]
#> [[1]]
#>      [,1] [,2]
#> [1,]    2    2
#> [2,]    2    3
#> [3,]    3    2
#> [4,]    2    2
#> 
#> attr(,"class")
#> [1] "XY"      "POLYGON" "sfg"    
#> 
#> attr(,"n_empty")
#> [1] 0
#> attr(,"crs")
#> $input
#> [1] NA
#> 
#> $wkt
#> [1] NA
#> 
#> attr(,"class")
#> [1] "crs"
#> attr(,"class")
#> [1] "sfc_POLYGON" "sfc"        
#> attr(,"precision")
#> [1] 0
#> attr(,"bbox")
#> xmin ymin xmax ymax 
#>    0    0    3    3 
#> attr(,"class")
#> [1] "bbox"

df <- data.frame(
  ml_id = c(1,1,1,1,1,1,1,1,1,2,2,2,2,2,2)
  , l_id = c(1,1,1,2,2,2,3,3,3,1,1,1,2,2,2)
  , x = rnorm(15)
  , y = rnorm(15)
  , z = rnorm(15)
  , m = rnorm(15)
)

sfc_polygon( obj = df, x = "x", y = "y")
#> [[1]]
#> [[1]]
#>             [,1]         [,2]
#>  [1,] -0.5643478  0.040552373
#>  [2,] -1.1299415  1.623137689
#>  [3,]  2.0278518  0.899184460
#>  [4,]  0.3706005  0.737795252
#>  [5,] -0.9224405  0.129568237
#>  [6,]  2.9812062 -1.181236223
#>  [7,] -1.0445528 -0.299830169
#>  [8,]  0.3332412  0.634823653
#>  [9,] -0.3115026  0.266269989
#> [10,]  1.2225446  0.009578391
#> [11,]  1.4146348  1.964497760
#> [12,] -0.8663306 -1.522624977
#> [13,] -0.9129576  0.860824506
#> [14,]  0.2621673 -0.445331473
#> [15,]  0.2799509  1.561928366
#> [16,] -0.5643478  0.040552373
#> 
#> attr(,"class")
#> [1] "XY"      "POLYGON" "sfg"    
#> 
#> attr(,"n_empty")
#> [1] 0
#> attr(,"crs")
#> $input
#> [1] NA
#> 
#> $wkt
#> [1] NA
#> 
#> attr(,"class")
#> [1] "crs"
#> attr(,"class")
#> [1] "sfc_POLYGON" "sfc"        
#> attr(,"precision")
#> [1] 0
#> attr(,"bbox")
#>      xmin      ymin      xmax      ymax 
#> -1.129942 -1.522625  2.981206  1.964498 
#> attr(,"class")
#> [1] "bbox"
sfc_polygon( obj = df, x = "x", y = "y", z = "z")
#> [[1]]
#> [[1]]
#>             [,1]         [,2]        [,3]
#>  [1,] -0.5643478  0.040552373  0.06720844
#>  [2,] -1.1299415  1.623137689  0.24222636
#>  [3,]  2.0278518  0.899184460  0.79023216
#>  [4,]  0.3706005  0.737795252 -2.53517176
#>  [5,] -0.9224405  0.129568237 -0.50677042
#>  [6,]  2.9812062 -1.181236223  0.93565488
#>  [7,] -1.0445528 -0.299830169  2.16112782
#>  [8,]  0.3332412  0.634823653 -0.74614209
#>  [9,] -0.3115026  0.266269989 -0.44693629
#> [10,]  1.2225446  0.009578391  1.53780562
#> [11,]  1.4146348  1.964497760  0.45707173
#> [12,] -0.8663306 -1.522624977 -1.96984666
#> [13,] -0.9129576  0.860824506 -0.30566547
#> [14,]  0.2621673 -0.445331473 -0.41026820
#> [15,]  0.2799509  1.561928366  0.75219055
#> [16,] -0.5643478  0.040552373  0.06720844
#> 
#> attr(,"class")
#> [1] "XYZ"     "POLYGON" "sfg"    
#> 
#> attr(,"n_empty")
#> [1] 0
#> attr(,"crs")
#> $input
#> [1] NA
#> 
#> $wkt
#> [1] NA
#> 
#> attr(,"class")
#> [1] "crs"
#> attr(,"class")
#> [1] "sfc_POLYGON" "sfc"        
#> attr(,"precision")
#> [1] 0
#> attr(,"bbox")
#>      xmin      ymin      xmax      ymax 
#> -1.129942 -1.522625  2.981206  1.964498 
#> attr(,"class")
#> [1] "bbox"
#> attr(,"z_range")
#>      zmin      zmax 
#> -2.535172  2.161128 
#> attr(,"class")
#> [1] "z_range"
sfc_polygon( obj = df, x = "x", y = "y", z = "z", m = "m")
#> [[1]]
#> [[1]]
#>             [,1]         [,2]        [,3]        [,4]
#>  [1,] -0.5643478  0.040552373  0.06720844 -1.13929352
#>  [2,] -1.1299415  1.623137689  0.24222636  0.04390924
#>  [3,]  2.0278518  0.899184460  0.79023216  0.20026606
#>  [4,]  0.3706005  0.737795252 -2.53517176  1.21416710
#>  [5,] -0.9224405  0.129568237 -0.50677042 -0.28495673
#>  [6,]  2.9812062 -1.181236223  0.93565488 -0.47316332
#>  [7,] -1.0445528 -0.299830169  2.16112782 -1.55767613
#>  [8,]  0.3332412  0.634823653 -0.74614209  0.49698184
#>  [9,] -0.3115026  0.266269989 -0.44693629  0.76137667
#> [10,]  1.2225446  0.009578391  1.53780562  0.04805000
#> [11,]  1.4146348  1.964497760  0.45707173 -0.75543119
#> [12,] -0.8663306 -1.522624977 -1.96984666 -1.20529915
#> [13,] -0.9129576  0.860824506 -0.30566547 -0.53910836
#> [14,]  0.2621673 -0.445331473 -0.41026820  0.59312869
#> [15,]  0.2799509  1.561928366  0.75219055  0.34333361
#> [16,] -0.5643478  0.040552373  0.06720844 -1.13929352
#> 
#> attr(,"class")
#> [1] "XYZM"    "POLYGON" "sfg"    
#> 
#> attr(,"n_empty")
#> [1] 0
#> attr(,"crs")
#> $input
#> [1] NA
#> 
#> $wkt
#> [1] NA
#> 
#> attr(,"class")
#> [1] "crs"
#> attr(,"class")
#> [1] "sfc_POLYGON" "sfc"        
#> attr(,"precision")
#> [1] 0
#> attr(,"bbox")
#>      xmin      ymin      xmax      ymax 
#> -1.129942 -1.522625  2.981206  1.964498 
#> attr(,"class")
#> [1] "bbox"
#> attr(,"z_range")
#>      zmin      zmax 
#> -2.535172  2.161128 
#> attr(,"class")
#> [1] "z_range"
#> attr(,"m_range")
#>      mmin      mmax 
#> -1.557676  1.214167 
#> attr(,"class")
#> [1] "m_range"

sfc_polygon( obj = df, x = 2, y = 3)
#> [[1]]
#> [[1]]
#>       [,1]       [,2]
#>  [1,]    1 -0.5643478
#>  [2,]    1 -1.1299415
#>  [3,]    1  2.0278518
#>  [4,]    2  0.3706005
#>  [5,]    2 -0.9224405
#>  [6,]    2  2.9812062
#>  [7,]    3 -1.0445528
#>  [8,]    3  0.3332412
#>  [9,]    3 -0.3115026
#> [10,]    1  1.2225446
#> [11,]    1  1.4146348
#> [12,]    1 -0.8663306
#> [13,]    2 -0.9129576
#> [14,]    2  0.2621673
#> [15,]    2  0.2799509
#> [16,]    1 -0.5643478
#> 
#> attr(,"class")
#> [1] "XY"      "POLYGON" "sfg"    
#> 
#> attr(,"n_empty")
#> [1] 0
#> attr(,"crs")
#> $input
#> [1] NA
#> 
#> $wkt
#> [1] NA
#> 
#> attr(,"class")
#> [1] "crs"
#> attr(,"class")
#> [1] "sfc_POLYGON" "sfc"        
#> attr(,"precision")
#> [1] 0
#> attr(,"bbox")
#>      xmin      ymin      xmax      ymax 
#>  1.000000 -1.129942  3.000000  2.981206 
#> attr(,"class")
#> [1] "bbox"
sfc_polygon( obj = df, x = 2, y = 3, z = 4)
#> [[1]]
#> [[1]]
#>       [,1]       [,2]         [,3]
#>  [1,]    1 -0.5643478  0.040552373
#>  [2,]    1 -1.1299415  1.623137689
#>  [3,]    1  2.0278518  0.899184460
#>  [4,]    2  0.3706005  0.737795252
#>  [5,]    2 -0.9224405  0.129568237
#>  [6,]    2  2.9812062 -1.181236223
#>  [7,]    3 -1.0445528 -0.299830169
#>  [8,]    3  0.3332412  0.634823653
#>  [9,]    3 -0.3115026  0.266269989
#> [10,]    1  1.2225446  0.009578391
#> [11,]    1  1.4146348  1.964497760
#> [12,]    1 -0.8663306 -1.522624977
#> [13,]    2 -0.9129576  0.860824506
#> [14,]    2  0.2621673 -0.445331473
#> [15,]    2  0.2799509  1.561928366
#> [16,]    1 -0.5643478  0.040552373
#> 
#> attr(,"class")
#> [1] "XYZ"     "POLYGON" "sfg"    
#> 
#> attr(,"n_empty")
#> [1] 0
#> attr(,"crs")
#> $input
#> [1] NA
#> 
#> $wkt
#> [1] NA
#> 
#> attr(,"class")
#> [1] "crs"
#> attr(,"class")
#> [1] "sfc_POLYGON" "sfc"        
#> attr(,"precision")
#> [1] 0
#> attr(,"bbox")
#>      xmin      ymin      xmax      ymax 
#>  1.000000 -1.129942  3.000000  2.981206 
#> attr(,"class")
#> [1] "bbox"
#> attr(,"z_range")
#>      zmin      zmax 
#> -1.522625  1.964498 
#> attr(,"class")
#> [1] "z_range"
sfc_polygon( obj = df, x = 2, y = 3, z = 4, m = 5)
#> [[1]]
#> [[1]]
#>       [,1]       [,2]         [,3]        [,4]
#>  [1,]    1 -0.5643478  0.040552373  0.06720844
#>  [2,]    1 -1.1299415  1.623137689  0.24222636
#>  [3,]    1  2.0278518  0.899184460  0.79023216
#>  [4,]    2  0.3706005  0.737795252 -2.53517176
#>  [5,]    2 -0.9224405  0.129568237 -0.50677042
#>  [6,]    2  2.9812062 -1.181236223  0.93565488
#>  [7,]    3 -1.0445528 -0.299830169  2.16112782
#>  [8,]    3  0.3332412  0.634823653 -0.74614209
#>  [9,]    3 -0.3115026  0.266269989 -0.44693629
#> [10,]    1  1.2225446  0.009578391  1.53780562
#> [11,]    1  1.4146348  1.964497760  0.45707173
#> [12,]    1 -0.8663306 -1.522624977 -1.96984666
#> [13,]    2 -0.9129576  0.860824506 -0.30566547
#> [14,]    2  0.2621673 -0.445331473 -0.41026820
#> [15,]    2  0.2799509  1.561928366  0.75219055
#> [16,]    1 -0.5643478  0.040552373  0.06720844
#> 
#> attr(,"class")
#> [1] "XYZM"    "POLYGON" "sfg"    
#> 
#> attr(,"n_empty")
#> [1] 0
#> attr(,"crs")
#> $input
#> [1] NA
#> 
#> $wkt
#> [1] NA
#> 
#> attr(,"class")
#> [1] "crs"
#> attr(,"class")
#> [1] "sfc_POLYGON" "sfc"        
#> attr(,"precision")
#> [1] 0
#> attr(,"bbox")
#>      xmin      ymin      xmax      ymax 
#>  1.000000 -1.129942  3.000000  2.981206 
#> attr(,"class")
#> [1] "bbox"
#> attr(,"z_range")
#>      zmin      zmax 
#> -1.522625  1.964498 
#> attr(,"class")
#> [1] "z_range"
#> attr(,"m_range")
#>      mmin      mmax 
#> -2.535172  2.161128 
#> attr(,"class")
#> [1] "m_range"

sfc_polygon( obj = df, polygon_id = "ml_id", linestring_id = "l_id" )
#> [[1]]
#> [[1]]
#>            [,1]       [,2]       [,3]        [,4]
#> [1,] -0.5643478 0.04055237 0.06720844 -1.13929352
#> [2,] -1.1299415 1.62313769 0.24222636  0.04390924
#> [3,]  2.0278518 0.89918446 0.79023216  0.20026606
#> [4,] -0.5643478 0.04055237 0.06720844 -1.13929352
#> 
#> [[2]]
#>            [,1]       [,2]       [,3]       [,4]
#> [1,]  0.3706005  0.7377953 -2.5351718  1.2141671
#> [2,] -0.9224405  0.1295682 -0.5067704 -0.2849567
#> [3,]  2.9812062 -1.1812362  0.9356549 -0.4731633
#> [4,]  0.3706005  0.7377953 -2.5351718  1.2141671
#> 
#> [[3]]
#>            [,1]       [,2]       [,3]       [,4]
#> [1,] -1.0445528 -0.2998302  2.1611278 -1.5576761
#> [2,]  0.3332412  0.6348237 -0.7461421  0.4969818
#> [3,] -0.3115026  0.2662700 -0.4469363  0.7613767
#> [4,] -1.0445528 -0.2998302  2.1611278 -1.5576761
#> 
#> attr(,"class")
#> [1] "XYZM"    "POLYGON" "sfg"    
#> 
#> [[2]]
#> [[1]]
#>            [,1]         [,2]       [,3]       [,4]
#> [1,]  1.2225446  0.009578391  1.5378056  0.0480500
#> [2,]  1.4146348  1.964497760  0.4570717 -0.7554312
#> [3,] -0.8663306 -1.522624977 -1.9698467 -1.2052992
#> [4,]  1.2225446  0.009578391  1.5378056  0.0480500
#> 
#> [[2]]
#>            [,1]       [,2]       [,3]       [,4]
#> [1,] -0.9129576  0.8608245 -0.3056655 -0.5391084
#> [2,]  0.2621673 -0.4453315 -0.4102682  0.5931287
#> [3,]  0.2799509  1.5619284  0.7521906  0.3433336
#> [4,] -0.9129576  0.8608245 -0.3056655 -0.5391084
#> 
#> attr(,"class")
#> [1] "XYZM"    "POLYGON" "sfg"    
#> 
#> attr(,"n_empty")
#> [1] 0
#> attr(,"crs")
#> $input
#> [1] NA
#> 
#> $wkt
#> [1] NA
#> 
#> attr(,"class")
#> [1] "crs"
#> attr(,"class")
#> [1] "sfc_POLYGON" "sfc"        
#> attr(,"precision")
#> [1] 0
#> attr(,"bbox")
#>      xmin      ymin      xmax      ymax 
#> -1.129942 -1.522625  2.981206  1.964498 
#> attr(,"class")
#> [1] "bbox"
#> attr(,"z_range")
#>      zmin      zmax 
#> -2.535172  2.161128 
#> attr(,"class")
#> [1] "z_range"
#> attr(,"m_range")
#>      mmin      mmax 
#> -1.557676  1.214167 
#> attr(,"class")
#> [1] "m_range"
sfc_polygon( obj = df, polygon_id = 1, linestring_id = 2 )
#> [[1]]
#> [[1]]
#>            [,1]       [,2]       [,3]        [,4]
#> [1,] -0.5643478 0.04055237 0.06720844 -1.13929352
#> [2,] -1.1299415 1.62313769 0.24222636  0.04390924
#> [3,]  2.0278518 0.89918446 0.79023216  0.20026606
#> [4,] -0.5643478 0.04055237 0.06720844 -1.13929352
#> 
#> [[2]]
#>            [,1]       [,2]       [,3]       [,4]
#> [1,]  0.3706005  0.7377953 -2.5351718  1.2141671
#> [2,] -0.9224405  0.1295682 -0.5067704 -0.2849567
#> [3,]  2.9812062 -1.1812362  0.9356549 -0.4731633
#> [4,]  0.3706005  0.7377953 -2.5351718  1.2141671
#> 
#> [[3]]
#>            [,1]       [,2]       [,3]       [,4]
#> [1,] -1.0445528 -0.2998302  2.1611278 -1.5576761
#> [2,]  0.3332412  0.6348237 -0.7461421  0.4969818
#> [3,] -0.3115026  0.2662700 -0.4469363  0.7613767
#> [4,] -1.0445528 -0.2998302  2.1611278 -1.5576761
#> 
#> attr(,"class")
#> [1] "XYZM"    "POLYGON" "sfg"    
#> 
#> [[2]]
#> [[1]]
#>            [,1]         [,2]       [,3]       [,4]
#> [1,]  1.2225446  0.009578391  1.5378056  0.0480500
#> [2,]  1.4146348  1.964497760  0.4570717 -0.7554312
#> [3,] -0.8663306 -1.522624977 -1.9698467 -1.2052992
#> [4,]  1.2225446  0.009578391  1.5378056  0.0480500
#> 
#> [[2]]
#>            [,1]       [,2]       [,3]       [,4]
#> [1,] -0.9129576  0.8608245 -0.3056655 -0.5391084
#> [2,]  0.2621673 -0.4453315 -0.4102682  0.5931287
#> [3,]  0.2799509  1.5619284  0.7521906  0.3433336
#> [4,] -0.9129576  0.8608245 -0.3056655 -0.5391084
#> 
#> attr(,"class")
#> [1] "XYZM"    "POLYGON" "sfg"    
#> 
#> attr(,"n_empty")
#> [1] 0
#> attr(,"crs")
#> $input
#> [1] NA
#> 
#> $wkt
#> [1] NA
#> 
#> attr(,"class")
#> [1] "crs"
#> attr(,"class")
#> [1] "sfc_POLYGON" "sfc"        
#> attr(,"precision")
#> [1] 0
#> attr(,"bbox")
#>      xmin      ymin      xmax      ymax 
#> -1.129942 -1.522625  2.981206  1.964498 
#> attr(,"class")
#> [1] "bbox"
#> attr(,"z_range")
#>      zmin      zmax 
#> -2.535172  2.161128 
#> attr(,"class")
#> [1] "z_range"
#> attr(,"m_range")
#>      mmin      mmax 
#> -1.557676  1.214167 
#> attr(,"class")
#> [1] "m_range"