sfc POLYGON

constructs an sfc of POLYGON objects

sfc_polygon(obj = NULL, x = NULL, y = NULL, z = NULL, m = NULL,
  polygon_id = NULL, linestring_id = NULL)

Arguments

obj
sorted matrix or data.frame

x
x geometry column

y
y geometry column

z
z geometry column

m
m geometry column

polygon_id
column of ids for polygons

linestring_id
column of ids for lines (within polygons)

notes

sfheaders functions do not perform any validity checks on the geometries. Nor do they set Coordinate Reference Systems, EPSG, PROJ4 or precision attributes.

The data.frame and matrices you send into the sfheader functions must be ordered.

Examples

m <- matrix(c(0,0,0,0,1,1), ncol = 3 )
sfc_polygon( m )

#> [[1]]
#> [[1]]
#>      [,1] [,2] [,3]
#> [1,]    0    0    1
#> [2,]    0    0    1
#> 
#> attr(,"class")
#> [1] "XYZ"     "POLYGON" "sfg"    
#> 
#> attr(,"n_empty")
#> [1] 0
#> attr(,"crs")
#> $epsg
#> [1] NA
#> 
#> $proj4string
#> [1] NA
#> 
#> attr(,"class")
#> [1] "crs"
#> attr(,"class")
#> [1] "sfc_POLYGON" "sfc"        
#> attr(,"precision")
#> [1] 0
#> attr(,"bbox")
#> xmin ymin xmax ymax 
#>    0    0    0    0 
#> attr(,"class")
#> [1] "bbox"
#> attr(,"z_range")
#> zmin zmax 
#>    1    1 
#> attr(,"class")
#> [1] "z_range"
#> attr(,"m_range")
#> mmin mmax 
#>   NA   NA 
#> attr(,"class")
#> [1] "m_range"

m <- matrix(c(0,0,0,0,0,1,0,1,1,1,2,2,1,2,3), ncol = 3, byrow = TRUE)
sfc_polygon( obj = m )

#> [[1]]
#> [[1]]
#>      [,1] [,2] [,3]
#> [1,]    0    0    0
#> [2,]    0    0    1
#> [3,]    0    1    1
#> [4,]    1    2    2
#> [5,]    1    2    3
#> 
#> attr(,"class")
#> [1] "XYZ"     "POLYGON" "sfg"    
#> 
#> attr(,"n_empty")
#> [1] 0
#> attr(,"crs")
#> $epsg
#> [1] NA
#> 
#> $proj4string
#> [1] NA
#> 
#> attr(,"class")
#> [1] "crs"
#> attr(,"class")
#> [1] "sfc_POLYGON" "sfc"        
#> attr(,"precision")
#> [1] 0
#> attr(,"bbox")
#> xmin ymin xmax ymax 
#>    0    0    1    2 
#> attr(,"class")
#> [1] "bbox"
#> attr(,"z_range")
#> zmin zmax 
#>    0    3 
#> attr(,"class")
#> [1] "z_range"
#> attr(,"m_range")
#> mmin mmax 
#>   NA   NA 
#> attr(,"class")
#> [1] "m_range"

sfc_polygon( obj = m, polygon_id = 1 )

#> [[1]]
#> [[1]]
#>      [,1] [,2]
#> [1,]    0    0
#> [2,]    0    1
#> [3,]    1    1
#> 
#> attr(,"class")
#> [1] "XY"      "POLYGON" "sfg"    
#> 
#> [[2]]
#> [[1]]
#>      [,1] [,2]
#> [1,]    2    2
#> [2,]    2    3
#> 
#> attr(,"class")
#> [1] "XY"      "POLYGON" "sfg"    
#> 
#> attr(,"n_empty")
#> [1] 0
#> attr(,"crs")
#> $epsg
#> [1] NA
#> 
#> $proj4string
#> [1] NA
#> 
#> attr(,"class")
#> [1] "crs"
#> attr(,"class")
#> [1] "sfc_POLYGON" "sfc"        
#> attr(,"precision")
#> [1] 0
#> attr(,"bbox")
#> xmin ymin xmax ymax 
#>    0    0    2    3 
#> attr(,"class")
#> [1] "bbox"
#> attr(,"z_range")
#> zmin zmax 
#>   NA   NA 
#> attr(,"class")
#> [1] "z_range"
#> attr(,"m_range")
#> mmin mmax 
#>   NA   NA 
#> attr(,"class")
#> [1] "m_range"

sfc_polygon( obj = m, linestring_id = 1 )

#> [[1]]
#> [[1]]
#>      [,1] [,2]
#> [1,]    0    0
#> [2,]    0    1
#> [3,]    1    1
#> 
#> [[2]]
#>      [,1] [,2]
#> [1,]    2    2
#> [2,]    2    3
#> 
#> attr(,"class")
#> [1] "XY"      "POLYGON" "sfg"    
#> 
#> attr(,"n_empty")
#> [1] 0
#> attr(,"crs")
#> $epsg
#> [1] NA
#> 
#> $proj4string
#> [1] NA
#> 
#> attr(,"class")
#> [1] "crs"
#> attr(,"class")
#> [1] "sfc_POLYGON" "sfc"        
#> attr(,"precision")
#> [1] 0
#> attr(,"bbox")
#> xmin ymin xmax ymax 
#>    0    0    2    3 
#> attr(,"class")
#> [1] "bbox"
#> attr(,"z_range")
#> zmin zmax 
#>   NA   NA 
#> attr(,"class")
#> [1] "z_range"
#> attr(,"m_range")
#> mmin mmax 
#>   NA   NA 
#> attr(,"class")
#> [1] "m_range"

sfc_polygon( obj = m, linestring_id = 1, polygon_id = 1 )

#> [[1]]
#> [[1]]
#>      [,1] [,2]
#> [1,]    0    0
#> [2,]    0    1
#> [3,]    1    1
#> 
#> attr(,"class")
#> [1] "XY"      "POLYGON" "sfg"    
#> 
#> [[2]]
#> [[1]]
#>      [,1] [,2]
#> [1,]    2    2
#> [2,]    2    3
#> 
#> attr(,"class")
#> [1] "XY"      "POLYGON" "sfg"    
#> 
#> attr(,"n_empty")
#> [1] 0
#> attr(,"crs")
#> $epsg
#> [1] NA
#> 
#> $proj4string
#> [1] NA
#> 
#> attr(,"class")
#> [1] "crs"
#> attr(,"class")
#> [1] "sfc_POLYGON" "sfc"        
#> attr(,"precision")
#> [1] 0
#> attr(,"bbox")
#> xmin ymin xmax ymax 
#>    0    0    2    3 
#> attr(,"class")
#> [1] "bbox"
#> attr(,"z_range")
#> zmin zmax 
#>   NA   NA 
#> attr(,"class")
#> [1] "z_range"
#> attr(,"m_range")
#> mmin mmax 
#>   NA   NA 
#> attr(,"class")
#> [1] "m_range"

sfc_polygon( obj = m, x = 2, y = 3 )

#> [[1]]
#> [[1]]
#>      [,1] [,2]
#> [1,]    0    0
#> [2,]    0    1
#> [3,]    1    1
#> [4,]    2    2
#> [5,]    2    3
#> 
#> attr(,"class")
#> [1] "XY"      "POLYGON" "sfg"    
#> 
#> attr(,"n_empty")
#> [1] 0
#> attr(,"crs")
#> $epsg
#> [1] NA
#> 
#> $proj4string
#> [1] NA
#> 
#> attr(,"class")
#> [1] "crs"
#> attr(,"class")
#> [1] "sfc_POLYGON" "sfc"        
#> attr(,"precision")
#> [1] 0
#> attr(,"bbox")
#> xmin ymin xmax ymax 
#>    0    0    2    3 
#> attr(,"class")
#> [1] "bbox"
#> attr(,"z_range")
#> zmin zmax 
#>   NA   NA 
#> attr(,"class")
#> [1] "z_range"
#> attr(,"m_range")
#> mmin mmax 
#>   NA   NA 
#> attr(,"class")
#> [1] "m_range"

sfc_polygon( obj = m, x = 1, y = 2, z = 3 )

#> [[1]]
#> [[1]]
#>      [,1] [,2] [,3]
#> [1,]    0    0    0
#> [2,]    0    0    1
#> [3,]    0    1    1
#> [4,]    1    2    2
#> [5,]    1    2    3
#> 
#> attr(,"class")
#> [1] "XYZ"     "POLYGON" "sfg"    
#> 
#> attr(,"n_empty")
#> [1] 0
#> attr(,"crs")
#> $epsg
#> [1] NA
#> 
#> $proj4string
#> [1] NA
#> 
#> attr(,"class")
#> [1] "crs"
#> attr(,"class")
#> [1] "sfc_POLYGON" "sfc"        
#> attr(,"precision")
#> [1] 0
#> attr(,"bbox")
#> xmin ymin xmax ymax 
#>    0    0    1    2 
#> attr(,"class")
#> [1] "bbox"
#> attr(,"z_range")
#> zmin zmax 
#>    0    3 
#> attr(,"class")
#> [1] "z_range"
#> attr(,"m_range")
#> mmin mmax 
#>   NA   NA 
#> attr(,"class")
#> [1] "m_range"

sfc_polygon( obj = m, x = 2, y = 3, linestring_id = 1, polygon_id = 1 )

#> [[1]]
#> [[1]]
#>      [,1] [,2]
#> [1,]    0    0
#> [2,]    0    1
#> [3,]    1    1
#> 
#> attr(,"class")
#> [1] "XY"      "POLYGON" "sfg"    
#> 
#> [[2]]
#> [[1]]
#>      [,1] [,2]
#> [1,]    2    2
#> [2,]    2    3
#> 
#> attr(,"class")
#> [1] "XY"      "POLYGON" "sfg"    
#> 
#> attr(,"n_empty")
#> [1] 0
#> attr(,"crs")
#> $epsg
#> [1] NA
#> 
#> $proj4string
#> [1] NA
#> 
#> attr(,"class")
#> [1] "crs"
#> attr(,"class")
#> [1] "sfc_POLYGON" "sfc"        
#> attr(,"precision")
#> [1] 0
#> attr(,"bbox")
#> xmin ymin xmax ymax 
#>    0    0    2    3 
#> attr(,"class")
#> [1] "bbox"
#> attr(,"z_range")
#> zmin zmax 
#>   NA   NA 
#> attr(,"class")
#> [1] "z_range"
#> attr(,"m_range")
#> mmin mmax 
#>   NA   NA 
#> attr(,"class")
#> [1] "m_range"

df <- data.frame(
  ml_id = c(1,1,1,1,1,1,1,1,2,2,2,2,2)
  , l_id = c(1,1,1,2,2,3,3,3,1,1,1,2,2)
  , x = rnorm(13)
  , y = rnorm(13)
  , z = rnorm(13)
  , m = rnorm(13)
)

sfc_polygon( obj = df, x = "x", y = "y")

#> [[1]]
#> [[1]]
#>              [,1]       [,2]
#>  [1,]  0.03849955  0.5176691
#>  [2,] -0.35638119 -0.1029087
#>  [3,]  0.78284410 -0.9740696
#>  [4,]  0.80441162  1.2706723
#>  [5,] -1.90006082  0.9608648
#>  [6,]  0.93578429  0.7687214
#>  [7,] -0.30905150  1.0359308
#>  [8,]  0.26306668 -0.4738871
#>  [9,] -1.79059186 -1.2753349
#> [10,] -0.78825884 -0.3056207
#> [11,] -1.13302167  2.2117695
#> [12,]  0.36365257 -1.0416684
#> [13,] -0.28588791 -1.1465239
#> 
#> attr(,"class")
#> [1] "XY"      "POLYGON" "sfg"    
#> 
#> attr(,"n_empty")
#> [1] 0
#> attr(,"crs")
#> $epsg
#> [1] NA
#> 
#> $proj4string
#> [1] NA
#> 
#> attr(,"class")
#> [1] "crs"
#> attr(,"class")
#> [1] "sfc_POLYGON" "sfc"        
#> attr(,"precision")
#> [1] 0
#> attr(,"bbox")
#>       xmin       ymin       xmax       ymax 
#> -1.9000608 -1.2753349  0.9357843  2.2117695 
#> attr(,"class")
#> [1] "bbox"
#> attr(,"z_range")
#> zmin zmax 
#>   NA   NA 
#> attr(,"class")
#> [1] "z_range"
#> attr(,"m_range")
#> mmin mmax 
#>   NA   NA 
#> attr(,"class")
#> [1] "m_range"

sfc_polygon( obj = df, x = "x", y = "y", z = "z")

#> [[1]]
#> [[1]]
#>              [,1]       [,2]       [,3]
#>  [1,]  0.03849955  0.5176691 -1.6753273
#>  [2,] -0.35638119 -0.1029087  1.5259387
#>  [3,]  0.78284410 -0.9740696  0.5541855
#>  [4,]  0.80441162  1.2706723  1.9931103
#>  [5,] -1.90006082  0.9608648 -0.1541207
#>  [6,]  0.93578429  0.7687214  2.5644083
#>  [7,] -0.30905150  1.0359308  1.0619991
#>  [8,]  0.26306668 -0.4738871  1.1426949
#>  [9,] -1.79059186 -1.2753349  1.1238388
#> [10,] -0.78825884 -0.3056207 -0.3970015
#> [11,] -1.13302167  2.2117695 -0.8232612
#> [12,]  0.36365257 -1.0416684 -0.5788846
#> [13,] -0.28588791 -1.1465239  1.7637894
#> 
#> attr(,"class")
#> [1] "XYZ"     "POLYGON" "sfg"    
#> 
#> attr(,"n_empty")
#> [1] 0
#> attr(,"crs")
#> $epsg
#> [1] NA
#> 
#> $proj4string
#> [1] NA
#> 
#> attr(,"class")
#> [1] "crs"
#> attr(,"class")
#> [1] "sfc_POLYGON" "sfc"        
#> attr(,"precision")
#> [1] 0
#> attr(,"bbox")
#>       xmin       ymin       xmax       ymax 
#> -1.9000608 -1.2753349  0.9357843  2.2117695 
#> attr(,"class")
#> [1] "bbox"
#> attr(,"z_range")
#> zmin zmax 
#>   NA   NA 
#> attr(,"class")
#> [1] "z_range"
#> attr(,"m_range")
#> mmin mmax 
#>   NA   NA 
#> attr(,"class")
#> [1] "m_range"

sfc_polygon( obj = df, x = "x", y = "y", z = "z", m = "m")

#> [[1]]
#> [[1]]
#>              [,1]       [,2]       [,3]        [,4]
#>  [1,]  0.03849955  0.5176691 -1.6753273  0.13299215
#>  [2,] -0.35638119 -0.1029087  1.5259387  0.37649933
#>  [3,]  0.78284410 -0.9740696  0.5541855  1.13870765
#>  [4,]  0.80441162  1.2706723  1.9931103  1.24126308
#>  [5,] -1.90006082  0.9608648 -0.1541207  0.61209094
#>  [6,]  0.93578429  0.7687214  2.5644083 -0.42938009
#>  [7,] -0.30905150  1.0359308  1.0619991  1.36046133
#>  [8,]  0.26306668 -0.4738871  1.1426949 -0.07085743
#>  [9,] -1.79059186 -1.2753349  1.1238388 -0.27215368
#> [10,] -0.78825884 -0.3056207 -0.3970015 -2.44668003
#> [11,] -1.13302167  2.2117695 -0.8232612  0.06548664
#> [12,]  0.36365257 -1.0416684 -0.5788846 -1.09850890
#> [13,] -0.28588791 -1.1465239  1.7637894 -0.63317818
#> 
#> attr(,"class")
#> [1] "XYZM"    "POLYGON" "sfg"    
#> 
#> attr(,"n_empty")
#> [1] 0
#> attr(,"crs")
#> $epsg
#> [1] NA
#> 
#> $proj4string
#> [1] NA
#> 
#> attr(,"class")
#> [1] "crs"
#> attr(,"class")
#> [1] "sfc_POLYGON" "sfc"        
#> attr(,"precision")
#> [1] 0
#> attr(,"bbox")
#>       xmin       ymin       xmax       ymax 
#> -1.9000608 -1.2753349  0.9357843  2.2117695 
#> attr(,"class")
#> [1] "bbox"
#> attr(,"z_range")
#> zmin zmax 
#>   -2    1 
#> attr(,"class")
#> [1] "z_range"
#> attr(,"m_range")
#> mmin mmax 
#>   -2    1 
#> attr(,"class")
#> [1] "m_range"

sfc_polygon( obj = df, x = 2, y = 3)

#> [[1]]
#> [[1]]
#>       l_id           x
#>  [1,]    1  0.03849955
#>  [2,]    1 -0.35638119
#>  [3,]    1  0.78284410
#>  [4,]    2  0.80441162
#>  [5,]    2 -1.90006082
#>  [6,]    3  0.93578429
#>  [7,]    3 -0.30905150
#>  [8,]    3  0.26306668
#>  [9,]    1 -1.79059186
#> [10,]    1 -0.78825884
#> [11,]    1 -1.13302167
#> [12,]    2  0.36365257
#> [13,]    2 -0.28588791
#> 
#> attr(,"class")
#> [1] "XY"      "POLYGON" "sfg"    
#> 
#> attr(,"n_empty")
#> [1] 0
#> attr(,"crs")
#> $epsg
#> [1] NA
#> 
#> $proj4string
#> [1] NA
#> 
#> attr(,"class")
#> [1] "crs"
#> attr(,"class")
#> [1] "sfc_POLYGON" "sfc"        
#> attr(,"precision")
#> [1] 0
#> attr(,"bbox")
#>       xmin       ymin       xmax       ymax 
#>  1.0000000 -1.9000608  3.0000000  0.9357843 
#> attr(,"class")
#> [1] "bbox"
#> attr(,"z_range")
#> zmin zmax 
#>   NA   NA 
#> attr(,"class")
#> [1] "z_range"
#> attr(,"m_range")
#> mmin mmax 
#>   NA   NA 
#> attr(,"class")
#> [1] "m_range"

sfc_polygon( obj = df, x = 2, y = 3, z = 4)

#> [[1]]
#> [[1]]
#>       l_id           x          y
#>  [1,]    1  0.03849955  0.5176691
#>  [2,]    1 -0.35638119 -0.1029087
#>  [3,]    1  0.78284410 -0.9740696
#>  [4,]    2  0.80441162  1.2706723
#>  [5,]    2 -1.90006082  0.9608648
#>  [6,]    3  0.93578429  0.7687214
#>  [7,]    3 -0.30905150  1.0359308
#>  [8,]    3  0.26306668 -0.4738871
#>  [9,]    1 -1.79059186 -1.2753349
#> [10,]    1 -0.78825884 -0.3056207
#> [11,]    1 -1.13302167  2.2117695
#> [12,]    2  0.36365257 -1.0416684
#> [13,]    2 -0.28588791 -1.1465239
#> 
#> attr(,"class")
#> [1] "XYZ"     "POLYGON" "sfg"    
#> 
#> attr(,"n_empty")
#> [1] 0
#> attr(,"crs")
#> $epsg
#> [1] NA
#> 
#> $proj4string
#> [1] NA
#> 
#> attr(,"class")
#> [1] "crs"
#> attr(,"class")
#> [1] "sfc_POLYGON" "sfc"        
#> attr(,"precision")
#> [1] 0
#> attr(,"bbox")
#>       xmin       ymin       xmax       ymax 
#>  1.0000000 -1.9000608  3.0000000  0.9357843 
#> attr(,"class")
#> [1] "bbox"
#> attr(,"z_range")
#> zmin zmax 
#>   NA   NA 
#> attr(,"class")
#> [1] "z_range"
#> attr(,"m_range")
#> mmin mmax 
#>   NA   NA 
#> attr(,"class")
#> [1] "m_range"

sfc_polygon( obj = df, x = 2, y = 3, z = 4, m = 5)

#> [[1]]
#> [[1]]
#>       l_id           x          y          z
#>  [1,]    1  0.03849955  0.5176691 -1.6753273
#>  [2,]    1 -0.35638119 -0.1029087  1.5259387
#>  [3,]    1  0.78284410 -0.9740696  0.5541855
#>  [4,]    2  0.80441162  1.2706723  1.9931103
#>  [5,]    2 -1.90006082  0.9608648 -0.1541207
#>  [6,]    3  0.93578429  0.7687214  2.5644083
#>  [7,]    3 -0.30905150  1.0359308  1.0619991
#>  [8,]    3  0.26306668 -0.4738871  1.1426949
#>  [9,]    1 -1.79059186 -1.2753349  1.1238388
#> [10,]    1 -0.78825884 -0.3056207 -0.3970015
#> [11,]    1 -1.13302167  2.2117695 -0.8232612
#> [12,]    2  0.36365257 -1.0416684 -0.5788846
#> [13,]    2 -0.28588791 -1.1465239  1.7637894
#> 
#> attr(,"class")
#> [1] "XYZM"    "POLYGON" "sfg"    
#> 
#> attr(,"n_empty")
#> [1] 0
#> attr(,"crs")
#> $epsg
#> [1] NA
#> 
#> $proj4string
#> [1] NA
#> 
#> attr(,"class")
#> [1] "crs"
#> attr(,"class")
#> [1] "sfc_POLYGON" "sfc"        
#> attr(,"precision")
#> [1] 0
#> attr(,"bbox")
#>       xmin       ymin       xmax       ymax 
#>  1.0000000 -1.9000608  3.0000000  0.9357843 
#> attr(,"class")
#> [1] "bbox"
#> attr(,"z_range")
#>      zmin      zmax 
#> -1.675327  2.564408 
#> attr(,"class")
#> [1] "z_range"
#> attr(,"m_range")
#>      mmin      mmax 
#> -1.675327  2.564408 
#> attr(,"class")
#> [1] "m_range"

sfc_polygon( obj = df, polygon_id = "ml_id", linestring_id = "l_id" )

#> [[1]]
#> [[1]]
#>             [,1]       [,2]       [,3]      [,4]
#> [1,]  0.03849955  0.5176691 -1.6753273 0.1329921
#> [2,] -0.35638119 -0.1029087  1.5259387 0.3764993
#> [3,]  0.78284410 -0.9740696  0.5541855 1.1387077
#> 
#> [[2]]
#>            [,1]      [,2]       [,3]      [,4]
#> [1,]  0.8044116 1.2706723  1.9931103 1.2412631
#> [2,] -1.9000608 0.9608648 -0.1541207 0.6120909
#> 
#> [[3]]
#>            [,1]       [,2]     [,3]        [,4]
#> [1,]  0.9357843  0.7687214 2.564408 -0.42938009
#> [2,] -0.3090515  1.0359308 1.061999  1.36046133
#> [3,]  0.2630667 -0.4738871 1.142695 -0.07085743
#> 
#> attr(,"class")
#> [1] "XYZM"    "POLYGON" "sfg"    
#> 
#> [[2]]
#> [[1]]
#>            [,1]       [,2]       [,3]        [,4]
#> [1,] -1.7905919 -1.2753349  1.1238388 -0.27215368
#> [2,] -0.7882588 -0.3056207 -0.3970015 -2.44668003
#> [3,] -1.1330217  2.2117695 -0.8232612  0.06548664
#> 
#> [[2]]
#>            [,1]      [,2]       [,3]       [,4]
#> [1,]  0.3636526 -1.041668 -0.5788846 -1.0985089
#> [2,] -0.2858879 -1.146524  1.7637894 -0.6331782
#> 
#> attr(,"class")
#> [1] "XYZM"    "POLYGON" "sfg"    
#> 
#> attr(,"n_empty")
#> [1] 0
#> attr(,"crs")
#> $epsg
#> [1] NA
#> 
#> $proj4string
#> [1] NA
#> 
#> attr(,"class")
#> [1] "crs"
#> attr(,"class")
#> [1] "sfc_POLYGON" "sfc"        
#> attr(,"precision")
#> [1] 0
#> attr(,"bbox")
#>       xmin       ymin       xmax       ymax 
#> -1.9000608 -1.2753349  0.9357843  2.2117695 
#> attr(,"class")
#> [1] "bbox"
#> attr(,"z_range")
#> zmin zmax 
#>   -2    1 
#> attr(,"class")
#> [1] "z_range"
#> attr(,"m_range")
#> mmin mmax 
#>   -2    1 
#> attr(,"class")
#> [1] "m_range"

sfc_polygon( obj = df, polygon_id = 1, linestring_id = 2 )

#> [[1]]
#> [[1]]
#>             [,1]       [,2]       [,3]      [,4]
#> [1,]  0.03849955  0.5176691 -1.6753273 0.1329921
#> [2,] -0.35638119 -0.1029087  1.5259387 0.3764993
#> [3,]  0.78284410 -0.9740696  0.5541855 1.1387077
#> 
#> [[2]]
#>            [,1]      [,2]       [,3]      [,4]
#> [1,]  0.8044116 1.2706723  1.9931103 1.2412631
#> [2,] -1.9000608 0.9608648 -0.1541207 0.6120909
#> 
#> [[3]]
#>            [,1]       [,2]     [,3]        [,4]
#> [1,]  0.9357843  0.7687214 2.564408 -0.42938009
#> [2,] -0.3090515  1.0359308 1.061999  1.36046133
#> [3,]  0.2630667 -0.4738871 1.142695 -0.07085743
#> 
#> attr(,"class")
#> [1] "XYZM"    "POLYGON" "sfg"    
#> 
#> [[2]]
#> [[1]]
#>            [,1]       [,2]       [,3]        [,4]
#> [1,] -1.7905919 -1.2753349  1.1238388 -0.27215368
#> [2,] -0.7882588 -0.3056207 -0.3970015 -2.44668003
#> [3,] -1.1330217  2.2117695 -0.8232612  0.06548664
#> 
#> [[2]]
#>            [,1]      [,2]       [,3]       [,4]
#> [1,]  0.3636526 -1.041668 -0.5788846 -1.0985089
#> [2,] -0.2858879 -1.146524  1.7637894 -0.6331782
#> 
#> attr(,"class")
#> [1] "XYZM"    "POLYGON" "sfg"    
#> 
#> attr(,"n_empty")
#> [1] 0
#> attr(,"crs")
#> $epsg
#> [1] NA
#> 
#> $proj4string
#> [1] NA
#> 
#> attr(,"class")
#> [1] "crs"
#> attr(,"class")
#> [1] "sfc_POLYGON" "sfc"        
#> attr(,"precision")
#> [1] 0
#> attr(,"bbox")
#>       xmin       ymin       xmax       ymax 
#> -1.9000608 -1.2753349  0.9357843  2.2117695 
#> attr(,"class")
#> [1] "bbox"
#> attr(,"z_range")
#>      zmin      zmax 
#> -2.446680  1.360461 
#> attr(,"class")
#> [1] "z_range"
#> attr(,"m_range")
#>      mmin      mmax 
#> -2.446680  1.360461 
#> attr(,"class")
#> [1] "m_range"