sfc POLYGON

constructs an sfc of POLYGON objects

sfc_polygon(
  obj = NULL,
  x = NULL,
  y = NULL,
  z = NULL,
  m = NULL,
  polygon_id = NULL,
  linestring_id = NULL,
  close = TRUE
)

Arguments

obj

sorted matrix or data.frame

x

x geometry column

y

y geometry column

z

z geometry column

m

m geometry column

polygon_id

column of ids for polygons

linestring_id

column of ids for lines (within polygons)

close

logical indicating whether polygons should be closed. If TRUE, all polygons will be checked and force closed if possible

Value

sfc object of POLYGON geometries

notes

sfheaders functions do not perform any validity checks on the geometries. Nor do they set Coordinate Reference Systems, EPSG, PROJ4 or precision attributes.

The data.frame and matrices you send into the sfheader functions must be ordered.

Examples

m <- matrix(c(0,0,0,0,1,1), ncol = 2 )
sfc_polygon( m )
#> [[1]]
#> [[1]]
#>      [,1] [,2]
#> [1,]    0    0
#> [2,]    0    1
#> [3,]    0    1
#> [4,]    0    0
#> 
#> attr(,"class")
#> [1] "XY"      "POLYGON" "sfg"    
#> 
#> attr(,"n_empty")
#> [1] 0
#> attr(,"crs")
#> $input
#> [1] NA
#> 
#> $wkt
#> [1] NA
#> 
#> attr(,"class")
#> [1] "crs"
#> attr(,"class")
#> [1] "sfc_POLYGON" "sfc"        
#> attr(,"precision")
#> [1] 0
#> attr(,"bbox")
#> xmin ymin xmax ymax 
#>    0    0    0    1 
#> attr(,"class")
#> [1] "bbox"

m <- matrix(c(0,0,0,0,0,1,0,1,1,1,2,2,1,2,3,1,3,2), ncol = 3, byrow = TRUE)
sfc_polygon( obj = m )
#> [[1]]
#> [[1]]
#>      [,1] [,2] [,3]
#> [1,]    0    0    0
#> [2,]    0    0    1
#> [3,]    0    1    1
#> [4,]    1    2    2
#> [5,]    1    2    3
#> [6,]    1    3    2
#> [7,]    0    0    0
#> 
#> attr(,"class")
#> [1] "XYZ"     "POLYGON" "sfg"    
#> 
#> attr(,"n_empty")
#> [1] 0
#> attr(,"crs")
#> $input
#> [1] NA
#> 
#> $wkt
#> [1] NA
#> 
#> attr(,"class")
#> [1] "crs"
#> attr(,"class")
#> [1] "sfc_POLYGON" "sfc"        
#> attr(,"precision")
#> [1] 0
#> attr(,"bbox")
#> xmin ymin xmax ymax 
#>    0    0    1    3 
#> attr(,"class")
#> [1] "bbox"
#> attr(,"z_range")
#> zmin zmax 
#>    0    3 
#> attr(,"class")
#> [1] "z_range"
sfc_polygon( obj = m, polygon_id = 1 )
#> [[1]]
#> [[1]]
#>      [,1] [,2]
#> [1,]    0    0
#> [2,]    0    1
#> [3,]    1    1
#> [4,]    0    0
#> 
#> attr(,"class")
#> [1] "XY"      "POLYGON" "sfg"    
#> 
#> [[2]]
#> [[1]]
#>      [,1] [,2]
#> [1,]    2    2
#> [2,]    2    3
#> [3,]    3    2
#> [4,]    2    2
#> 
#> attr(,"class")
#> [1] "XY"      "POLYGON" "sfg"    
#> 
#> attr(,"n_empty")
#> [1] 0
#> attr(,"crs")
#> $input
#> [1] NA
#> 
#> $wkt
#> [1] NA
#> 
#> attr(,"class")
#> [1] "crs"
#> attr(,"class")
#> [1] "sfc_POLYGON" "sfc"        
#> attr(,"precision")
#> [1] 0
#> attr(,"bbox")
#> xmin ymin xmax ymax 
#>    0    0    3    3 
#> attr(,"class")
#> [1] "bbox"
sfc_polygon( obj = m, linestring_id = 1 )
#> [[1]]
#> [[1]]
#>      [,1] [,2]
#> [1,]    0    0
#> [2,]    0    1
#> [3,]    1    1
#> [4,]    0    0
#> 
#> [[2]]
#>      [,1] [,2]
#> [1,]    2    2
#> [2,]    2    3
#> [3,]    3    2
#> [4,]    2    2
#> 
#> attr(,"class")
#> [1] "XY"      "POLYGON" "sfg"    
#> 
#> attr(,"n_empty")
#> [1] 0
#> attr(,"crs")
#> $input
#> [1] NA
#> 
#> $wkt
#> [1] NA
#> 
#> attr(,"class")
#> [1] "crs"
#> attr(,"class")
#> [1] "sfc_POLYGON" "sfc"        
#> attr(,"precision")
#> [1] 0
#> attr(,"bbox")
#> xmin ymin xmax ymax 
#>    0    0    3    3 
#> attr(,"class")
#> [1] "bbox"

sfc_polygon( obj = m, linestring_id = 1, polygon_id = 1 )
#> [[1]]
#> [[1]]
#>      [,1] [,2]
#> [1,]    0    0
#> [2,]    0    1
#> [3,]    1    1
#> [4,]    0    0
#> 
#> attr(,"class")
#> [1] "XY"      "POLYGON" "sfg"    
#> 
#> [[2]]
#> [[1]]
#>      [,1] [,2]
#> [1,]    2    2
#> [2,]    2    3
#> [3,]    3    2
#> [4,]    2    2
#> 
#> attr(,"class")
#> [1] "XY"      "POLYGON" "sfg"    
#> 
#> attr(,"n_empty")
#> [1] 0
#> attr(,"crs")
#> $input
#> [1] NA
#> 
#> $wkt
#> [1] NA
#> 
#> attr(,"class")
#> [1] "crs"
#> attr(,"class")
#> [1] "sfc_POLYGON" "sfc"        
#> attr(,"precision")
#> [1] 0
#> attr(,"bbox")
#> xmin ymin xmax ymax 
#>    0    0    3    3 
#> attr(,"class")
#> [1] "bbox"

sfc_polygon( obj = m, x = 2, y = 3 )
#> [[1]]
#> [[1]]
#>      [,1] [,2]
#> [1,]    0    0
#> [2,]    0    1
#> [3,]    1    1
#> [4,]    2    2
#> [5,]    2    3
#> [6,]    3    2
#> [7,]    0    0
#> 
#> attr(,"class")
#> [1] "XY"      "POLYGON" "sfg"    
#> 
#> attr(,"n_empty")
#> [1] 0
#> attr(,"crs")
#> $input
#> [1] NA
#> 
#> $wkt
#> [1] NA
#> 
#> attr(,"class")
#> [1] "crs"
#> attr(,"class")
#> [1] "sfc_POLYGON" "sfc"        
#> attr(,"precision")
#> [1] 0
#> attr(,"bbox")
#> xmin ymin xmax ymax 
#>    0    0    3    3 
#> attr(,"class")
#> [1] "bbox"
sfc_polygon( obj = m, x = 1, y = 2, z = 3 )
#> [[1]]
#> [[1]]
#>      [,1] [,2] [,3]
#> [1,]    0    0    0
#> [2,]    0    0    1
#> [3,]    0    1    1
#> [4,]    1    2    2
#> [5,]    1    2    3
#> [6,]    1    3    2
#> [7,]    0    0    0
#> 
#> attr(,"class")
#> [1] "XYZ"     "POLYGON" "sfg"    
#> 
#> attr(,"n_empty")
#> [1] 0
#> attr(,"crs")
#> $input
#> [1] NA
#> 
#> $wkt
#> [1] NA
#> 
#> attr(,"class")
#> [1] "crs"
#> attr(,"class")
#> [1] "sfc_POLYGON" "sfc"        
#> attr(,"precision")
#> [1] 0
#> attr(,"bbox")
#> xmin ymin xmax ymax 
#>    0    0    1    3 
#> attr(,"class")
#> [1] "bbox"
#> attr(,"z_range")
#> zmin zmax 
#>    0    3 
#> attr(,"class")
#> [1] "z_range"
sfc_polygon( obj = m, x = 2, y = 3, linestring_id = 1, polygon_id = 1 )
#> [[1]]
#> [[1]]
#>      [,1] [,2]
#> [1,]    0    0
#> [2,]    0    1
#> [3,]    1    1
#> [4,]    0    0
#> 
#> attr(,"class")
#> [1] "XY"      "POLYGON" "sfg"    
#> 
#> [[2]]
#> [[1]]
#>      [,1] [,2]
#> [1,]    2    2
#> [2,]    2    3
#> [3,]    3    2
#> [4,]    2    2
#> 
#> attr(,"class")
#> [1] "XY"      "POLYGON" "sfg"    
#> 
#> attr(,"n_empty")
#> [1] 0
#> attr(,"crs")
#> $input
#> [1] NA
#> 
#> $wkt
#> [1] NA
#> 
#> attr(,"class")
#> [1] "crs"
#> attr(,"class")
#> [1] "sfc_POLYGON" "sfc"        
#> attr(,"precision")
#> [1] 0
#> attr(,"bbox")
#> xmin ymin xmax ymax 
#>    0    0    3    3 
#> attr(,"class")
#> [1] "bbox"

df <- data.frame(
  ml_id = c(1,1,1,1,1,1,1,1,1,2,2,2,2,2,2)
  , l_id = c(1,1,1,2,2,2,3,3,3,1,1,1,2,2,2)
  , x = rnorm(15)
  , y = rnorm(15)
  , z = rnorm(15)
  , m = rnorm(15)
)

sfc_polygon( obj = df, x = "x", y = "y")
#> [[1]]
#> [[1]]
#>              [,1]       [,2]
#>  [1,] -0.03102642 -1.5807109
#>  [2,]  0.52220417 -0.1967497
#>  [3,]  0.88364472  0.8003670
#>  [4,] -0.74658348 -0.2651682
#>  [5,]  0.44185129  0.4338654
#>  [6,] -0.48360132 -0.3224765
#>  [7,]  1.46364447  0.6914412
#>  [8,] -0.44298189 -1.3391636
#>  [9,] -0.15667588  1.0465541
#> [10,] -2.78411814  0.7124653
#> [11,]  0.63092671 -0.4579592
#> [12,] -0.94036861 -0.1499387
#> [13,] -0.06026807  0.5605560
#> [14,]  1.22139564 -0.7730841
#> [15,] -1.05698084 -0.1142629
#> [16,] -0.03102642 -1.5807109
#> 
#> attr(,"class")
#> [1] "XY"      "POLYGON" "sfg"    
#> 
#> attr(,"n_empty")
#> [1] 0
#> attr(,"crs")
#> $input
#> [1] NA
#> 
#> $wkt
#> [1] NA
#> 
#> attr(,"class")
#> [1] "crs"
#> attr(,"class")
#> [1] "sfc_POLYGON" "sfc"        
#> attr(,"precision")
#> [1] 0
#> attr(,"bbox")
#>      xmin      ymin      xmax      ymax 
#> -2.784118 -1.580711  1.463644  1.046554 
#> attr(,"class")
#> [1] "bbox"
sfc_polygon( obj = df, x = "x", y = "y", z = "z")
#> [[1]]
#> [[1]]
#>              [,1]       [,2]        [,3]
#>  [1,] -0.03102642 -1.5807109 -0.11690490
#>  [2,]  0.52220417 -0.1967497  0.98465679
#>  [3,]  0.88364472  0.8003670  1.36481338
#>  [4,] -0.74658348 -0.2651682  0.29919144
#>  [5,]  0.44185129  0.4338654 -0.86006571
#>  [6,] -0.48360132 -0.3224765 -0.95913695
#>  [7,]  1.46364447  0.6914412 -0.02466081
#>  [8,] -0.44298189 -1.3391636 -1.00912655
#>  [9,] -0.15667588  1.0465541 -0.61691373
#> [10,] -2.78411814  0.7124653 -0.29405660
#> [11,]  0.63092671 -0.4579592 -2.15482431
#> [12,] -0.94036861 -0.1499387 -0.41461687
#> [13,] -0.06026807  0.5605560 -0.02928361
#> [14,]  1.22139564 -0.7730841 -0.63828849
#> [15,] -1.05698084 -0.1142629  2.04032655
#> [16,] -0.03102642 -1.5807109 -0.11690490
#> 
#> attr(,"class")
#> [1] "XYZ"     "POLYGON" "sfg"    
#> 
#> attr(,"n_empty")
#> [1] 0
#> attr(,"crs")
#> $input
#> [1] NA
#> 
#> $wkt
#> [1] NA
#> 
#> attr(,"class")
#> [1] "crs"
#> attr(,"class")
#> [1] "sfc_POLYGON" "sfc"        
#> attr(,"precision")
#> [1] 0
#> attr(,"bbox")
#>      xmin      ymin      xmax      ymax 
#> -2.784118 -1.580711  1.463644  1.046554 
#> attr(,"class")
#> [1] "bbox"
#> attr(,"z_range")
#>      zmin      zmax 
#> -2.154824  2.040327 
#> attr(,"class")
#> [1] "z_range"
sfc_polygon( obj = df, x = "x", y = "y", z = "z", m = "m")
#> [[1]]
#> [[1]]
#>              [,1]       [,2]        [,3]       [,4]
#>  [1,] -0.03102642 -1.5807109 -0.11690490 -0.3243311
#>  [2,]  0.52220417 -0.1967497  0.98465679 -0.8530671
#>  [3,]  0.88364472  0.8003670  1.36481338  0.1006867
#>  [4,] -0.74658348 -0.2651682  0.29919144 -1.3592587
#>  [5,]  0.44185129  0.4338654 -0.86006571  0.7176586
#>  [6,] -0.48360132 -0.3224765 -0.95913695 -1.6193590
#>  [7,]  1.46364447  0.6914412 -0.02466081  0.5247028
#>  [8,] -0.44298189 -1.3391636 -1.00912655  0.3694097
#>  [9,] -0.15667588  1.0465541 -0.61691373 -1.7175126
#> [10,] -2.78411814  0.7124653 -0.29405660  1.2564200
#> [11,]  0.63092671 -0.4579592 -2.15482431  0.3229175
#> [12,] -0.94036861 -0.1499387 -0.41461687  0.5931802
#> [13,] -0.06026807  0.5605560 -0.02928361 -0.1439221
#> [14,]  1.22139564 -0.7730841 -0.63828849  0.4124129
#> [15,] -1.05698084 -0.1142629  2.04032655  0.6161192
#> [16,] -0.03102642 -1.5807109 -0.11690490 -0.3243311
#> 
#> attr(,"class")
#> [1] "XYZM"    "POLYGON" "sfg"    
#> 
#> attr(,"n_empty")
#> [1] 0
#> attr(,"crs")
#> $input
#> [1] NA
#> 
#> $wkt
#> [1] NA
#> 
#> attr(,"class")
#> [1] "crs"
#> attr(,"class")
#> [1] "sfc_POLYGON" "sfc"        
#> attr(,"precision")
#> [1] 0
#> attr(,"bbox")
#>      xmin      ymin      xmax      ymax 
#> -2.784118 -1.580711  1.463644  1.046554 
#> attr(,"class")
#> [1] "bbox"
#> attr(,"z_range")
#>      zmin      zmax 
#> -2.154824  2.040327 
#> attr(,"class")
#> [1] "z_range"
#> attr(,"m_range")
#>      mmin      mmax 
#> -1.717513  1.256420 
#> attr(,"class")
#> [1] "m_range"

sfc_polygon( obj = df, x = 2, y = 3)
#> [[1]]
#> [[1]]
#>       [,1]        [,2]
#>  [1,]    1 -0.03102642
#>  [2,]    1  0.52220417
#>  [3,]    1  0.88364472
#>  [4,]    2 -0.74658348
#>  [5,]    2  0.44185129
#>  [6,]    2 -0.48360132
#>  [7,]    3  1.46364447
#>  [8,]    3 -0.44298189
#>  [9,]    3 -0.15667588
#> [10,]    1 -2.78411814
#> [11,]    1  0.63092671
#> [12,]    1 -0.94036861
#> [13,]    2 -0.06026807
#> [14,]    2  1.22139564
#> [15,]    2 -1.05698084
#> [16,]    1 -0.03102642
#> 
#> attr(,"class")
#> [1] "XY"      "POLYGON" "sfg"    
#> 
#> attr(,"n_empty")
#> [1] 0
#> attr(,"crs")
#> $input
#> [1] NA
#> 
#> $wkt
#> [1] NA
#> 
#> attr(,"class")
#> [1] "crs"
#> attr(,"class")
#> [1] "sfc_POLYGON" "sfc"        
#> attr(,"precision")
#> [1] 0
#> attr(,"bbox")
#>      xmin      ymin      xmax      ymax 
#>  1.000000 -2.784118  3.000000  1.463644 
#> attr(,"class")
#> [1] "bbox"
sfc_polygon( obj = df, x = 2, y = 3, z = 4)
#> [[1]]
#> [[1]]
#>       [,1]        [,2]       [,3]
#>  [1,]    1 -0.03102642 -1.5807109
#>  [2,]    1  0.52220417 -0.1967497
#>  [3,]    1  0.88364472  0.8003670
#>  [4,]    2 -0.74658348 -0.2651682
#>  [5,]    2  0.44185129  0.4338654
#>  [6,]    2 -0.48360132 -0.3224765
#>  [7,]    3  1.46364447  0.6914412
#>  [8,]    3 -0.44298189 -1.3391636
#>  [9,]    3 -0.15667588  1.0465541
#> [10,]    1 -2.78411814  0.7124653
#> [11,]    1  0.63092671 -0.4579592
#> [12,]    1 -0.94036861 -0.1499387
#> [13,]    2 -0.06026807  0.5605560
#> [14,]    2  1.22139564 -0.7730841
#> [15,]    2 -1.05698084 -0.1142629
#> [16,]    1 -0.03102642 -1.5807109
#> 
#> attr(,"class")
#> [1] "XYZ"     "POLYGON" "sfg"    
#> 
#> attr(,"n_empty")
#> [1] 0
#> attr(,"crs")
#> $input
#> [1] NA
#> 
#> $wkt
#> [1] NA
#> 
#> attr(,"class")
#> [1] "crs"
#> attr(,"class")
#> [1] "sfc_POLYGON" "sfc"        
#> attr(,"precision")
#> [1] 0
#> attr(,"bbox")
#>      xmin      ymin      xmax      ymax 
#>  1.000000 -2.784118  3.000000  1.463644 
#> attr(,"class")
#> [1] "bbox"
#> attr(,"z_range")
#>      zmin      zmax 
#> -1.580711  1.046554 
#> attr(,"class")
#> [1] "z_range"
sfc_polygon( obj = df, x = 2, y = 3, z = 4, m = 5)
#> [[1]]
#> [[1]]
#>       [,1]        [,2]       [,3]        [,4]
#>  [1,]    1 -0.03102642 -1.5807109 -0.11690490
#>  [2,]    1  0.52220417 -0.1967497  0.98465679
#>  [3,]    1  0.88364472  0.8003670  1.36481338
#>  [4,]    2 -0.74658348 -0.2651682  0.29919144
#>  [5,]    2  0.44185129  0.4338654 -0.86006571
#>  [6,]    2 -0.48360132 -0.3224765 -0.95913695
#>  [7,]    3  1.46364447  0.6914412 -0.02466081
#>  [8,]    3 -0.44298189 -1.3391636 -1.00912655
#>  [9,]    3 -0.15667588  1.0465541 -0.61691373
#> [10,]    1 -2.78411814  0.7124653 -0.29405660
#> [11,]    1  0.63092671 -0.4579592 -2.15482431
#> [12,]    1 -0.94036861 -0.1499387 -0.41461687
#> [13,]    2 -0.06026807  0.5605560 -0.02928361
#> [14,]    2  1.22139564 -0.7730841 -0.63828849
#> [15,]    2 -1.05698084 -0.1142629  2.04032655
#> [16,]    1 -0.03102642 -1.5807109 -0.11690490
#> 
#> attr(,"class")
#> [1] "XYZM"    "POLYGON" "sfg"    
#> 
#> attr(,"n_empty")
#> [1] 0
#> attr(,"crs")
#> $input
#> [1] NA
#> 
#> $wkt
#> [1] NA
#> 
#> attr(,"class")
#> [1] "crs"
#> attr(,"class")
#> [1] "sfc_POLYGON" "sfc"        
#> attr(,"precision")
#> [1] 0
#> attr(,"bbox")
#>      xmin      ymin      xmax      ymax 
#>  1.000000 -2.784118  3.000000  1.463644 
#> attr(,"class")
#> [1] "bbox"
#> attr(,"z_range")
#>      zmin      zmax 
#> -1.580711  1.046554 
#> attr(,"class")
#> [1] "z_range"
#> attr(,"m_range")
#>      mmin      mmax 
#> -2.154824  2.040327 
#> attr(,"class")
#> [1] "m_range"

sfc_polygon( obj = df, polygon_id = "ml_id", linestring_id = "l_id" )
#> [[1]]
#> [[1]]
#>             [,1]       [,2]       [,3]       [,4]
#> [1,] -0.03102642 -1.5807109 -0.1169049 -0.3243311
#> [2,]  0.52220417 -0.1967497  0.9846568 -0.8530671
#> [3,]  0.88364472  0.8003670  1.3648134  0.1006867
#> [4,] -0.03102642 -1.5807109 -0.1169049 -0.3243311
#> 
#> [[2]]
#>            [,1]       [,2]       [,3]       [,4]
#> [1,] -0.7465835 -0.2651682  0.2991914 -1.3592587
#> [2,]  0.4418513  0.4338654 -0.8600657  0.7176586
#> [3,] -0.4836013 -0.3224765 -0.9591370 -1.6193590
#> [4,] -0.7465835 -0.2651682  0.2991914 -1.3592587
#> 
#> [[3]]
#>            [,1]       [,2]        [,3]       [,4]
#> [1,]  1.4636445  0.6914412 -0.02466081  0.5247028
#> [2,] -0.4429819 -1.3391636 -1.00912655  0.3694097
#> [3,] -0.1566759  1.0465541 -0.61691373 -1.7175126
#> [4,]  1.4636445  0.6914412 -0.02466081  0.5247028
#> 
#> attr(,"class")
#> [1] "XYZM"    "POLYGON" "sfg"    
#> 
#> [[2]]
#> [[1]]
#>            [,1]       [,2]       [,3]      [,4]
#> [1,] -2.7841181  0.7124653 -0.2940566 1.2564200
#> [2,]  0.6309267 -0.4579592 -2.1548243 0.3229175
#> [3,] -0.9403686 -0.1499387 -0.4146169 0.5931802
#> [4,] -2.7841181  0.7124653 -0.2940566 1.2564200
#> 
#> [[2]]
#>             [,1]       [,2]        [,3]       [,4]
#> [1,] -0.06026807  0.5605560 -0.02928361 -0.1439221
#> [2,]  1.22139564 -0.7730841 -0.63828849  0.4124129
#> [3,] -1.05698084 -0.1142629  2.04032655  0.6161192
#> [4,] -0.06026807  0.5605560 -0.02928361 -0.1439221
#> 
#> attr(,"class")
#> [1] "XYZM"    "POLYGON" "sfg"    
#> 
#> attr(,"n_empty")
#> [1] 0
#> attr(,"crs")
#> $input
#> [1] NA
#> 
#> $wkt
#> [1] NA
#> 
#> attr(,"class")
#> [1] "crs"
#> attr(,"class")
#> [1] "sfc_POLYGON" "sfc"        
#> attr(,"precision")
#> [1] 0
#> attr(,"bbox")
#>      xmin      ymin      xmax      ymax 
#> -2.784118 -1.580711  1.463644  1.046554 
#> attr(,"class")
#> [1] "bbox"
#> attr(,"z_range")
#>      zmin      zmax 
#> -2.154824  2.040327 
#> attr(,"class")
#> [1] "z_range"
#> attr(,"m_range")
#>      mmin      mmax 
#> -1.717513  1.256420 
#> attr(,"class")
#> [1] "m_range"
sfc_polygon( obj = df, polygon_id = 1, linestring_id = 2 )
#> [[1]]
#> [[1]]
#>             [,1]       [,2]       [,3]       [,4]
#> [1,] -0.03102642 -1.5807109 -0.1169049 -0.3243311
#> [2,]  0.52220417 -0.1967497  0.9846568 -0.8530671
#> [3,]  0.88364472  0.8003670  1.3648134  0.1006867
#> [4,] -0.03102642 -1.5807109 -0.1169049 -0.3243311
#> 
#> [[2]]
#>            [,1]       [,2]       [,3]       [,4]
#> [1,] -0.7465835 -0.2651682  0.2991914 -1.3592587
#> [2,]  0.4418513  0.4338654 -0.8600657  0.7176586
#> [3,] -0.4836013 -0.3224765 -0.9591370 -1.6193590
#> [4,] -0.7465835 -0.2651682  0.2991914 -1.3592587
#> 
#> [[3]]
#>            [,1]       [,2]        [,3]       [,4]
#> [1,]  1.4636445  0.6914412 -0.02466081  0.5247028
#> [2,] -0.4429819 -1.3391636 -1.00912655  0.3694097
#> [3,] -0.1566759  1.0465541 -0.61691373 -1.7175126
#> [4,]  1.4636445  0.6914412 -0.02466081  0.5247028
#> 
#> attr(,"class")
#> [1] "XYZM"    "POLYGON" "sfg"    
#> 
#> [[2]]
#> [[1]]
#>            [,1]       [,2]       [,3]      [,4]
#> [1,] -2.7841181  0.7124653 -0.2940566 1.2564200
#> [2,]  0.6309267 -0.4579592 -2.1548243 0.3229175
#> [3,] -0.9403686 -0.1499387 -0.4146169 0.5931802
#> [4,] -2.7841181  0.7124653 -0.2940566 1.2564200
#> 
#> [[2]]
#>             [,1]       [,2]        [,3]       [,4]
#> [1,] -0.06026807  0.5605560 -0.02928361 -0.1439221
#> [2,]  1.22139564 -0.7730841 -0.63828849  0.4124129
#> [3,] -1.05698084 -0.1142629  2.04032655  0.6161192
#> [4,] -0.06026807  0.5605560 -0.02928361 -0.1439221
#> 
#> attr(,"class")
#> [1] "XYZM"    "POLYGON" "sfg"    
#> 
#> attr(,"n_empty")
#> [1] 0
#> attr(,"crs")
#> $input
#> [1] NA
#> 
#> $wkt
#> [1] NA
#> 
#> attr(,"class")
#> [1] "crs"
#> attr(,"class")
#> [1] "sfc_POLYGON" "sfc"        
#> attr(,"precision")
#> [1] 0
#> attr(,"bbox")
#>      xmin      ymin      xmax      ymax 
#> -2.784118 -1.580711  1.463644  1.046554 
#> attr(,"class")
#> [1] "bbox"
#> attr(,"z_range")
#>      zmin      zmax 
#> -2.154824  2.040327 
#> attr(,"class")
#> [1] "z_range"
#> attr(,"m_range")
#>      mmin      mmax 
#> -1.717513  1.256420 
#> attr(,"class")
#> [1] "m_range"